All examples

These examples are fully documented:

ExampleData.CobbleData

ExampleData.DeValues

ExampleData.Fading

ExampleData.portableOSL

ExampleData.RLum.Data.Image

ExampleData.ScaleGammaDose

ExampleData.SurfaceExposure

ExampleData.TR_OSL

The following ones are also available, but are visible only via the internal documentation, as they correspond to unpublished or synthetic data:

ExampleData.Al2O3C

ExampleData.BINfileData

ExampleData.CW_OSL_Curve

ExampleData.FittingLM

ExampleData.LxTxData

ExampleData.LxTxOSLData

ExampleData.MortarData

ExampleData.RF70Curves

ExampleData.RLum.Analysis

ExampleData.XSYG