ExampleData.RLum.Data.Image {Luminescence}	R Documentation

Example data as RLum.Data.Image objects

Description

Measurement of Princton Instruments camera imported with the function read_SPE2R to R to produce an RLum.Data.Image object.

Format

Object of class RLum.Data.Image

Version

0.1

Source

ExampleData.RLum.Data.Image

These data were kindly provided by Regina DeWitt.

Lab.: 	 Department of Physics, East-Carolina University, NC, USA

Lab-Code: 	 -

Location: 	 -

Material: 	 -

Reference: 	 -

Image data is a measurement of fluorescent ceiling lights with a cooled Princeton Instruments (TM) camera fitted on Risoe DA-20 TL/OSL reader.

Examples



##load data
data(ExampleData.RLum.Data.Image, envir = environment())

##plot data
plot_RLum(ExampleData.RLum.Data.Image)


[Package Luminescence version 0.9.15 Index]