ExampleData.XSYG {Luminescence}	R Documentation

Example data for a SAR OSL measurement and a TL spectrum using a lexsyg reader

Description

Example data from a SAR OSL measurement and a TL spectrum for package Luminescence imported from a Freiberg Instruments XSYG file using the function read_XSYG2R.

Format

OSL.SARMeasurement: SAR OSL measurement data

The data contain two elements: (a) $Sequence.Header is a data.frame with metadata from the measurement,(b) Sequence.Object contains an RLum.Analysis object for further analysis.

TL.Spectrum: TL spectrum data

RLum.Data.Spectrum object for further analysis. The spectrum was cleaned from cosmic-rays using the function

apply_CosmicRayRemoval. Note that no quantum efficiency calibration was performed.

Version

0.1

Source

OSL.SARMeasurement

Lab: 	 Luminescence Laboratory Giessen

Lab-Code: 	 no code

Location: 	 not specified

Material: 	 Coarse grain quartz on steel cups on lexsyg research reader

Reference: 	 unpublished

TL.Spectrum

Lab: 	 Luminescence Laboratory Giessen

Lab-Code: 	 BT753

Location: 	 Dolni Vestonice/Czech Republic

Material: 	 Fine grain polymineral on steel cups on lexsyg research reader

Reference: 	 Fuchs et al., 2013

Spectrum: 	 Integration time 19 s, channel time 20 s

Heating: 	 1 K/s, up to 500 deg. C

References

Unpublished data measured to serve as example data for that package. Location origin of sample BT753 is given here:

Fuchs, M., Kreutzer, S., Rousseau, D.D., Antoine, P., Hatte, C., Lagroix, F., Moine, O., Gauthier, C., Svoboda, J., Lisa, L., 2013. The loess sequence of Dolni Vestonice, Czech Republic: A new OSL-based chronology of the Last Climatic Cycle. Boreas, 42, 664–677.

See Also

read_XSYG2R, RLum.Analysis, RLum.Data.Spectrum, plot_RLum, plot_RLum.Analysis, plot_RLum.Data.Spectrum

Examples


##show data
data(ExampleData.XSYG, envir = environment())

## =========================================
##(1) OSL.SARMeasurement
OSL.SARMeasurement

##show $Sequence.Object
OSL.SARMeasurement$Sequence.Object

##grep OSL curves and plot the first curve
OSLcurve <- get_RLum(OSL.SARMeasurement$Sequence.Object,
recordType="OSL")[[1]]
plot_RLum(OSLcurve)

## =========================================
##(2) TL.Spectrum
TL.Spectrum

##plot simple spectrum (2D)
plot_RLum.Data.Spectrum(TL.Spectrum,
                        plot.type="contour",
                        xlim = c(310,750),
                        ylim = c(0,300),
                        bin.rows=10,
                        bin.cols = 1)

##plot 3d spectrum (uncomment for usage)
# plot_RLum.Data.Spectrum(TL.Spectrum, plot.type="persp",
# xlim = c(310,750), ylim = c(0,300), bin.rows=10,
# bin.cols = 1)


[Package Luminescence version 0.9.15 Index]