PSL2Risoe.BINfileData {Luminescence}	R Documentation

Convert portable OSL data to an Risoe.BINfileData object

Description

Converts an RLum.Analysis object produced by the function read_PSL2R() to an Risoe.BINfileData object (BETA).

Usage


PSL2Risoe.BINfileData(object, ...)

Arguments

object
RLum.Analysis (required): RLum.Analysis object produced by read_PSL2R

...
currently not used.

Details

This function converts an RLum.Analysis object that was produced by the read_PSL2R function to an Risoe.BINfileData. The Risoe.BINfileData can be used to write a Risoe BIN file via write_R2BIN.

Value

Returns an S4 Risoe.BINfileData object that can be used to write a BIN file using write_R2BIN.

Function version

0.0.1

How to cite

Burow, C., 2021. PSL2Risoe.BINfileData(): Convert portable OSL data to an Risoe.BINfileData object. Function version 0.0.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Christoph Burow, University of Cologne (Germany) , RLum Developer Team

See Also

RLum.Analysis, RLum.Data.Curve, Risoe.BINfileData

Examples



# (1) load and plot example data set
data("ExampleData.portableOSL", envir = environment())
plot_RLum(ExampleData.portableOSL)

# (2) merge all RLum.Analysis objects into one
merged <- merge_RLum(ExampleData.portableOSL)
merged

# (3) convert to RisoeBINfile object
bin <- PSL2Risoe.BINfileData(merged)
bin

# (4) write Risoe BIN file
## Not run: 
write_R2BIN(bin, "~/portableOSL.binx")

## End(Not run)


[Package Luminescence version 0.9.15 Index]