Class "RLum"

Abstract class for data in the package Luminescence Sublasses are:

# S4 method for RLum
replicate_RLum(object, times = NULL)

Arguments

object
RLum (required): an object of class RLum

times
integer (optional): number for times each element is repeated element

Details

RLum-class | |----RLum.Data |----|-- RLum.Data.Curve |----|-- RLum.Data.Spectrum |----|-- RLum.Data.Image |----RLum.Analysis |----RLum.Results

Methods (by generic)

replicate_RLum: Replication method RLum-objects

Slots

originator

Object of class character containing the name of the producing function for the object. Set automatically by using the function set_RLum.

info

Object of class list for additional information on the object itself

.uid

Object of class character for a unique object identifier. This id is usually calculated using the internal function create_UID() if the funtion set_RLum is called.

.pid

Object of class character for a parent id. This allows nesting RLum-objects at will. The parent id can be the uid of another object.

Note

RLum is a virtual class.

Objects from the Class

A virtual Class: No objects can be created from it.

Class version

0.4.0

See also

RLum.Data, RLum.Data.Curve, RLum.Data.Spectrum, RLum.Data.Image, RLum.Analysis, RLum.Results, methods_RLum

How to cite

Kreutzer, S. (2018). RLum-class(): Class 'RLum'. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence

Examples

showClass("RLum")

#> Virtual Class "RLum" [package "Luminescence"]
#> 
#> Slots:
#>                                                   
#> Name:  originator       info       .uid       .pid
#> Class:  character       list  character  character
#> 
#> Known Subclasses: 
#> Class "RLum.Analysis", directly
#> Class "RLum.Data", directly
#> Class "RLum.Results", directly
#> Class "RLum.Data.Curve", by class "RLum.Data", distance 2
#> Class "RLum.Data.Image", by class "RLum.Data", distance 2
#> Class "RLum.Data.Spectrum", by class "RLum.Data", distance 2