Class "RLum.Data"

Generalized virtual data class for luminescence data.

Usage

# S4 method for class 'RLum.Data'
add_metadata(object, info_element) <- value

# S4 method for class 'RLum.Data'
rename_metadata(object, info_element) <- value

# S4 method for class 'RLum.Data'
replace_metadata(object, info_element, subset = NULL, verbose = TRUE) <- value

# S4 method for class 'RLum.Data'
view(object, ...)

Arguments

object

(required) an object of class RLum.Data

info_element

character (required) name of the metadata field to replace or remove

value

(required) The value assigned to the selected elements of the metadata field. If NULL the elements named in info_element will be removed.

subset

expression (optional) logical expression to limit the substitution only to the selected subset of elements

verbose

logical (with default) enable/disable output to the terminal.

...

other arguments that might be passed

Methods (by generic)

add_metadata(RLum.Data) <- value: Add metadata entries to RLum.Data objects

rename_metadata(RLum.Data) <- value: Rename a metadata entry of RLum.Data objects

replace_metadata(RLum.Data) <- value: Replaces or removes metadata of RLum.Data objects

view(RLum.Data): View method for RLum.Data objects

Objects from the Class

A virtual class: no objects can be created from it.

Class version

0.2.1

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)

Examples

showClass("RLum.Data")
#> Virtual Class "RLum.Data" [package "Luminescence"]
#> 
#> Slots:
#>                                                   
#> Name:  originator       info       .uid       .pid
#> Class:  character       list  character  character
#> 
#> Extends: "RLum"
#> 
#> Known Subclasses: "RLum.Data.Curve", "RLum.Data.Image", "RLum.Data.Spectrum"