Analyse fading measurements and returns the fading rate per decade (g-value)

The function analysis fading measurements and returns a fading rate including an error estimation. The function is not limited to standard fading measurements, as can be seen, e.g., Huntley and Lamothe 2001. Additionally, the density of recombination centres (rho') is estimated after Kars et al. 2008.

analyse_FadingMeasurement(object, structure = c("Lx", "Tx"), signal.integral,
  background.integral, t_star = "half", n.MC = 100, verbose = TRUE,
  plot = TRUE, plot.single = FALSE, ...)

Arguments

object
RLum.Analysis (required): input object with the measurement data. Alternatively, a list containing RLum.Analysis objects or a data.frame with three columns (x = LxTx, y = LxTx error, z = time since irradiation) can be provided. Can also be a wide table, i.e. a data.frame with a number of colums divisible by 3 and where each triplet has the before mentioned column structure.

If data from multiple aliquots are provided please see the details below with regard to Lx/Tx normalisation.

structure
character (with default): sets the structure of the measurement data. Allowed are 'Lx' or c('Lx','Tx'). Other input is ignored

signal.integral
vector (required): vector with the limits for the signal integral. Not required if a data.frame with LxTx values are provided.

background.integral
vector (required): vector with the bounds for the background integral. Not required if a data.frame with LxTx values are provided.

t_star
character (with default): method for calculating the time elasped since irradiaton. Options are: 'half', which is \(t_star := t_1 + (t_2 - t_1)/2\) (Auclair et al., 2003) and 'end', which takes the time between irradiation and the measurement step. Default is 'half'

n.MC
integer (with default): number for Monte Carlo runs for the error estimation

verbose
logical (with default): enables/disables verbose mode

plot
logical (with default): enables/disables plot output

plot.single
logical (with default): enables/disables single plot mode, i.e. one plot window per plot. Alternatively a vector specifying the plot to be drawn, e.g., plot.single = c(3,4) draws only the last two plots

...
(optional) further arguments that can be passed to internally used functions (see details)

Value

An RLum.Results object is returned:

Slot: @data

OBJECT 	 TYPE 	 COMMENT

fading_results 	 data.frame 	 results of the fading measurement in a table

fit 	 lm 	 object returned by the used linear fitting function stats::lm

rho_prime 	 data.frame 	 results of rho' estimation after Kars et al. 2008

LxTx_table 	 data.frame 	 Lx/Tx table, if curve data had been provided

irr.times 	 integer 	 vector with the irradiation times in seconds

Slot: @info

OBJECT 	 TYPE 	 COMMENT

call 	 call 	 the original function call

Details

All provided output corresponds to the \(tc\) value obtained by this analysis. Additionally in the output object the g-value normalised to 2-days is provided. The output of this function can be passed to the function calc_FadingCorr. Fitting and error estimation

For the fitting the function stats::lm is used without applying weights. For the error estimation all input values, except tc, as the precision can be consdiered as sufficiently high enough with regard to the underlying problem, are sampled assuming a normal distribution for each value with the value as the mean and the provided uncertainty as standard deviation. Density of recombination centres

The density of recombination centres, expressed by the dimensionless variable rho', is estimated by fitting equation 5 in Kars et al. 2008 to the data. For the fitting the function stats::nls is used without applying weights. For the error estimation the same procedure as for the g-value is applied (see above). Multiple aliquots & Lx/Tx normalisation

Be aware that this function will always normalise all Lx/Tx values by the Lx/Tx value of the prompt measurement of the first aliquot. This implicitly assumes that there are no systematic inter-aliquot variations in Lx/Tx values. If deemed necessary to normalise the Lx/Tx values of each aliquot by its individual prompt measurement please do so before running analyse_FadingMeasurement and provide the already normalised values for object instead.

Note

This function has BETA status and should not be used for publication work!

Function version

0.1.5 (2018-01-21 17:22:38)

How to cite

Kreutzer, S., Burow, C. (2018). analyse_FadingMeasurement(): Analyse fading measurements and returns the fading rate per decade (g-value). Function version 0.1.5. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.1. https://CRAN.R-project.org/package=Luminescence

References

Auclair, M., Lamothe, M., Huot, S., 2003. Measurement of anomalous fading for feldpsar IRSL using SAR. Radiation Measurements 37, 487-492. doi:10.1016/S1350-4487(03)00018-0

Huntley, D.J., Lamothe, M., 2001. Ubiquity of anomalous fading in K-feldspars and the measurement and correction for it in optical dating. Canadian Journal of Earth Sciences 38, 1093-1106. doi:10.1139/cjes-38-7-1093

Kars, R.H., Wallinga, J., Cohen, K.M., 2008. A new approach towards anomalous fading correction for feldspar IRSL dating-tests on samples in field saturation. Radiation Measurements 43, 786-790. doi:10.1016/j.radmeas.2008.01.021

See also

calc_OSLLxTxRatio, read_BIN2R, read_XSYG2R, extract_IrradiationTimes

Examples

## load example data (sample UNIL/NB123, see ?ExampleData.Fading)
data("ExampleData.Fading", envir = environment())

##(1) get fading measurement data (here a three column data.frame)
fading_data <- ExampleData.Fading$fading.data$IR50

##(2) run analysis
g_value <- analyse_FadingMeasurement(
fading_data,
plot = TRUE,
verbose = TRUE,
n.MC = 10)

#> Error in plot.new(): figure margins too large

##(3) this can be further used in the function
## to correct the age according to Huntley & Lamothe, 2001
results <- calc_FadingCorr(
age.faded = c(100,2),
g_value = g_value,
n.MC = 10)

#> Error in calc_FadingCorr(age.faded = c(100, 2), g_value = g_value, n.MC = 10): object 'g_value' not found