calc_FastRatio {Luminescence}	R Documentation

Calculate the Fast Ratio for CW-OSL curves

Description

Function to calculate the fast ratio of quartz CW-OSL single grain or single aliquot curves after Durcan & Duller (2011).

Usage


calc_FastRatio(
  object,
  stimulation.power = 30.6,
  wavelength = 470,
  sigmaF = 2.6e-17,
  sigmaM = 4.28e-18,
  Ch_L1 = 1,
  Ch_L2 = NULL,
  Ch_L3 = NULL,
  x = 1,
  x2 = 0.1,
  dead.channels = c(0, 0),
  fitCW.sigma = FALSE,
  fitCW.curve = FALSE,
  plot = TRUE,
  ...
)

Arguments

object
RLum.Analysis, RLum.Data.Curve or data.frame (required): x, y data of measured values (time and counts).

stimulation.power
numeric (with default): Stimulation power in mW/cm^2

wavelength
numeric (with default): Stimulation wavelength in nm

sigmaF
numeric (with default): Photoionisation cross-section (cm^2) of the fast component. Default value after Durcan & Duller (2011).

sigmaM
numeric (with default): Photoionisation cross-section (cm^2) of the medium component. Default value after Durcan & Duller (2011).

Ch_L1
numeric (with default): An integer specifying the channel for L1.

Ch_L2
numeric (optional): An integer specifying the channel for L2.

Ch_L3
numeric (optional): A vector of length 2 with integer values specifying the start and end channels for L3 (e.g., c(40, 50)).

x
numeric (with default): \ Used to define the location of L2 and L3 (start).

x2
numeric (with default): \ Used to define the location of L3 (end).

dead.channels
numeric (with default): Vector of length 2 in the form of c(x, y). Channels that do not contain OSL data, i.e. at the start or end of measurement.

fitCW.sigma
logical (optional): fit CW-OSL curve using fit_CWCurve to calculate sigmaF and sigmaM (experimental).

fitCW.curve
logical (optional): fit CW-OSL curve using fit_CWCurve and derive the counts of L2 and L3 from the fitted OSL curve (experimental).

plot
logical (with default): plot output (TRUE/FALSE)

...
available options: verbose (logical). Further arguments passed to fit_CWCurve.

Details

This function follows the equations of Durcan & Duller (2011). The energy required to reduce the fast and medium quartz OSL components to x and x2 \ and end). The fast ratio is then calculated from: (L1-L3)/(L2-L3).

Value

Returns a plot (optional) and an S4 object of type RLum.Results. The slot data contains a list with the following elements:

summary
data.frame summary of all relevant results

data
the original input data

fit
RLum.Results object if either fitCW.sigma or fitCW.curve is TRUE

args
list of used arguments

call
[call] the function call

Function version

0.1.1

How to cite

King, G.E., Durcan, J., Burow, C., 2021. calc_FastRatio(): Calculate the Fast Ratio for CW-OSL curves. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Georgina E. King, University of Bern (Switzerland)
Julie A. Durcan, University of Oxford (United Kingdom)
Christoph Burow, University of Cologne (Germany) , RLum Developer Team

References

Durcan, J.A. & Duller, G.A.T., 2011. The fast ratio: A rapid measure for testing the dominance of the fast component in the initial OSL signal from quartz. Radiation Measurements 46, 1065-1072.

Madsen, A.T., Duller, G.A.T., Donnelly, J.P., Roberts, H.M. & Wintle, A.G., 2009. A chronology of hurricane landfalls at Little Sippewissett Marsh, Massachusetts, USA, using optical dating. Geomorphology 109, 36-45.

Further reading

Steffen, D., Preusser, F. & Schlunegger, 2009. OSL quartz age underestimation due to unstable signal components. Quaternary Geochronology 4, 353-362.

See Also

fit_CWCurve, get_RLum, RLum.Analysis, RLum.Results, RLum.Data.Curve

Examples


# load example CW-OSL curve
data("ExampleData.CW_OSL_Curve")

# calculate the fast ratio w/o further adjustments
res <- calc_FastRatio(ExampleData.CW_OSL_Curve)

# show the summary table
get_RLum(res)


[Package Luminescence version 0.9.15 Index]