calc_IEU {Luminescence}	R Documentation

Apply the internal-external-uncertainty (IEU) model after Thomsen et al. (2007) to a given De distribution

Description

Function to calculate the IEU De for a De data set.

Usage


calc_IEU(data, a, b, interval, decimal.point = 2, plot = TRUE, ...)

Arguments

data
RLum.Results or data.frame (required): for data.frame: two columns with De (data[,1]) and De error (values[,2])

a
numeric (required): slope

b
numeric (required): intercept

interval
numeric (required): fixed interval (e.g. 5 Gy) used for iteration of Dbar, from the mean to Lowest.De used to create Graph.IEU [Dbar.Fixed vs Z]

decimal.point
numeric (with default): number of decimal points for rounding calculations (e.g. 2)

plot
logical (with default): plot output

...
further arguments (trace, verbose).

Details

This function uses the equations of Thomsen et al. (2007). The parameters a and b are estimated from dose-recovery experiments.

Value

Returns a plot (optional) and terminal output. In addition an RLum.Results object is returned containing the following elements:

.$summary
data.frame summary of all relevant model results.

.$data
data.frame original input data

.$args
list used arguments

.$call
call the function call

.$tables
list a list of data frames containing all calculation tables

The output should be accessed using the function get_RLum.

Function version

0.1.1

How to cite

Smedley, R.K., 2021. calc_IEU(): Apply the internal-external-uncertainty (IEU) model after Thomsen et al. (2007) to a given De distribution. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Rachel Smedley, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
Based on an excel spreadsheet and accompanying macro written by Kristina Thomsen. , RLum Developer Team

References

Smedley, R.K., 2015. A new R function for the Internal External Uncertainty (IEU) model. Ancient TL 33, 16-21.

Thomsen, K.J., Murray, A.S., Boetter-Jensen, L. & Kinahan, J., 2007. Determination of burial dose in incompletely bleached fluvial samples using single grains of quartz. Radiation Measurements 42, 370-379.

See Also

plot, calc_CommonDose, calc_CentralDose, calc_FiniteMixture, calc_FuchsLang2001, calc_MinDose

Examples



## load data
data(ExampleData.DeValues, envir = environment())

## apply the IEU model
ieu <- calc_IEU(ExampleData.DeValues$CA1, a = 0.2, b = 1.9, interval = 1)


[Package Luminescence version 0.9.15 Index]