calc_Lamothe2003 {Luminescence}	R Documentation

Apply fading correction after Lamothe et al., 2003

Description

This function applies the fading correction for the prediction of long-term fading as suggested by Lamothe et al., 2003. The function basically adjusts the Ln/Tn values and fit a new dose-response curve using the function plot_GrowthCurve.

Usage


calc_Lamothe2003(
  object,
  dose_rate.envir,
  dose_rate.source,
  g_value,
  tc = NULL,
  tc.g_value = tc,
  verbose = TRUE,
  plot = TRUE,
  ...
)

Arguments

object
RLum.Results data.frame (required): Input data for applying the fading correction. Allow are (1) data.frame with three columns (dose, LxTx, LxTx error), (2) RLum.Results object created by the function analyse_SAR.CWOSL or analyse_pIRIRSequence

dose_rate.envir
numeric vector of length 2 (required): Environmental dose rate in mGy/a

dose_rate.source
numeric vector of length 2 (required): Irradiation source dose rate in Gy/s, which is, according to Lamothe et al. (2003) De/t*.

g_value
numeric vector of length 2 (required): g_value in \ the equivalent dose was calculated, i.e. tc is either similar for the g-value measurement and the De measurement or needs be to recalculated (cf. calc_FadingCorr). Inserting a normalised g-value, e.g., normalised to 2-days , will lead to wrong results

tc
numeric (optional): time in seconds between irradiation and the prompt measurement used in the De estimation (cf. Huntley & Lamothe 2001). If set to NULL it is assumed that tc is similar for the equivalent dose estimation and the g-value estimation

tc.g_value
numeric (with default): the time in seconds between irradiation and the prompt measurement used for estimating the g-value. If the g-value was normalised to, e.g., 2 days, this time in seconds (i.e., 172800) should be given here along with the time used for the De estimation. If nothing is provided the time is set to tc, which is usual case for g-values obtained using the SAR method and g-values that had been not normalised to 2 days. Note: If this value is not NULL the functions expects a numeric value for tc.

verbose
logical (with default): Enables/disables terminal verbose mode

plot
logical (with default): Enables/disables plot output

...
further arguments passed to the function plot_GrowthCurve

Value

The function returns are graphical output produced by the function plot_GrowthCurve and an RLum.Results.

———————————–
[ NUMERICAL OUTPUT ]
———————————–

RLum.Results-object

slot: @data

Element 	 Type 	 Description

$data 	 data.frame 	 the fading corrected values

$fit 	 nls 	 the object returned by the dose response curve fitting

'slot: @info

The original function call

Function version

0.1.0

How to cite

Kreutzer, S., Mercier, N., 2021. calc_Lamothe2003(): Apply fading correction after Lamothe et al., 2003. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom), Norbert Mercier, IRAMAT-CRP2A, Université Bordeaux Montaigne (France) , RLum Developer Team

References

Huntley, D.J., Lamothe, M., 2001. Ubiquity of anomalous fading in K-feldspars and the measurement and correction for it in optical dating. Canadian Journal of Earth Sciences 38, 1093-1106.

Lamothe, M., Auclair, M., Hamzaoui, C., Huot, S., 2003. Towards a prediction of long-term anomalous fading of feldspar IRSL. Radiation Measurements 37, 493-498.

See Also

plot_GrowthCurve, calc_FadingCorr, analyse_SAR.CWOSL, analyse_pIRIRSequence

Examples



##load data
##ExampleData.BINfileData contains two BINfileData objects
##CWOSL.SAR.Data and TL.SAR.Data
data(ExampleData.BINfileData, envir = environment())

##transform the values from the first position in a RLum.Analysis object
object <- Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos=1)

##perform SAR analysis and set rejection criteria
results <- analyse_SAR.CWOSL(
object = object,
signal.integral.min = 1,
signal.integral.max = 2,
background.integral.min = 900,
background.integral.max = 1000,
verbose = FALSE,
plot = FALSE,
onlyLxTxTable = TRUE
)

##run fading correction
results_corr <- calc_Lamothe2003(
  object = results,
  dose_rate.envir =  c(1.676 , 0.180),
  dose_rate.source = c(0.184, 0.003),
  g_value =  c(2.36, 0.6),
  plot = TRUE,
  fit.method = "EXP")



[Package Luminescence version 0.9.15 Index]