calc_SourceDoseRate {Luminescence}	R Documentation

Calculation of the source dose rate via the date of measurement

Description

Calculating the dose rate of the irradiation source via the date of measurement based on: source calibration date, source dose rate, dose rate error. The function returns a data.frame that provides the input argument dose_rate for the function Second2Gray.

Usage


calc_SourceDoseRate(
  measurement.date = Sys.Date(),
  calib.date,
  calib.dose.rate,
  calib.error,
  source.type = "Sr-90",
  dose.rate.unit = "Gy/s",
  predict = NULL
)

Arguments

measurement.date
character or Date (with default): Date of measurement in "YYYY-MM-DD". If no value is provided, the date will be set to today. The argument can be provided as vector.

calib.date
character or Date (required): date of source calibration in "YYYY-MM-DD"

calib.dose.rate
numeric (required): dose rate at date of calibration in Gy/s or Gy/min

calib.error
numeric (required): error of dose rate at date of calibration Gy/s or Gy/min

source.type
character (with default): specify irradiation source (Sr-90, Co-60, Cs-137, Am-214), see details for further information

dose.rate.unit
character (with default): specify dose rate unit for input (Gy/min or Gy/s), the output is given in Gy/s as valid for the function Second2Gray

predict
integer (with default): option allowing to predict the dose rate of the source over time in days set by the provided value. Starting date is the value set with measurement.date, e.g., calc_SourceDoseRate(..., predict = 100) calculates the source dose rate for the next 100 days.

Details

Calculation of the source dose rate based on the time elapsed since the last calibration of the irradiation source. Decay parameters assume a Sr-90 beta source.

dose.rate = D0 * exp(-log(2) / T.1/2 * t)

with: D0 <- calibration dose rate T.1/2 <- half-life of the source nuclide (here in days) t <- time since source calibration (in days) log(2) / T.1/2 equals the decay constant lambda

Information on the date of measurements may be taken from the data's original .BIN file (using e.g., BINfile <- readBIN2R() and the slot BINfile@METADATA$DATE)

Allowed source types and related values

# 	 Source type 	 T.1/2 	 Reference

[1] 	 Sr-90 	 28.90 y 	 NNDC, Brookhaven National Laboratory

[2]	 Am-214 	 432.6 y 	 NNDC, Brookhaven National Laboratory

[3] 	 Co-60 	 5.274 y 	 NNDC, Brookhaven National Laboratory

[4 	 Cs-137 	 30.08 y 	 NNDC, Brookhaven National Laboratory

Value

Returns an S4 object of type RLum.Results. Slot data contains a list with the following structure:

$ dose.rate (data.frame)
.. $ dose.rate
.. $ dose.rate.error
.. $ date (corresponding measurement date)
$ parameters (list)
.. $ source.type
.. $ halflife
.. $ dose.rate.unit
$ call (the original function call)

The output should be accessed using the function get_RLum.
A plot method of the output is provided via plot_RLum

Function version

0.3.2

How to cite

Fuchs, M.C., Kreutzer, S., 2021. calc_SourceDoseRate(): Calculation of the source dose rate via the date of measurement. Function version 0.3.2. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

Please be careful when using the option predict, especially when a multiple set for measurement.date and calib.date is provided. For the source dose rate prediction the function takes the last value measurement.date and predicts from that the the source source dose rate for the number of days requested, means: the (multiple) original input will be replaced. However, the function do not change entries for the calibration dates, but mix them up. Therefore, it is not recommended to use this option when multiple calibration dates (calib.date) are provided.

Author(s)

Margret C. Fuchs, HZDR, Helmholtz-Institute Freiberg for Resource Technology (Germany)
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

References

NNDC, Brookhaven National Laboratory http://www.nndc.bnl.gov/

See Also

Second2Gray, get_RLum, plot_RLum

Examples




##(1) Simple function usage
##Basic calculation of the dose rate for a specific date
dose.rate <-  calc_SourceDoseRate(measurement.date = "2012-01-27",
                                  calib.date = "2014-12-19",
                                  calib.dose.rate = 0.0438,
                                  calib.error = 0.0019)

##show results
get_RLum(dose.rate)

##(2) Usage in combination with another function (e.g., Second2Gray() )
## load example data
data(ExampleData.DeValues, envir = environment())

## use the calculated variable dose.rate as input argument
## to convert De(s) to De(Gy)
Second2Gray(ExampleData.DeValues$BT998, dose.rate)

##(3) source rate prediction and plotting
dose.rate <-  calc_SourceDoseRate(measurement.date = "2012-01-27",
                                  calib.date = "2014-12-19",
                                  calib.dose.rate = 0.0438,
                                  calib.error = 0.0019,
                                  predict = 1000)
plot_RLum(dose.rate)


##(4) export output to a LaTeX table (example using the package 'xtable')
## Not run: 
xtable::xtable(get_RLum(dose.rate))


## End(Not run)


[Package Luminescence version 0.9.15 Index]