Convert Nuclide Activities to Concentrations and Vice Versa

The function performs the conversion of the specific activities into concentrations and vice versa for the nuclides U-238, Th-232 and K-40 to harmonise the measurement unit with the required data input unit of potential analytical tools for, e.g. dose rate calculation or related functions such as use_DRAC.

convert_Activity2Concentration(data, input_unit = "Bq/kg", verbose = TRUE)

Arguments

data
data.frame (required): provide dose rate data (activity or concentration) in three columns. The first column indicates the nuclides, the 2nd column measured value and in the 3rd column its error value. Allowed nuclide data are 'U-238', 'Th-232' and 'K-40'. See examples for an example.

input_unit
character (with default): specify unit of input data given in the dose rate data frame, choose between 'Bq/kg' and 'ppm/%' the default is 'Bq/kg'

verbose
logical (with default): enable or disable verbose mode

Details

The conversion from nuclide activity of a sample to nuclide concentration is performed using conversion factors that are based on the mass-related specific activity of the respective nuclides. The factors can be calculated using the equation: $$ A = avogadronumber * N.freq / N.mol.mass * ln(2) / N.half.life $$ $$ f = A / 10^6 $$

where:

A - specific activity of the nuclide

N.freq - natural frequency of the isotop

N.mol.mass molare mass

n.half.life half-life of the nuclide

example for U238:

\(avogadronumber = 6.02214199*10^23\)

\(uran.half.life = 1.41*10^17\) (in s)

\(uran.mol.mass = 0.23802891\) (in kg/mol)

\(uran.freq = 0.992745\) (in mol)

\(A.U = avogadronumber * uran.freq / uran.mol.mass * ln(2) / uran.half.life\) (specific activity in Bq/kg)

\(f.U = A.kg / 10^6\)

Function version

0.1.0 (2018-01-21 17:22:38)

How to cite

Fuchs, M.C. (2018). convert_Activity2Concentration(): Convert Nuclide Activities to Concentrations and Vice Versa. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence

References

Debertin, K., Helmer, R.G., 1988. Gamma- and X-ray Spectrometry with Semiconductor Detectors, Elsevier Science Publishers, p.283

Wiechen, A., Ruehle, H., Vogl, K., 2013. Bestimmung der massebezogenen Aktivitaet von Radionukliden. AEQUIVAL/MASSAKT, ISSN 1865-8725, http://www.bmub.bund.de/fileadmin/Daten_BMU/Download_PDF/Strahlenschutz/aequival-massakt_v2013-07_bf.pdf

Examples

##construct data.frame
data <- data.frame(
 NUCLIDES = c("U-238", "Th-232", "K-40"),
 VALUE = c(40,80,100),
 VALUE_ERROR = c(4,8,10),
 stringsAsFactors = FALSE)

##perform analysis
convert_Activity2Concentration(data)

#>   NUCLIDE ACTIVIY (Bq/kg) ACTIVIY ERROR (Bq/kg) CONC. (ppm/%)
#> 1   U-238              40                     4     3.2388664
#> 2  Th-232              80                     8    19.7190042
#> 3    K-40             100                    10     0.3236246
#>   CONC. ERROR (ppm/%)
#> 1          0.32388664
#> 2          1.97190042
#> 3          0.03236246