Converts RLum.Analysis-objects and RLum.Data.Curve-objects to RLum2Risoe.BINfileData-objects

The functions converts RLum.Analysis and RLum.Data.Curve objects and a list of those to Risoe.BINfileData objects. The funtion intends to provide a minimum of compatiblility between both formats. The created RLum.Analysis object can be later exported to a BIN-file using the function write_R2BIN.

convert_RLum2Risoe.BINfileData(object, keep.position.number = FALSE)

Arguments

object
RLum.Analysis or RLum.Data.Curve (required): input object to be converted

keep.position.number
logical (with default): keeps the original position number or re-calculate the numbers to avoid doubling

Value

The function returns a Risoe.BINfileData object.

Note

The conversion can be never perfect. The RLum objects may contain information which are not part of the Risoe.BINfileData definition.

Function version

0.1.3

See also

Author

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

How to cite

Kreutzer, S., 2020. convert_RLum2Risoe.BINfileData(): Converts RLum.Analysis-objects and RLum.Data.Curve-objects to RLum2Risoe.BINfileData-objects. Function version 0.1.3. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Riedesel, S., Autzen, M., Mittelstrass, D., 2020. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.10. https://CRAN.R-project.org/package=Luminescence

Examples

##simple conversion using the example dataset
data(ExampleData.RLum.Analysis, envir = environment())
convert_RLum2Risoe.BINfileData(IRSAR.RF.Data)

#> 
#> [Risoe.BINfileData object]
#> 
#> 	BIN/BINX version      08
#> 	Object date:          20201210
#> 	User:                 RLum.Dat
#> 	System ID:            0 (unknown)
#> 	Overall records:      2
#> 	Records type:         RL	(n = 2)
#> 	Position range:       0 : 0
#> 	Grain range:          0 : 0
#> 	Run range:            1 : 2
#> 	Set range:            1 : 1