Export XSYG-file(s) to CSV-files

This function is a wrapper function around the functions read_XSYG2R and write_RLum2CSV and it imports an XSYG-file and directly exports its content to CSV-files. If nothing is set for the argument path (write_RLum2CSV) the input folder will become the output folder.

convert_XSYG2CSV(file, ...)

Arguments

file
character (required): name of the XSYG-file to be converted to CSV-files

...
further arguments that will be passed to the function read_XSYG2R and write_RLum2CSV

Value

The function returns either a CSV-file (or many of them) or for the option export = FALSE a list comprising objects of type data.frame and matrix

Function version

0.1.0 (2018-01-21 17:22:38)

See also

RLum.Analysis, RLum.Data, RLum.Results, utils::write.table, write_RLum2CSV, read_XSYG2R

How to cite

Kreutzer, S. (2018). convert_XSYG2CSV(): Export XSYG-file(s) to CSV-files. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence

Examples

##transform XSYG-file values to a list
data(ExampleData.XSYG, envir = environment())
convert_XSYG2CSV(OSL.SARMeasurement$Sequence.Object[1:10], export = FALSE)

#> $`1_TL (UVVIS)`
#>          [,1] [,2]
#>    [1,]   0.1    9
#>    [2,]   0.2    7
#>    [3,]   0.3    7
#>    [4,]   0.4    9
#>    [5,]   0.5   20
#>    [6,]   0.6   12
#>    [7,]   0.7   13
#>    [8,]   0.8   17
#>    [9,]   0.9   10
#>   [10,]   1.0   11
#>   [11,]   1.1   13
#>   [12,]   1.2   13
#>   [13,]   1.3   10
#>   [14,]   1.4    3
#>   [15,]   1.5   12
#>   [16,]   1.6   18
#>   [17,]   1.7   11
#>   [18,]   1.8   11
#>   [19,]   1.9   12
#>   [20,]   2.0    9
#>   [21,]   2.1   10
#>   [22,]   2.2    0
#>   [23,]   2.3   15
#>   [24,]   2.4   14
#>   [25,]   2.5   14
#>   [26,]   2.6   14
#>   [27,]   2.7   14
#>   [28,]   2.8    7
#>   [29,]   2.9   18
#>   [30,]   3.0   16
#>   [31,]   3.1   10
#>   [32,]   3.2   14
#>   [33,]   3.3   10
#>   [34,]   3.4   10
#>   [35,]   3.5    5
#>   [36,]   3.6   11
#>   [37,]   3.7   17
#>   [38,]   3.8   18
#>   [39,]   3.9   11
#>   [40,]   4.0    4
#>   [41,]   4.1   11
#>   [42,]   4.2   12
#>   [43,]   4.3   21
#>   [44,]   4.4    7
#>   [45,]   4.5   18
#>   [46,]   4.6    8
#>   [47,]   4.7   19
#>   [48,]   4.8    9
#>   [49,]   4.9   14
#>   [50,]   5.0   22
#>   [51,]   5.1   12
#>   [52,]   5.2   20
#>   [53,]   5.3   16
#>   [54,]   5.4   19
#>   [55,]   5.5   10
#>   [56,]   5.6   18
#>   [57,]   5.7   20
#>   [58,]   5.8   18
#>   [59,]   5.9   14
#>   [60,]   6.0    9
#>   [61,]   6.1    9
#>   [62,]   6.2   17
#>   [63,]   6.3   26
#>   [64,]   6.4   12
#>   [65,]   6.5   13
#>   [66,]   6.6   10
#>   [67,]   6.7   12
#>   [68,]   6.8    8
#>   [69,]   6.9   12
#>   [70,]   7.0   16
#>   [71,]   7.1    9
#>   [72,]   7.2   11
#>   [73,]   7.3   18
#>   [74,]   7.4    5
#>   [75,]   7.5    9
#>   [76,]   7.6    6
#>   [77,]   7.7    8
#>   [78,]   7.8   17
#>   [79,]   7.9   18
#>   [80,]   8.0   18
#>   [81,]   8.1   11
#>   [82,]   8.2   10
#>   [83,]   8.3   16
#>   [84,]   8.4   11
#>   [85,]   8.5   14
#>   [86,]   8.6   10
#>   [87,]   8.7   10
#>   [88,]   8.8    8
#>   [89,]   8.9   15
#>   [90,]   9.0    7
#>   [91,]   9.1   11
#>   [92,]   9.2   11
#>   [93,]   9.3   16
#>   [94,]   9.4   10
#>   [95,]   9.5   14
#>   [96,]   9.6   13
#>   [97,]   9.7   11
#>   [98,]   9.8   10
#>   [99,]   9.9    8
#>  [100,]  10.0   10
#>  [101,]  10.1    9
#>  [102,]  10.2   10
#>  [103,]  10.3   13
#>  [104,]  10.4   12
#>  [105,]  10.5   15
#>  [106,]  10.6   12
#>  [107,]  10.7    6
#>  [108,]  10.8   13
#>  [109,]  10.9   10
#>  [110,]  11.0   15
#>  [111,]  11.1   14
#>  [112,]  11.2    9
#>  [113,]  11.3    4
#>  [114,]  11.4   14
#>  [115,]  11.5   15
#>  [116,]  11.6    0
#>  [117,]  11.7   19
#>  [118,]  11.8   12
#>  [119,]  11.9   10
#>  [120,]  12.0    4
#>  [121,]  12.1   16
#>  [122,]  12.2   11
#>  [123,]  12.3   19
#>  [124,]  12.4   18
#>  [125,]  12.5    3
#>  [126,]  12.6   13
#>  [127,]  12.7   15
#>  [128,]  12.8    4
#>  [129,]  12.9   13
#>  [130,]  13.0    0
#>  [131,]  13.1   10
#>  [132,]  13.2   13
#>  [133,]  13.3   12
#>  [134,]  13.4    9
#>  [135,]  13.5   12
#>  [136,]  13.6    8
#>  [137,]  13.7   10
#>  [138,]  13.8   14
#>  [139,]  13.9   14
#>  [140,]  14.0   15
#>  [141,]  14.1   12
#>  [142,]  14.2    7
#>  [143,]  14.3   14
#>  [144,]  14.4    8
#>  [145,]  14.5   22
#>  [146,]  14.6   12
#>  [147,]  14.7   14
#>  [148,]  14.8   19
#>  [149,]  14.9   13
#>  [150,]  15.0    9
#>  [151,]  15.1   19
#>  [152,]  15.2   13
#>  [153,]  15.3   14
#>  [154,]  15.4   23
#>  [155,]  15.5   15
#>  [156,]  15.6   10
#>  [157,]  15.7   15
#>  [158,]  15.8   20
#>  [159,]  15.9   19
#>  [160,]  16.0   11
#>  [161,]  16.1   11
#>  [162,]  16.2    7
#>  [163,]  16.3    9
#>  [164,]  16.4    0
#>  [165,]  16.5   10
#>  [166,]  16.6   10
#>  [167,]  16.7    8
#>  [168,]  16.8   14
#>  [169,]  16.9   10
#>  [170,]  17.0    7
#>  [171,]  17.1   10
#>  [172,]  17.2    8
#>  [173,]  17.3    9
#>  [174,]  17.4    9
#>  [175,]  17.5   10
#>  [176,]  17.6   22
#>  [177,]  17.7   14
#>  [178,]  17.8   11
#>  [179,]  17.9   17
#>  [180,]  18.0   14
#>  [181,]  18.1    8
#>  [182,]  18.2    8
#>  [183,]  18.3   15
#>  [184,]  18.4    9
#>  [185,]  18.5   13
#>  [186,]  18.6   10
#>  [187,]  18.7   12
#>  [188,]  18.8    9
#>  [189,]  18.9   20
#>  [190,]  19.0   10
#>  [191,]  19.1    9
#>  [192,]  19.2    7
#>  [193,]  19.3    8
#>  [194,]  19.4   23
#>  [195,]  19.5   16
#>  [196,]  19.6   17
#>  [197,]  19.7   11
#>  [198,]  19.8   13
#>  [199,]  19.9   17
#>  [200,]  20.0   24
#>  [201,]  20.1   14
#>  [202,]  20.2   19
#>  [203,]  20.3   11
#>  [204,]  20.4    9
#>  [205,]  20.5   13
#>  [206,]  20.6   23
#>  [207,]  20.7   10
#>  [208,]  20.8   12
#>  [209,]  20.9   22
#>  [210,]  21.0    5
#>  [211,]  21.1   11
#>  [212,]  21.2   16
#>  [213,]  21.3   17
#>  [214,]  21.4   11
#>  [215,]  21.5   16
#>  [216,]  21.6   11
#>  [217,]  21.7    9
#>  [218,]  21.8   13
#>  [219,]  21.9   12
#>  [220,]  22.0    7
#>  [221,]  22.1   13
#>  [222,]  22.2   14
#>  [223,]  22.3   21
#>  [224,]  22.4    8
#>  [225,]  22.5    4
#>  [226,]  22.6   17
#>  [227,]  22.7   19
#>  [228,]  22.8    8
#>  [229,]  22.9   13
#>  [230,]  23.0   14
#>  [231,]  23.1   11
#>  [232,]  23.2   14
#>  [233,]  23.3   15
#>  [234,]  23.4   13
#>  [235,]  23.5   15
#>  [236,]  23.6    9
#>  [237,]  23.7   17
#>  [238,]  23.8   10
#>  [239,]  23.9   16
#>  [240,]  24.0    8
#>  [241,]  24.1    8
#>  [242,]  24.2   17
#>  [243,]  24.3    8
#>  [244,]  24.4    7
#>  [245,]  24.5   19
#>  [246,]  24.6   14
#>  [247,]  24.7   17
#>  [248,]  24.8    7
#>  [249,]  24.9   14
#>  [250,]  25.0   12
#>  [251,]  25.1   11
#>  [252,]  25.2   11
#>  [253,]  25.3   15
#>  [254,]  25.4   11
#>  [255,]  25.5   18
#>  [256,]  25.6   21
#>  [257,]  25.7   13
#>  [258,]  25.8   11
#>  [259,]  25.9   13
#>  [260,]  26.0   14
#>  [261,]  26.1   10
#>  [262,]  26.2   19
#>  [263,]  26.3   13
#>  [264,]  26.4    9
#>  [265,]  26.5    8
#>  [266,]  26.6    7
#>  [267,]  26.7   19
#>  [268,]  26.8   19
#>  [269,]  26.9   10
#>  [270,]  27.0   25
#>  [271,]  27.1   13
#>  [272,]  27.2   12
#>  [273,]  27.3    8
#>  [274,]  27.4   14
#>  [275,]  27.5   18
#>  [276,]  27.6   19
#>  [277,]  27.7    6
#>  [278,]  27.8   11
#>  [279,]  27.9    9
#>  [280,]  28.0   22
#>  [281,]  28.1   18
#>  [282,]  28.2   11
#>  [283,]  28.3   14
#>  [284,]  28.4   20
#>  [285,]  28.5   12
#>  [286,]  28.6   17
#>  [287,]  28.7    9
#>  [288,]  28.8    0
#>  [289,]  28.9    9
#>  [290,]  29.0    7
#>  [291,]  29.1    5
#>  [292,]  29.2    8
#>  [293,]  29.3   11
#>  [294,]  29.4   21
#>  [295,]  29.5   35
#>  [296,]  29.6    9
#>  [297,]  29.7   12
#>  [298,]  29.8   20
#>  [299,]  29.9   10
#>  [300,]  30.0   16
#>  [301,]  30.1   13
#>  [302,]  30.2    9
#>  [303,]  30.3   14
#>  [304,]  30.4   10
#>  [305,]  30.5   13
#>  [306,]  30.6   15
#>  [307,]  30.7    0
#>  [308,]  30.8    8
#>  [309,]  30.9    6
#>  [310,]  31.0   14
#>  [311,]  31.1   19
#>  [312,]  31.2   27
#>  [313,]  31.3   12
#>  [314,]  31.4   15
#>  [315,]  31.5    8
#>  [316,]  31.6   11
#>  [317,]  31.7   13
#>  [318,]  31.8   11
#>  [319,]  31.9    7
#>  [320,]  32.0   13
#>  [321,]  32.1   15
#>  [322,]  32.2    7
#>  [323,]  32.3   16
#>  [324,]  32.4    5
#>  [325,]  32.5    7
#>  [326,]  32.6   11
#>  [327,]  32.7   15
#>  [328,]  32.8   10
#>  [329,]  32.9    9
#>  [330,]  33.0   15
#>  [331,]  33.1   14
#>  [332,]  33.2   13
#>  [333,]  33.3    8
#>  [334,]  33.4   22
#>  [335,]  33.5   15
#>  [336,]  33.6   16
#>  [337,]  33.7   11
#>  [338,]  33.8   15
#>  [339,]  33.9    7
#>  [340,]  34.0    2
#>  [341,]  34.1    9
#>  [342,]  34.2   14
#>  [343,]  34.3   10
#>  [344,]  34.4   12
#>  [345,]  34.5   12
#>  [346,]  34.6    9
#>  [347,]  34.7   16
#>  [348,]  34.8   12
#>  [349,]  34.9   15
#>  [350,]  35.0   16
#>  [351,]  35.1   13
#>  [352,]  35.2   16
#>  [353,]  35.3   15
#>  [354,]  35.4   16
#>  [355,]  35.5    8
#>  [356,]  35.6   12
#>  [357,]  35.7   17
#>  [358,]  35.8   18
#>  [359,]  35.9    9
#>  [360,]  36.0    8
#>  [361,]  36.1   16
#>  [362,]  36.2   17
#>  [363,]  36.3   19
#>  [364,]  36.4    9
#>  [365,]  36.5    6
#>  [366,]  36.6   15
#>  [367,]  36.7    4
#>  [368,]  36.8   16
#>  [369,]  36.9    6
#>  [370,]  37.0    9
#>  [371,]  37.1    8
#>  [372,]  37.2    9
#>  [373,]  37.3   10
#>  [374,]  37.4   12
#>  [375,]  37.5   18
#>  [376,]  37.6   10
#>  [377,]  37.7   11
#>  [378,]  37.8   11
#>  [379,]  37.9   16
#>  [380,]  38.0   15
#>  [381,]  38.1    9
#>  [382,]  38.2   11
#>  [383,]  38.3   11
#>  [384,]  38.4   14
#>  [385,]  38.5    9
#>  [386,]  38.6   13
#>  [387,]  38.7    9
#>  [388,]  38.8   10
#>  [389,]  38.9   14
#>  [390,]  39.0    7
#>  [391,]  39.1   10
#>  [392,]  39.2   14
#>  [393,]  39.3   20
#>  [394,]  39.4   11
#>  [395,]  39.5   14
#>  [396,]  39.6    8
#>  [397,]  39.7   14
#>  [398,]  39.8    9
#>  [399,]  39.9    9
#>  [400,]  40.0    7
#>  [401,]  40.1   21
#>  [402,]  40.2   16
#>  [403,]  40.3   11
#>  [404,]  40.4   19
#>  [405,]  40.5   14
#>  [406,]  40.6    6
#>  [407,]  40.7   11
#>  [408,]  40.8   11
#>  [409,]  40.9    7
#>  [410,]  41.0   17
#>  [411,]  41.1   14
#>  [412,]  41.2   16
#>  [413,]  41.3    8
#>  [414,]  41.4    8
#>  [415,]  41.5   14
#>  [416,]  41.6   15
#>  [417,]  41.7    9
#>  [418,]  41.8    7
#>  [419,]  41.9   11
#>  [420,]  42.0   11
#>  [421,]  42.1    7
#>  [422,]  42.2    0
#>  [423,]  42.3   10
#>  [424,]  42.4   10
#>  [425,]  42.5   15
#>  [426,]  42.6   11
#>  [427,]  42.7   15
#>  [428,]  42.8    9
#>  [429,]  42.9   12
#>  [430,]  43.0   14
#>  [431,]  43.1    7
#>  [432,]  43.2    6
#>  [433,]  43.3   15
#>  [434,]  43.4   11
#>  [435,]  43.5   11
#>  [436,]  43.6    9
#>  [437,]  43.7    4
#>  [438,]  43.8    5
#>  [439,]  43.9   11
#>  [440,]  44.0   13
#>  [441,]  44.1   10
#>  [442,]  44.2   11
#>  [443,]  44.3   14
#>  [444,]  44.4   10
#>  [445,]  44.5   13
#>  [446,]  44.6   10
#>  [447,]  44.7   11
#>  [448,]  44.8    5
#>  [449,]  44.9   12
#>  [450,]  45.0   15
#>  [451,]  45.1   12
#>  [452,]  45.2    7
#>  [453,]  45.3   15
#>  [454,]  45.4   15
#>  [455,]  45.5   15
#>  [456,]  45.6   21
#>  [457,]  45.7   22
#>  [458,]  45.8    9
#>  [459,]  45.9   11
#>  [460,]  46.0   15
#>  [461,]  46.1   10
#>  [462,]  46.2   16
#>  [463,]  46.3   13
#>  [464,]  46.4    9
#>  [465,]  46.5   12
#>  [466,]  46.6    5
#>  [467,]  46.7    9
#>  [468,]  46.8   12
#>  [469,]  46.9    6
#>  [470,]  47.0   12
#>  [471,]  47.1   17
#>  [472,]  47.2   15
#>  [473,]  47.3   11
#>  [474,]  47.4    6
#>  [475,]  47.5   15
#>  [476,]  47.6    8
#>  [477,]  47.7    4
#>  [478,]  47.8   13
#>  [479,]  47.9   20
#>  [480,]  48.0   11
#>  [481,]  48.1   11
#>  [482,]  48.2    9
#>  [483,]  48.3    8
#>  [484,]  48.4   11
#>  [485,]  48.5   20
#>  [486,]  48.6   10
#>  [487,]  48.7   11
#>  [488,]  48.8   10
#>  [489,]  48.9   13
#>  [490,]  49.0   19
#>  [491,]  49.1   16
#>  [492,]  49.2   17
#>  [493,]  49.3   10
#>  [494,]  49.4   16
#>  [495,]  49.5   13
#>  [496,]  49.6    9
#>  [497,]  49.7   15
#>  [498,]  49.8   14
#>  [499,]  49.9    8
#>  [500,]  50.0   11
#>  [ reached getOption("max.print") -- omitted 680 rows ]
#> 
#> $`2_TL (NA)`
#>      [,1] [,2]
#> [1,]    0   25
#> [2,]  118  260
#> [3,]  128  260
#> [4,]  148   60
#> [5,]  268   60
#> 
#> $`3_TL (NA)`
#>          [,1] [,2]
#>    [1,]   0.0   26
#>    [2,]   0.1   27
#>    [3,]   0.2   26
#>    [4,]   0.3   29
#>    [5,]   0.4   29
#>    [6,]   0.5   28
#>    [7,]   0.6   28
#>    [8,]   0.7   28
#>    [9,]   0.8   27
#>   [10,]   0.9   27
#>   [11,]   1.0   27
#>   [12,]   1.1   27
#>   [13,]   1.2   27
#>   [14,]   1.3   26
#>   [15,]   1.4   26
#>   [16,]   1.5   26
#>   [17,]   1.6   26
#>   [18,]   1.7   26
#>   [19,]   1.8   26
#>   [20,]   1.9   26
#>   [21,]   2.0   26
#>   [22,]   2.1   26
#>   [23,]   2.2   26
#>   [24,]   2.3   26
#>   [25,]   2.4   26
#>   [26,]   2.5   26
#>   [27,]   2.6   26
#>   [28,]   2.7   26
#>   [29,]   2.8   27
#>   [30,]   2.9   27
#>   [31,]   3.0   27
#>   [32,]   3.1   27
#>   [33,]   3.2   27
#>   [34,]   3.3   26
#>   [35,]   3.4   26
#>   [36,]   3.5   26
#>   [37,]   3.6   27
#>   [38,]   3.7   27
#>   [39,]   3.8   26
#>   [40,]   3.9   26
#>   [41,]   4.0   26
#>   [42,]   4.1   26
#>   [43,]   4.2   26
#>   [44,]   4.3   27
#>   [45,]   4.4   27
#>   [46,]   4.5   27
#>   [47,]   4.6   27
#>   [48,]   4.7   27
#>   [49,]   4.8   26
#>   [50,]   4.9   26
#>   [51,]   5.0   26
#>   [52,]   5.1   26
#>   [53,]   5.2   27
#>   [54,]   5.3   27
#>   [55,]   5.4   27
#>   [56,]   5.5   27
#>   [57,]   5.6   27
#>   [58,]   5.7   26
#>   [59,]   5.8   27
#>   [60,]   5.9   27
#>   [61,]   6.0   27
#>   [62,]   6.1   27
#>   [63,]   6.2   26
#>   [64,]   6.3   26
#>   [65,]   6.4   27
#>   [66,]   6.5   27
#>   [67,]   6.6   27
#>   [68,]   6.7   27
#>   [69,]   6.8   28
#>   [70,]   6.9   28
#>   [71,]   7.0   28
#>   [72,]   7.1   29
#>   [73,]   7.2   29
#>   [74,]   7.3   29
#>   [75,]   7.4   30
#>   [76,]   7.5   30
#>   [77,]   7.6   31
#>   [78,]   7.7   31
#>   [79,]   7.8   31
#>   [80,]   7.9   31
#>   [81,]   8.0   32
#>   [82,]   8.1   32
#>   [83,]   8.2   32
#>   [84,]   8.3   33
#>   [85,]   8.4   33
#>   [86,]   8.5   33
#>   [87,]   8.6   33
#>   [88,]   8.7   33
#>   [89,]   8.8   34
#>   [90,]   8.9   34
#>   [91,]   9.0   34
#>   [92,]   9.1   35
#>   [93,]   9.2   35
#>   [94,]   9.3   35
#>   [95,]   9.4   35
#>   [96,]   9.5   36
#>   [97,]   9.6   36
#>   [98,]   9.7   37
#>   [99,]   9.8   37
#>  [100,]   9.9   37
#>  [101,]  10.0   37
#>  [102,]  10.1   38
#>  [103,]  10.2   38
#>  [104,]  10.3   38
#>  [105,]  10.4   39
#>  [106,]  10.5   39
#>  [107,]  10.6   39
#>  [108,]  10.7   39
#>  [109,]  10.8   40
#>  [110,]  10.9   40
#>  [111,]  11.0   41
#>  [112,]  11.1   41
#>  [113,]  11.2   41
#>  [114,]  11.3   41
#>  [115,]  11.4   42
#>  [116,]  11.5   42
#>  [117,]  11.6   42
#>  [118,]  11.7   43
#>  [119,]  11.8   43
#>  [120,]  11.9   43
#>  [121,]  12.0   43
#>  [122,]  12.1   43
#>  [123,]  12.2   44
#>  [124,]  12.3   44
#>  [125,]  12.4   44
#>  [126,]  12.5   44
#>  [127,]  12.6   45
#>  [128,]  12.7   45
#>  [129,]  12.8   45
#>  [130,]  12.9   45
#>  [131,]  13.0   46
#>  [132,]  13.1   46
#>  [133,]  13.2   47
#>  [134,]  13.3   47
#>  [135,]  13.4   47
#>  [136,]  13.5   48
#>  [137,]  13.6   48
#>  [138,]  13.7   48
#>  [139,]  13.8   48
#>  [140,]  13.9   49
#>  [141,]  14.0   49
#>  [142,]  14.1   49
#>  [143,]  14.2   49
#>  [144,]  14.3   49
#>  [145,]  14.4   49
#>  [146,]  14.5   50
#>  [147,]  14.6   50
#>  [148,]  14.7   51
#>  [149,]  14.8   51
#>  [150,]  14.9   51
#>  [151,]  15.0   51
#>  [152,]  15.1   51
#>  [153,]  15.2   51
#>  [154,]  15.3   51
#>  [155,]  15.4   52
#>  [156,]  15.5   52
#>  [157,]  15.6   52
#>  [158,]  15.7   52
#>  [159,]  15.8   53
#>  [160,]  15.9   53
#>  [161,]  16.0   54
#>  [162,]  16.1   54
#>  [163,]  16.2   54
#>  [164,]  16.3   54
#>  [165,]  16.4   55
#>  [166,]  16.5   55
#>  [167,]  16.6   55
#>  [168,]  16.7   55
#>  [169,]  16.8   55
#>  [170,]  16.9   55
#>  [171,]  17.0   55
#>  [172,]  17.1   55
#>  [173,]  17.2   55
#>  [174,]  17.3   55
#>  [175,]  17.4   55
#>  [176,]  17.5   55
#>  [177,]  17.6   55
#>  [178,]  17.7   56
#>  [179,]  17.8   57
#>  [180,]  17.9   57
#>  [181,]  18.0   58
#>  [182,]  18.1   59
#>  [183,]  18.2   60
#>  [184,]  18.3   61
#>  [185,]  18.4   62
#>  [186,]  18.5   62
#>  [187,]  18.6   63
#>  [188,]  18.7   63
#>  [189,]  18.8   63
#>  [190,]  18.9   63
#>  [191,]  19.0   63
#>  [192,]  19.1   62
#>  [193,]  19.2   62
#>  [194,]  19.3   61
#>  [195,]  19.4   61
#>  [196,]  19.5   62
#>  [197,]  19.6   62
#>  [198,]  19.7   63
#>  [199,]  19.8   63
#>  [200,]  19.9   63
#>  [201,]  20.0   63
#>  [202,]  20.1   64
#>  [203,]  20.2   64
#>  [204,]  20.3   64
#>  [205,]  20.4   64
#>  [206,]  20.5   65
#>  [207,]  20.6   65
#>  [208,]  20.7   64
#>  [209,]  20.8   64
#>  [210,]  20.9   64
#>  [211,]  21.0   65
#>  [212,]  21.1   65
#>  [213,]  21.2   65
#>  [214,]  21.3   65
#>  [215,]  21.4   65
#>  [216,]  21.5   65
#>  [217,]  21.6   66
#>  [218,]  21.7   66
#>  [219,]  21.8   66
#>  [220,]  21.9   66
#>  [221,]  22.0   67
#>  [222,]  22.1   67
#>  [223,]  22.2   67
#>  [224,]  22.3   67
#>  [225,]  22.4   68
#>  [226,]  22.5   68
#>  [227,]  22.6   68
#>  [228,]  22.7   68
#>  [229,]  22.8   68
#>  [230,]  22.9   69
#>  [231,]  23.0   69
#>  [232,]  23.1   69
#>  [233,]  23.2   69
#>  [234,]  23.3   70
#>  [235,]  23.4   69
#>  [236,]  23.5   69
#>  [237,]  23.6   70
#>  [238,]  23.7   70
#>  [239,]  23.8   70
#>  [240,]  23.9   70
#>  [241,]  24.0   71
#>  [242,]  24.1   71
#>  [243,]  24.2   71
#>  [244,]  24.3   71
#>  [245,]  24.4   72
#>  [246,]  24.5   72
#>  [247,]  24.6   72
#>  [248,]  24.7   72
#>  [249,]  24.8   72
#>  [250,]  24.9   72
#>  [251,]  25.0   73
#>  [252,]  25.1   73
#>  [253,]  25.2   73
#>  [254,]  25.3   73
#>  [255,]  25.4   74
#>  [256,]  25.5   74
#>  [257,]  25.6   74
#>  [258,]  25.7   74
#>  [259,]  25.8   75
#>  [260,]  25.9   75
#>  [261,]  26.0   75
#>  [262,]  26.1   75
#>  [263,]  26.2   75
#>  [264,]  26.3   75
#>  [265,]  26.4   76
#>  [266,]  26.5   77
#>  [267,]  26.6   77
#>  [268,]  26.7   77
#>  [269,]  26.8   77
#>  [270,]  26.9   77
#>  [271,]  27.0   77
#>  [272,]  27.1   78
#>  [273,]  27.2   78
#>  [274,]  27.3   78
#>  [275,]  27.4   78
#>  [276,]  27.5   78
#>  [277,]  27.6   79
#>  [278,]  27.7   79
#>  [279,]  27.8   79
#>  [280,]  27.9   79
#>  [281,]  28.0   79
#>  [282,]  28.1   79
#>  [283,]  28.2   80
#>  [284,]  28.3   80
#>  [285,]  28.4   80
#>  [286,]  28.5   81
#>  [287,]  28.6   80
#>  [288,]  28.7   80
#>  [289,]  28.8   81
#>  [290,]  28.9   81
#>  [291,]  29.0   81
#>  [292,]  29.1   81
#>  [293,]  29.2   81
#>  [294,]  29.3   82
#>  [295,]  29.4   82
#>  [296,]  29.5   82
#>  [297,]  29.6   83
#>  [298,]  29.7   83
#>  [299,]  29.8   83
#>  [300,]  29.9   83
#>  [301,]  30.0   83
#>  [302,]  30.1   84
#>  [303,]  30.2   84
#>  [304,]  30.3   85
#>  [305,]  30.4   86
#>  [306,]  30.5   86
#>  [307,]  30.6   86
#>  [308,]  30.7   86
#>  [309,]  30.8   86
#>  [310,]  30.9   87
#>  [311,]  31.0   87
#>  [312,]  31.1   87
#>  [313,]  31.2   87
#>  [314,]  31.3   86
#>  [315,]  31.4   86
#>  [316,]  31.5   86
#>  [317,]  31.6   87
#>  [318,]  31.7   87
#>  [319,]  31.8   87
#>  [320,]  31.9   87
#>  [321,]  32.0   88
#>  [322,]  32.1   88
#>  [323,]  32.2   88
#>  [324,]  32.3   88
#>  [325,]  32.4   89
#>  [326,]  32.5   89
#>  [327,]  32.6   90
#>  [328,]  32.7   90
#>  [329,]  32.8   90
#>  [330,]  32.9   90
#>  [331,]  33.0   90
#>  [332,]  33.1   90
#>  [333,]  33.2   90
#>  [334,]  33.3   91
#>  [335,]  33.4   90
#>  [336,]  33.5   90
#>  [337,]  33.6   90
#>  [338,]  33.7   91
#>  [339,]  33.8   91
#>  [340,]  33.9   91
#>  [341,]  34.0   91
#>  [342,]  34.1   91
#>  [343,]  34.2   92
#>  [344,]  34.3   92
#>  [345,]  34.4   92
#>  [346,]  34.5   92
#>  [347,]  34.6   93
#>  [348,]  34.7   93
#>  [349,]  34.8   93
#>  [350,]  34.9   93
#>  [351,]  35.0   93
#>  [352,]  35.1   94
#>  [353,]  35.2   94
#>  [354,]  35.3   94
#>  [355,]  35.4   94
#>  [356,]  35.5   94
#>  [357,]  35.6   94
#>  [358,]  35.7   95
#>  [359,]  35.8   95
#>  [360,]  35.9   95
#>  [361,]  36.0   95
#>  [362,]  36.1   95
#>  [363,]  36.2   96
#>  [364,]  36.3   96
#>  [365,]  36.4   96
#>  [366,]  36.5   96
#>  [367,]  36.6   97
#>  [368,]  36.7   97
#>  [369,]  36.8   97
#>  [370,]  36.9   97
#>  [371,]  37.0   98
#>  [372,]  37.1   98
#>  [373,]  37.2   97
#>  [374,]  37.3   98
#>  [375,]  37.4   98
#>  [376,]  37.5   98
#>  [377,]  37.6   99
#>  [378,]  37.7   99
#>  [379,]  37.8   99
#>  [380,]  37.9   99
#>  [381,]  38.0   99
#>  [382,]  38.1  100
#>  [383,]  38.2  101
#>  [384,]  38.3  101
#>  [385,]  38.4  101
#>  [386,]  38.5  101
#>  [387,]  38.6  101
#>  [388,]  38.7  101
#>  [389,]  38.8  101
#>  [390,]  38.9  101
#>  [391,]  39.0  101
#>  [392,]  39.1  101
#>  [393,]  39.2  102
#>  [394,]  39.3  102
#>  [395,]  39.4  102
#>  [396,]  39.5  102
#>  [397,]  39.6  102
#>  [398,]  39.7  103
#>  [399,]  39.8  103
#>  [400,]  39.9  103
#>  [401,]  40.0  103
#>  [402,]  40.1  103
#>  [403,]  40.2  104
#>  [404,]  40.3  104
#>  [405,]  40.4  104
#>  [406,]  40.5  105
#>  [407,]  40.6  105
#>  [408,]  40.7  105
#>  [409,]  40.8  105
#>  [410,]  40.9  105
#>  [411,]  41.0  105
#>  [412,]  41.1  106
#>  [413,]  41.2  106
#>  [414,]  41.3  106
#>  [415,]  41.4  107
#>  [416,]  41.5  106
#>  [417,]  41.6  107
#>  [418,]  41.7  107
#>  [419,]  41.8  107
#>  [420,]  41.9  107
#>  [421,]  42.0  108
#>  [422,]  42.1  108
#>  [423,]  42.2  108
#>  [424,]  42.3  108
#>  [425,]  42.4  108
#>  [426,]  42.5  109
#>  [427,]  42.6  109
#>  [428,]  42.7  109
#>  [429,]  42.8  109
#>  [430,]  42.9  109
#>  [431,]  43.0  110
#>  [432,]  43.1  110
#>  [433,]  43.2  110
#>  [434,]  43.3  110
#>  [435,]  43.4  111
#>  [436,]  43.5  111
#>  [437,]  43.6  111
#>  [438,]  43.7  111
#>  [439,]  43.8  112
#>  [440,]  43.9  111
#>  [441,]  44.0  111
#>  [442,]  44.1  111
#>  [443,]  44.2  111
#>  [444,]  44.3  111
#>  [445,]  44.4  111
#>  [446,]  44.5  111
#>  [447,]  44.6  112
#>  [448,]  44.7  112
#>  [449,]  44.8  112
#>  [450,]  44.9  113
#>  [451,]  45.0  113
#>  [452,]  45.1  114
#>  [453,]  45.2  114
#>  [454,]  45.3  114
#>  [455,]  45.4  115
#>  [456,]  45.5  115
#>  [457,]  45.6  115
#>  [458,]  45.7  115
#>  [459,]  45.8  115
#>  [460,]  45.9  115
#>  [461,]  46.0  115
#>  [462,]  46.1  115
#>  [463,]  46.2  115
#>  [464,]  46.3  115
#>  [465,]  46.4  115
#>  [466,]  46.5  116
#>  [467,]  46.6  116
#>  [468,]  46.7  116
#>  [469,]  46.8  116
#>  [470,]  46.9  117
#>  [471,]  47.0  117
#>  [472,]  47.1  117
#>  [473,]  47.2  118
#>  [474,]  47.3  118
#>  [475,]  47.4  118
#>  [476,]  47.5  118
#>  [477,]  47.6  118
#>  [478,]  47.7  119
#>  [479,]  47.8  119
#>  [480,]  47.9  119
#>  [481,]  48.0  119
#>  [482,]  48.1  120
#>  [483,]  48.2  120
#>  [484,]  48.3  120
#>  [485,]  48.4  120
#>  [486,]  48.5  120
#>  [487,]  48.6  121
#>  [488,]  48.7  121
#>  [489,]  48.8  121
#>  [490,]  48.9  121
#>  [491,]  49.0  121
#>  [492,]  49.1  122
#>  [493,]  49.2  122
#>  [494,]  49.3  122
#>  [495,]  49.4  123
#>  [496,]  49.5  123
#>  [497,]  49.6  123
#>  [498,]  49.7  123
#>  [499,]  49.8  124
#>  [500,]  49.9  124
#>  [ reached getOption("max.print") -- omitted 1240 rows ]
#> 
#> $`4_OSL (UVVIS)`
#>        [,1] [,2]
#>   [1,]  0.1  891
#>   [2,]  0.2  775
#>   [3,]  0.3  655
#>   [4,]  0.4  497
#>   [5,]  0.5  424
#>   [6,]  0.6  328
#>   [7,]  0.7  293
#>   [8,]  0.8  199
#>   [9,]  0.9  204
#>  [10,]  1.0  157
#>  [11,]  1.1  125
#>  [12,]  1.2  108
#>  [13,]  1.3   94
#>  [14,]  1.4   85
#>  [15,]  1.5   77
#>  [16,]  1.6   61
#>  [17,]  1.7   73
#>  [18,]  1.8   50
#>  [19,]  1.9   42
#>  [20,]  2.0   41
#>  [21,]  2.1   32
#>  [22,]  2.2   42
#>  [23,]  2.3   34
#>  [24,]  2.4   43
#>  [25,]  2.5   26
#>  [26,]  2.6   41
#>  [27,]  2.7   28
#>  [28,]  2.8   29
#>  [29,]  2.9   34
#>  [30,]  3.0   35
#>  [31,]  3.1   19
#>  [32,]  3.2   18
#>  [33,]  3.3   17
#>  [34,]  3.4   33
#>  [35,]  3.5   18
#>  [36,]  3.6   18
#>  [37,]  3.7   15
#>  [38,]  3.8   31
#>  [39,]  3.9   15
#>  [40,]  4.0   14
#>  [41,]  4.1   12
#>  [42,]  4.2   16
#>  [43,]  4.3   18
#>  [44,]  4.4   12
#>  [45,]  4.5   16
#>  [46,]  4.6   15
#>  [47,]  4.7    9
#>  [48,]  4.8   21
#>  [49,]  4.9   16
#>  [50,]  5.0   17
#>  [51,]  5.1   18
#>  [52,]  5.2   17
#>  [53,]  5.3   18
#>  [54,]  5.4   13
#>  [55,]  5.5   19
#>  [56,]  5.6   12
#>  [57,]  5.7   11
#>  [58,]  5.8   11
#>  [59,]  5.9   16
#>  [60,]  6.0    9
#>  [61,]  6.1   11
#>  [62,]  6.2   15
#>  [63,]  6.3    9
#>  [64,]  6.4   12
#>  [65,]  6.5   13
#>  [66,]  6.6   16
#>  [67,]  6.7    8
#>  [68,]  6.8   16
#>  [69,]  6.9   12
#>  [70,]  7.0   16
#>  [71,]  7.1    8
#>  [72,]  7.2   15
#>  [73,]  7.3    9
#>  [74,]  7.4    7
#>  [75,]  7.5   11
#>  [76,]  7.6   15
#>  [77,]  7.7   19
#>  [78,]  7.8   14
#>  [79,]  7.9   13
#>  [80,]  8.0   18
#>  [81,]  8.1   10
#>  [82,]  8.2   16
#>  [83,]  8.3    7
#>  [84,]  8.4   14
#>  [85,]  8.5   15
#>  [86,]  8.6   11
#>  [87,]  8.7   11
#>  [88,]  8.8   14
#>  [89,]  8.9   20
#>  [90,]  9.0   12
#>  [91,]  9.1   10
#>  [92,]  9.2   16
#>  [93,]  9.3    5
#>  [94,]  9.4   11
#>  [95,]  9.5    9
#>  [96,]  9.6   15
#>  [97,]  9.7   16
#>  [98,]  9.8   15
#>  [99,]  9.9   11
#> [100,] 10.0   14
#> [101,] 10.1   11
#> [102,] 10.2   12
#> [103,] 10.3   12
#> [104,] 10.4   12
#> [105,] 10.5   16
#> [106,] 10.6    9
#> [107,] 10.7   13
#> [108,] 10.8   19
#> [109,] 10.9   10
#> [110,] 11.0    8
#> [111,] 11.1    2
#> [112,] 11.2    4
#> [113,] 11.3    9
#> [114,] 11.4   11
#> [115,] 11.5   12
#> [116,] 11.6   14
#> [117,] 11.7   13
#> [118,] 11.8    8
#> [119,] 11.9   15
#> [120,] 12.0    4
#> [121,] 12.1   16
#> [122,] 12.2   28
#> [123,] 12.3    5
#> [124,] 12.4   16
#> [125,] 12.5   15
#> [126,] 12.6   14
#> [127,] 12.7   10
#> [128,] 12.8   13
#> [129,] 12.9   13
#> [130,] 13.0   14
#> [131,] 13.1   12
#> [132,] 13.2   14
#> [133,] 13.3   19
#> [134,] 13.4   16
#> [135,] 13.5   13
#> [136,] 13.6   14
#> [137,] 13.7   10
#> [138,] 13.8   11
#> [139,] 13.9    9
#> [140,] 14.0   20
#> [141,] 14.1   14
#> [142,] 14.2   13
#> [143,] 14.3   15
#> [144,] 14.4   20
#> [145,] 14.5    9
#> [146,] 14.6   17
#> [147,] 14.7   10
#> [148,] 14.8   13
#> [149,] 14.9    8
#> [150,] 15.0   15
#> [151,] 15.1   14
#> [152,] 15.2    9
#> [153,] 15.3   19
#> [154,] 15.4   14
#> [155,] 15.5    9
#> [156,] 15.6   11
#> [157,] 15.7    7
#> [158,] 15.8   14
#> [159,] 15.9   11
#> [160,] 16.0   18
#> [161,] 16.1    8
#> [162,] 16.2    8
#> [163,] 16.3   15
#> [164,] 16.4   15
#> [165,] 16.5    8
#> [166,] 16.6    0
#> [167,] 16.7   17
#> [168,] 16.8   16
#> [169,] 16.9   14
#> [170,] 17.0   12
#> [171,] 17.1    5
#> [172,] 17.2   12
#> [173,] 17.3   11
#> [174,] 17.4   16
#> [175,] 17.5    6
#> [176,] 17.6   13
#> [177,] 17.7   12
#> [178,] 17.8    9
#> [179,] 17.9    3
#> [180,] 18.0   13
#> [181,] 18.1   20
#> [182,] 18.2   12
#> [183,] 18.3   21
#> [184,] 18.4    4
#> [185,] 18.5    8
#> [186,] 18.6    9
#> [187,] 18.7   11
#> [188,] 18.8   20
#> [189,] 18.9   11
#> [190,] 19.0   12
#> [191,] 19.1    6
#> [192,] 19.2   11
#> [193,] 19.3   21
#> [194,] 19.4   16
#> [195,] 19.5   13
#> [196,] 19.6    7
#> [197,] 19.7   16
#> [198,] 19.8    9
#> [199,] 19.9   10
#> [200,] 20.0   15
#> [201,] 20.1   12
#> [202,] 20.2    8
#> [203,] 20.3    9
#> [204,] 20.4   16
#> [205,] 20.5   11
#> [206,] 20.6   17
#> [207,] 20.7   18
#> [208,] 20.8   10
#> [209,] 20.9   15
#> [210,] 21.0    9
#> [211,] 21.1   13
#> [212,] 21.2    7
#> [213,] 21.3   17
#> [214,] 21.4   12
#> [215,] 21.5   29
#> [216,] 21.6   15
#> [217,] 21.7   10
#> [218,] 21.8   11
#> [219,] 21.9   10
#> [220,] 22.0   21
#> [221,] 22.1   12
#> [222,] 22.2    5
#> [223,] 22.3    7
#> [224,] 22.4   14
#> [225,] 22.5   10
#> [226,] 22.6   14
#> [227,] 22.7    4
#> [228,] 22.8   16
#> [229,] 22.9   12
#> [230,] 23.0   15
#> [231,] 23.1    9
#> [232,] 23.2   16
#> [233,] 23.3    6
#> [234,] 23.4   13
#> [235,] 23.5   13
#> [236,] 23.6   18
#> [237,] 23.7    7
#> [238,] 23.8   13
#> [239,] 23.9   13
#> [240,] 24.0   10
#> [241,] 24.1    4
#> [242,] 24.2   22
#> [243,] 24.3    9
#> [244,] 24.4   16
#> [245,] 24.5    4
#> [246,] 24.6   12
#> [247,] 24.7    9
#> [248,] 24.8   29
#> [249,] 24.9   10
#> [250,] 25.0   19
#> [251,] 25.1    9
#> [252,] 25.2   20
#> [253,] 25.3   18
#> [254,] 25.4   14
#> [255,] 25.5   14
#> [256,] 25.6   12
#> [257,] 25.7   24
#> [258,] 25.8    8
#> [259,] 25.9   16
#> [260,] 26.0    7
#> [261,] 26.1   17
#> [262,] 26.2    9
#> [263,] 26.3   15
#> [264,] 26.4   11
#> [265,] 26.5   11
#> [266,] 26.6   16
#> [267,] 26.7    9
#> [268,] 26.8   18
#> [269,] 26.9   21
#> [270,] 27.0    6
#> [271,] 27.1   10
#> [272,] 27.2    6
#> [273,] 27.3   11
#> [274,] 27.4   12
#> [275,] 27.5   16
#> [276,] 27.6   14
#> [277,] 27.7   15
#> [278,] 27.8   17
#> [279,] 27.9   13
#> [280,] 28.0    9
#> [281,] 28.1   16
#> [282,] 28.2   15
#> [283,] 28.3    6
#> [284,] 28.4   12
#> [285,] 28.5   11
#> [286,] 28.6   19
#> [287,] 28.7    7
#> [288,] 28.8   10
#> [289,] 28.9   11
#> [290,] 29.0    7
#> [291,] 29.1   14
#> [292,] 29.2   17
#> [293,] 29.3   25
#> [294,] 29.4   18
#> [295,] 29.5   13
#> [296,] 29.6   18
#> [297,] 29.7   10
#> [298,] 29.8   21
#> [299,] 29.9   11
#> [300,] 30.0   10
#> [301,] 30.1   18
#> [302,] 30.2    9
#> [303,] 30.3    9
#> [304,] 30.4    8
#> [305,] 30.5    7
#> [306,] 30.6   15
#> [307,] 30.7    7
#> [308,] 30.8   19
#> [309,] 30.9   22
#> [310,] 31.0    6
#> [311,] 31.1   19
#> [312,] 31.2   15
#> [313,] 31.3   10
#> [314,] 31.4   12
#> [315,] 31.5   20
#> [316,] 31.6   12
#> [317,] 31.7   17
#> [318,] 31.8   12
#> [319,] 31.9   16
#> [320,] 32.0   15
#> [321,] 32.1   13
#> [322,] 32.2   19
#> [323,] 32.3   12
#> [324,] 32.4   20
#> [325,] 32.5   10
#> [326,] 32.6   15
#> [327,] 32.7   17
#> [328,] 32.8   18
#> [329,] 32.9   21
#> [330,] 33.0    9
#> [331,] 33.1   17
#> [332,] 33.2   10
#> [333,] 33.3   15
#> [334,] 33.4    8
#> [335,] 33.5    9
#> [336,] 33.6   11
#> [337,] 33.7    8
#> [338,] 33.8    2
#> [339,] 33.9   18
#> [340,] 34.0    8
#> [341,] 34.1   19
#> [342,] 34.2   24
#> [343,] 34.3    8
#> [344,] 34.4   11
#> [345,] 34.5   11
#> [346,] 34.6   11
#> [347,] 34.7   12
#> [348,] 34.8   16
#> [349,] 34.9    9
#> [350,] 35.0   13
#> [351,] 35.1    8
#> [352,] 35.2   12
#> [353,] 35.3   14
#> [354,] 35.4    9
#> [355,] 35.5   19
#> [356,] 35.6   19
#> [357,] 35.7   12
#> [358,] 35.8    7
#> [359,] 35.9   13
#> [360,] 36.0   11
#> [361,] 36.1   12
#> [362,] 36.2   13
#> [363,] 36.3   16
#> [364,] 36.4   16
#> [365,] 36.5   21
#> [366,] 36.6    5
#> [367,] 36.7    7
#> [368,] 36.8   14
#> [369,] 36.9   12
#> [370,] 37.0   20
#> [371,] 37.1   11
#> [372,] 37.2   15
#> [373,] 37.3   17
#> [374,] 37.4   11
#> [375,] 37.5    8
#> [376,] 37.6   10
#> [377,] 37.7   14
#> [378,] 37.8   11
#> [379,] 37.9   11
#> [380,] 38.0   10
#> [381,] 38.1   19
#> [382,] 38.2   23
#> [383,] 38.3   17
#> [384,] 38.4    6
#> [385,] 38.5   15
#> [386,] 38.6    9
#> [387,] 38.7    7
#> [388,] 38.8    8
#> [389,] 38.9   16
#> [390,] 39.0   12
#> [391,] 39.1    8
#> [392,] 39.2   12
#> [393,] 39.3   15
#> [394,] 39.4   15
#> [395,] 39.5    6
#> [396,] 39.6   10
#> [397,] 39.7   22
#> [398,] 39.8   16
#> [399,] 39.9    9
#> [400,] 40.0   11
#> [401,] 40.1   10
#> [402,] 40.2   17
#> [403,] 40.3    9
#> [404,] 40.4   10
#> [405,] 40.5   10
#> [406,] 40.6   13
#> [407,] 40.7   14
#> [408,] 40.8    9
#> [409,] 40.9   17
#> [410,] 41.0   20
#> [411,] 41.1    8
#> [412,] 41.2   13
#> [413,] 41.3   13
#> [414,] 41.4    5
#> [415,] 41.5   16
#> [416,] 41.6   15
#> [417,] 41.7   18
#> [418,] 41.8   12
#> [419,] 41.9   17
#> [420,] 42.0   11
#> [421,] 42.1   23
#> [422,] 42.2   17
#> [423,] 42.3   10
#> [424,] 42.4   11
#> [425,] 42.5   12
#> [426,] 42.6   11
#> [427,] 42.7   22
#> [428,] 42.8    9
#> [429,] 42.9   18
#> [430,] 43.0   12
#> [431,] 43.1   16
#> [432,] 43.2   11
#> [433,] 43.3   15
#> [434,] 43.4   16
#> [435,] 43.5    7
#> [436,] 43.6   13
#> [437,] 43.7   20
#> [438,] 43.8    9
#> [439,] 43.9   20
#> [440,] 44.0   15
#> [441,] 44.1   10
#> [442,] 44.2   10
#> [443,] 44.3   14
#> [444,] 44.4   14
#> [445,] 44.5    4
#> [446,] 44.6   12
#> [447,] 44.7    5
#> [448,] 44.8   15
#> [449,] 44.9    6
#> [450,] 45.0   14
#> [451,] 45.1   15
#> [452,] 45.2   11
#> [453,] 45.3   12
#> [454,] 45.4   14
#> [455,] 45.5   12
#> [456,] 45.6    8
#> [457,] 45.7   16
#> [458,] 45.8    7
#> [459,] 45.9    9
#> [460,] 46.0   10
#> [461,] 46.1   15
#> [462,] 46.2   12
#> [463,] 46.3    5
#> [464,] 46.4   11
#> [465,] 46.5   23
#> [466,] 46.6    5
#> [467,] 46.7   10
#> [468,] 46.8   16
#> [469,] 46.9   18
#> [470,] 47.0   12
#> [471,] 47.1   16
#> [472,] 47.2   16
#> [473,] 47.3    8
#> [474,] 47.4   17
#> [475,] 47.5    8
#> [476,] 47.6   12
#> [477,] 47.7   10
#> [478,] 47.8   14
#> [479,] 47.9    8
#> [480,] 48.0   10
#> [481,] 48.1   11
#> [482,] 48.2   19
#> [483,] 48.3   12
#> [484,] 48.4    9
#> [485,] 48.5   13
#> [486,] 48.6   20
#> [487,] 48.7    9
#> [488,] 48.8   13
#> [489,] 48.9    7
#> [490,] 49.0   16
#> [491,] 49.1   10
#> [492,] 49.2   14
#> [493,] 49.3   12
#> [494,] 49.4   12
#> [495,] 49.5   10
#> [496,] 49.6    7
#> [497,] 49.7    8
#> [498,] 49.8   12
#> [499,] 49.9   19
#> [500,] 50.0   16
#> 
#> $`5_OSL (NA)`
#>      [,1] [,2]
#> [1,]    0   25
#> [2,]   20  125
#> [3,]   90  125
#> [4,]   96   60
#> [5,]  216   60
#> 
#> $`6_OSL (NA)`
#>          [,1] [,2]
#>    [1,]   0.0   50
#>    [2,]   0.1   49
#>    [3,]   0.2   48
#>    [4,]   0.3   48
#>    [5,]   0.4   48
#>    [6,]   0.5   48
#>    [7,]   0.6   48
#>    [8,]   0.7   48
#>    [9,]   0.8   48
#>   [10,]   0.9   48
#>   [11,]   1.0   47
#>   [12,]   1.1   47
#>   [13,]   1.2   47
#>   [14,]   1.3   47
#>   [15,]   1.4   47
#>   [16,]   1.5   47
#>   [17,]   1.6   47
#>   [18,]   1.7   47
#>   [19,]   1.8   47
#>   [20,]   1.9   47
#>   [21,]   2.0   47
#>   [22,]   2.1   47
#>   [23,]   2.2   47
#>   [24,]   2.3   47
#>   [25,]   2.4   46
#>   [26,]   2.5   46
#>   [27,]   2.6   46
#>   [28,]   2.7   46
#>   [29,]   2.8   46
#>   [30,]   2.9   46
#>   [31,]   3.0   46
#>   [32,]   3.1   46
#>   [33,]   3.2   46
#>   [34,]   3.3   46
#>   [35,]   3.4   46
#>   [36,]   3.5   46
#>   [37,]   3.6   46
#>   [38,]   3.7   46
#>   [39,]   3.8   46
#>   [40,]   3.9   46
#>   [41,]   4.0   46
#>   [42,]   4.1   45
#>   [43,]   4.2   45
#>   [44,]   4.3   45
#>   [45,]   4.4   45
#>   [46,]   4.5   45
#>   [47,]   4.6   45
#>   [48,]   4.7   45
#>   [49,]   4.8   44
#>   [50,]   4.9   43
#>   [51,]   5.0   43
#>   [52,]   5.1   43
#>   [53,]   5.2   44
#>   [54,]   5.3   44
#>   [55,]   5.4   43
#>   [56,]   5.5   43
#>   [57,]   5.6   43
#>   [58,]   5.7   43
#>   [59,]   5.8   43
#>   [60,]   5.9   44
#>   [61,]   6.0   44
#>   [62,]   6.1   44
#>   [63,]   6.2   44
#>   [64,]   6.3   44
#>   [65,]   6.4   44
#>   [66,]   6.5   43
#>   [67,]   6.6   43
#>   [68,]   6.7   44
#>   [69,]   6.8   44
#>   [70,]   6.9   44
#>   [71,]   7.0   44
#>   [72,]   7.1   43
#>   [73,]   7.2   43
#>   [74,]   7.3   43
#>   [75,]   7.4   43
#>   [76,]   7.5   43
#>   [77,]   7.6   43
#>   [78,]   7.7   43
#>   [79,]   7.8   44
#>   [80,]   7.9   44
#>   [81,]   8.0   44
#>   [82,]   8.1   45
#>   [83,]   8.2   45
#>   [84,]   8.3   46
#>   [85,]   8.4   47
#>   [86,]   8.5   47
#>   [87,]   8.6   48
#>   [88,]   8.7   49
#>   [89,]   8.8   49
#>   [90,]   8.9   50
#>   [91,]   9.0   50
#>   [92,]   9.1   51
#>   [93,]   9.2   51
#>   [94,]   9.3   52
#>   [95,]   9.4   53
#>   [96,]   9.5   53
#>   [97,]   9.6   54
#>   [98,]   9.7   54
#>   [99,]   9.8   55
#>  [100,]   9.9   56
#>  [101,]  10.0   58
#>  [102,]  10.1   59
#>  [103,]  10.2   59
#>  [104,]  10.3   60
#>  [105,]  10.4   61
#>  [106,]  10.5   62
#>  [107,]  10.6   64
#>  [108,]  10.7   65
#>  [109,]  10.8   65
#>  [110,]  10.9   65
#>  [111,]  11.0   66
#>  [112,]  11.1   66
#>  [113,]  11.2   67
#>  [114,]  11.3   67
#>  [115,]  11.4   68
#>  [116,]  11.5   69
#>  [117,]  11.6   69
#>  [118,]  11.7   70
#>  [119,]  11.8   70
#>  [120,]  11.9   71
#>  [121,]  12.0   71
#>  [122,]  12.1   72
#>  [123,]  12.2   73
#>  [124,]  12.3   73
#>  [125,]  12.4   74
#>  [126,]  12.5   75
#>  [127,]  12.6   75
#>  [128,]  12.7   76
#>  [129,]  12.8   77
#>  [130,]  12.9   78
#>  [131,]  13.0   78
#>  [132,]  13.1   79
#>  [133,]  13.2   79
#>  [134,]  13.3   80
#>  [135,]  13.4   81
#>  [136,]  13.5   82
#>  [137,]  13.6   82
#>  [138,]  13.7   83
#>  [139,]  13.8   84
#>  [140,]  13.9   85
#>  [141,]  14.0   86
#>  [142,]  14.1   86
#>  [143,]  14.2   87
#>  [144,]  14.3   87
#>  [145,]  14.4   89
#>  [146,]  14.5   89
#>  [147,]  14.6   90
#>  [148,]  14.7   90
#>  [149,]  14.8   91
#>  [150,]  14.9   90
#>  [151,]  15.0   90
#>  [152,]  15.1   91
#>  [153,]  15.2   91
#>  [154,]  15.3   92
#>  [155,]  15.4   92
#>  [156,]  15.5   92
#>  [157,]  15.6   93
#>  [158,]  15.7   93
#>  [159,]  15.8   94
#>  [160,]  15.9   95
#>  [161,]  16.0   97
#>  [162,]  16.1   97
#>  [163,]  16.2   98
#>  [164,]  16.3   98
#>  [165,]  16.4   99
#>  [166,]  16.5  100
#>  [167,]  16.6  100
#>  [168,]  16.7  101
#>  [169,]  16.8  102
#>  [170,]  16.9  103
#>  [171,]  17.0  103
#>  [172,]  17.1  103
#>  [173,]  17.2  104
#>  [174,]  17.3  105
#>  [175,]  17.4  105
#>  [176,]  17.5  106
#>  [177,]  17.6  106
#>  [178,]  17.7  107
#>  [179,]  17.8  107
#>  [180,]  17.9  107
#>  [181,]  18.0  108
#>  [182,]  18.1  109
#>  [183,]  18.2  110
#>  [184,]  18.3  110
#>  [185,]  18.4  111
#>  [186,]  18.5  111
#>  [187,]  18.6  111
#>  [188,]  18.7  112
#>  [189,]  18.8  113
#>  [190,]  18.9  113
#>  [191,]  19.0  114
#>  [192,]  19.1  114
#>  [193,]  19.2  116
#>  [194,]  19.3  116
#>  [195,]  19.4  117
#>  [196,]  19.5  117
#>  [197,]  19.6  118
#>  [198,]  19.7  119
#>  [199,]  19.8  119
#>  [200,]  19.9  119
#>  [201,]  20.0  120
#>  [202,]  20.1  121
#>  [203,]  20.2  120
#>  [204,]  20.3  121
#>  [205,]  20.4  121
#>  [206,]  20.5  122
#>  [207,]  20.6  122
#>  [208,]  20.7  123
#>  [209,]  20.8  123
#>  [210,]  20.9  123
#>  [211,]  21.0  124
#>  [212,]  21.1  124
#>  [213,]  21.2  124
#>  [214,]  21.3  124
#>  [215,]  21.4  124
#>  [216,]  21.5  124
#>  [217,]  21.6  124
#>  [218,]  21.7  124
#>  [219,]  21.8  124
#>  [220,]  21.9  124
#>  [221,]  22.0  124
#>  [222,]  22.1  124
#>  [223,]  22.2  124
#>  [224,]  22.3  124
#>  [225,]  22.4  124
#>  [226,]  22.5  124
#>  [227,]  22.6  124
#>  [228,]  22.7  124
#>  [229,]  22.8  123
#>  [230,]  22.9  123
#>  [231,]  23.0  124
#>  [232,]  23.1  124
#>  [233,]  23.2  124
#>  [234,]  23.3  123
#>  [235,]  23.4  124
#>  [236,]  23.5  124
#>  [237,]  23.6  124
#>  [238,]  23.7  125
#>  [239,]  23.8  125
#>  [240,]  23.9  125
#>  [241,]  24.0  125
#>  [242,]  24.1  125
#>  [243,]  24.2  125
#>  [244,]  24.3  126
#>  [245,]  24.4  126
#>  [246,]  24.5  126
#>  [247,]  24.6  127
#>  [248,]  24.7  126
#>  [249,]  24.8  126
#>  [250,]  24.9  126
#>  [251,]  25.0  126
#>  [252,]  25.1  126
#>  [253,]  25.2  126
#>  [254,]  25.3  126
#>  [255,]  25.4  126
#>  [256,]  25.5  126
#>  [257,]  25.6  125
#>  [258,]  25.7  125
#>  [259,]  25.8  125
#>  [260,]  25.9  125
#>  [261,]  26.0  125
#>  [262,]  26.1  125
#>  [263,]  26.2  125
#>  [264,]  26.3  125
#>  [265,]  26.4  124
#>  [266,]  26.5  124
#>  [267,]  26.6  124
#>  [268,]  26.7  124
#>  [269,]  26.8  124
#>  [270,]  26.9  124
#>  [271,]  27.0  124
#>  [272,]  27.1  124
#>  [273,]  27.2  124
#>  [274,]  27.3  124
#>  [275,]  27.4  125
#>  [276,]  27.5  125
#>  [277,]  27.6  125
#>  [278,]  27.7  125
#>  [279,]  27.8  125
#>  [280,]  27.9  125
#>  [281,]  28.0  125
#>  [282,]  28.1  125
#>  [283,]  28.2  125
#>  [284,]  28.3  125
#>  [285,]  28.4  125
#>  [286,]  28.5  125
#>  [287,]  28.6  125
#>  [288,]  28.7  125
#>  [289,]  28.8  125
#>  [290,]  28.9  125
#>  [291,]  29.0  125
#>  [292,]  29.1  125
#>  [293,]  29.2  125
#>  [294,]  29.3  125
#>  [295,]  29.4  125
#>  [296,]  29.5  125
#>  [297,]  29.6  125
#>  [298,]  29.7  125
#>  [299,]  29.8  125
#>  [300,]  29.9  125
#>  [301,]  30.0  125
#>  [302,]  30.1  125
#>  [303,]  30.2  125
#>  [304,]  30.3  125
#>  [305,]  30.4  125
#>  [306,]  30.5  125
#>  [307,]  30.6  125
#>  [308,]  30.7  125
#>  [309,]  30.8  126
#>  [310,]  30.9  126
#>  [311,]  31.0  126
#>  [312,]  31.1  126
#>  [313,]  31.2  126
#>  [314,]  31.3  126
#>  [315,]  31.4  126
#>  [316,]  31.5  126
#>  [317,]  31.6  127
#>  [318,]  31.7  127
#>  [319,]  31.8  126
#>  [320,]  31.9  126
#>  [321,]  32.0  126
#>  [322,]  32.1  126
#>  [323,]  32.2  126
#>  [324,]  32.3  126
#>  [325,]  32.4  126
#>  [326,]  32.5  126
#>  [327,]  32.6  127
#>  [328,]  32.7  126
#>  [329,]  32.8  126
#>  [330,]  32.9  126
#>  [331,]  33.0  126
#>  [332,]  33.1  126
#>  [333,]  33.2  127
#>  [334,]  33.3  127
#>  [335,]  33.4  127
#>  [336,]  33.5  127
#>  [337,]  33.6  126
#>  [338,]  33.7  126
#>  [339,]  33.8  126
#>  [340,]  33.9  126
#>  [341,]  34.0  126
#>  [342,]  34.1  126
#>  [343,]  34.2  126
#>  [344,]  34.3  126
#>  [345,]  34.4  126
#>  [346,]  34.5  126
#>  [347,]  34.6  126
#>  [348,]  34.7  126
#>  [349,]  34.8  126
#>  [350,]  34.9  126
#>  [351,]  35.0  126
#>  [352,]  35.1  126
#>  [353,]  35.2  126
#>  [354,]  35.3  126
#>  [355,]  35.4  126
#>  [356,]  35.5  126
#>  [357,]  35.6  126
#>  [358,]  35.7  126
#>  [359,]  35.8  125
#>  [360,]  35.9  125
#>  [361,]  36.0  125
#>  [362,]  36.1  125
#>  [363,]  36.2  125
#>  [364,]  36.3  125
#>  [365,]  36.4  125
#>  [366,]  36.5  125
#>  [367,]  36.6  125
#>  [368,]  36.7  125
#>  [369,]  36.8  125
#>  [370,]  36.9  125
#>  [371,]  37.0  125
#>  [372,]  37.1  125
#>  [373,]  37.2  125
#>  [374,]  37.3  125
#>  [375,]  37.4  125
#>  [376,]  37.5  125
#>  [377,]  37.6  125
#>  [378,]  37.7  125
#>  [379,]  37.8  125
#>  [380,]  37.9  126
#>  [381,]  38.0  126
#>  [382,]  38.1  126
#>  [383,]  38.2  125
#>  [384,]  38.3  126
#>  [385,]  38.4  126
#>  [386,]  38.5  126
#>  [387,]  38.6  126
#>  [388,]  38.7  126
#>  [389,]  38.8  125
#>  [390,]  38.9  125
#>  [391,]  39.0  125
#>  [392,]  39.1  125
#>  [393,]  39.2  125
#>  [394,]  39.3  125
#>  [395,]  39.4  125
#>  [396,]  39.5  125
#>  [397,]  39.6  125
#>  [398,]  39.7  125
#>  [399,]  39.8  125
#>  [400,]  39.9  125
#>  [401,]  40.0  125
#>  [402,]  40.1  125
#>  [403,]  40.2  125
#>  [404,]  40.3  125
#>  [405,]  40.4  125
#>  [406,]  40.5  126
#>  [407,]  40.6  126
#>  [408,]  40.7  126
#>  [409,]  40.8  126
#>  [410,]  40.9  126
#>  [411,]  41.0  126
#>  [412,]  41.1  126
#>  [413,]  41.2  126
#>  [414,]  41.3  126
#>  [415,]  41.4  126
#>  [416,]  41.5  126
#>  [417,]  41.6  126
#>  [418,]  41.7  125
#>  [419,]  41.8  125
#>  [420,]  41.9  125
#>  [421,]  42.0  125
#>  [422,]  42.1  125
#>  [423,]  42.2  125
#>  [424,]  42.3  125
#>  [425,]  42.4  125
#>  [426,]  42.5  125
#>  [427,]  42.6  125
#>  [428,]  42.7  125
#>  [429,]  42.8  125
#>  [430,]  42.9  126
#>  [431,]  43.0  126
#>  [432,]  43.1  126
#>  [433,]  43.2  126
#>  [434,]  43.3  125
#>  [435,]  43.4  126
#>  [436,]  43.5  126
#>  [437,]  43.6  126
#>  [438,]  43.7  126
#>  [439,]  43.8  126
#>  [440,]  43.9  126
#>  [441,]  44.0  125
#>  [442,]  44.1  126
#>  [443,]  44.2  125
#>  [444,]  44.3  125
#>  [445,]  44.4  125
#>  [446,]  44.5  126
#>  [447,]  44.6  126
#>  [448,]  44.7  126
#>  [449,]  44.8  126
#>  [450,]  44.9  126
#>  [451,]  45.0  126
#>  [452,]  45.1  126
#>  [453,]  45.2  126
#>  [454,]  45.3  126
#>  [455,]  45.4  126
#>  [456,]  45.5  126
#>  [457,]  45.6  126
#>  [458,]  45.7  126
#>  [459,]  45.8  126
#>  [460,]  45.9  126
#>  [461,]  46.0  127
#>  [462,]  46.1  126
#>  [463,]  46.2  126
#>  [464,]  46.3  126
#>  [465,]  46.4  127
#>  [466,]  46.5  127
#>  [467,]  46.6  127
#>  [468,]  46.7  127
#>  [469,]  46.8  127
#>  [470,]  46.9  127
#>  [471,]  47.0  127
#>  [472,]  47.1  127
#>  [473,]  47.2  127
#>  [474,]  47.3  127
#>  [475,]  47.4  126
#>  [476,]  47.5  126
#>  [477,]  47.6  126
#>  [478,]  47.7  126
#>  [479,]  47.8  126
#>  [480,]  47.9  125
#>  [481,]  48.0  125
#>  [482,]  48.1  125
#>  [483,]  48.2  125
#>  [484,]  48.3  125
#>  [485,]  48.4  125
#>  [486,]  48.5  125
#>  [487,]  48.6  125
#>  [488,]  48.7  124
#>  [489,]  48.8  124
#>  [490,]  48.9  124
#>  [491,]  49.0  124
#>  [492,]  49.1  124
#>  [493,]  49.2  124
#>  [494,]  49.3  124
#>  [495,]  49.4  124
#>  [496,]  49.5  124
#>  [497,]  49.6  124
#>  [498,]  49.7  125
#>  [499,]  49.8  125
#>  [500,]  49.9  125
#>  [ reached getOption("max.print") -- omitted 650 rows ]
#> 
#> $`7_OSL (NA)`
#>      [,1] [,2]
#> [1,]    0   40
#> [2,]   50   40
#> 
#> $`8_OSL (NA)`
#>          [,1] [,2]
#>   [1,]  0.497 39.8
#>   [2,]  0.997 39.8
#>   [3,]  1.497 40.0
#>   [4,]  1.997 40.1
#>   [5,]  2.497 40.0
#>   [6,]  2.997 40.0
#>   [7,]  3.497 40.0
#>   [8,]  3.997 40.1
#>   [9,]  4.497 39.9
#>  [10,]  4.997 40.0
#>  [11,]  5.497 40.2
#>  [12,]  5.997 40.0
#>  [13,]  6.497 39.9
#>  [14,]  6.997 40.1
#>  [15,]  7.497 40.1
#>  [16,]  7.997 39.9
#>  [17,]  8.497 40.1
#>  [18,]  8.997 40.1
#>  [19,]  9.497 40.0
#>  [20,]  9.997 40.0
#>  [21,] 10.497 40.0
#>  [22,] 10.997 40.0
#>  [23,] 11.497 40.0
#>  [24,] 11.997 39.9
#>  [25,] 12.497 39.9
#>  [26,] 12.997 40.0
#>  [27,] 13.497 39.9
#>  [28,] 13.997 40.0
#>  [29,] 14.497 40.0
#>  [30,] 14.997 40.0
#>  [31,] 15.497 40.0
#>  [32,] 15.997 40.0
#>  [33,] 16.497 40.0
#>  [34,] 16.997 40.0
#>  [35,] 17.497 40.0
#>  [36,] 17.997 40.0
#>  [37,] 18.497 39.9
#>  [38,] 18.997 40.0
#>  [39,] 19.497 40.0
#>  [40,] 19.997 40.0
#>  [41,] 20.497 40.0
#>  [42,] 20.997 40.1
#>  [43,] 21.497 40.0
#>  [44,] 21.997 39.9
#>  [45,] 22.497 40.0
#>  [46,] 22.997 40.1
#>  [47,] 23.497 40.0
#>  [48,] 23.997 40.0
#>  [49,] 24.497 40.0
#>  [50,] 24.997 39.9
#>  [51,] 25.497 40.0
#>  [52,] 25.997 40.0
#>  [53,] 26.497 39.9
#>  [54,] 26.997 39.9
#>  [55,] 27.497 40.0
#>  [56,] 27.997 40.0
#>  [57,] 28.497 40.0
#>  [58,] 28.997 39.9
#>  [59,] 29.497 40.0
#>  [60,] 29.997 40.0
#>  [61,] 30.497 39.9
#>  [62,] 30.997 40.0
#>  [63,] 31.497 39.9
#>  [64,] 31.997 40.0
#>  [65,] 32.497 40.0
#>  [66,] 32.997 40.0
#>  [67,] 33.497 40.1
#>  [68,] 33.997 40.1
#>  [69,] 34.497 40.0
#>  [70,] 34.997 40.0
#>  [71,] 35.497 40.1
#>  [72,] 35.997 40.0
#>  [73,] 36.497 40.1
#>  [74,] 36.997 40.1
#>  [75,] 37.497 39.9
#>  [76,] 37.997 39.9
#>  [77,] 38.497 40.1
#>  [78,] 38.997 40.1
#>  [79,] 39.497 40.0
#>  [80,] 39.997 40.0
#>  [81,] 40.497 40.0
#>  [82,] 40.997 40.0
#>  [83,] 41.497 39.9
#>  [84,] 41.997 39.9
#>  [85,] 42.497 40.0
#>  [86,] 42.997 40.0
#>  [87,] 43.497 39.9
#>  [88,] 43.997 39.9
#>  [89,] 44.497 39.9
#>  [90,] 44.997 39.9
#>  [91,] 45.497 40.0
#>  [92,] 45.997 40.0
#>  [93,] 46.497 40.0
#>  [94,] 46.997 40.1
#>  [95,] 47.497 40.1
#>  [96,] 47.997 40.0
#>  [97,] 48.497 40.0
#>  [98,] 48.997 40.0
#>  [99,] 49.497 39.9
#> [100,] 49.997 39.9
#> [101,] 50.497 40.0
#> [102,] 50.997 20.0
#> [103,] 51.497  0.0
#> [104,] 51.997  0.0
#> [105,] 52.497  0.0
#> [106,] 52.997  0.0
#> [107,] 53.497  0.0
#> [108,] 53.997  0.0
#> [109,] 54.497  0.0
#> [110,] 54.997  0.0
#> [111,] 55.497  0.0
#> [112,] 55.997  0.0
#> [113,] 56.497  0.0
#> [114,] 56.997  0.0
#> [115,] 57.497  0.0
#> [116,] 57.997  0.0
#> [117,] 58.497  0.0
#> [118,] 58.997  0.0
#> [119,] 59.497  0.0
#> [120,] 59.997  0.0
#> [121,] 60.497  0.0
#> [122,] 60.997  0.0
#> [123,] 61.497  0.0
#> [124,] 61.997  0.0
#> [125,] 62.497  0.0
#> [126,] 62.997  0.0
#> [127,] 63.497  0.0
#> [128,] 63.997  0.0
#> [129,] 64.497  0.0
#> [130,] 64.997  0.0
#> [131,] 65.497  0.0
#> [132,] 65.997  0.0
#> [133,] 66.497  0.0
#> [134,] 66.997  0.0
#> [135,] 67.497  0.0
#> [136,] 67.997  0.0
#> [137,] 68.497  0.0
#> [138,] 68.997  0.0
#> [139,] 69.497  0.0
#> [140,] 69.997  0.0
#> [141,] 70.497  0.0
#> [142,] 70.997  0.0
#> [143,] 71.497  0.0
#> [144,] 71.997  0.0
#> [145,] 72.497  0.0
#> [146,] 72.997  0.0
#> [147,] 73.497  0.0
#> [148,] 73.997  0.0
#> [149,] 74.497  0.0
#> [150,] 74.997  0.0
#> 
#> $`9_irradiation (NA)`
#>      [,1] [,2]
#> [1,]    0    1
#> [2,]   15    1
#> 
#> $`10_TL (UVVIS)`
#>        [,1] [,2]
#>   [1,]  0.1   46
#>   [2,]  0.2   47
#>   [3,]  0.3   35
#>   [4,]  0.4   34
#>   [5,]  0.5   30
#>   [6,]  0.6   36
#>   [7,]  0.7   40
#>   [8,]  0.8   39
#>   [9,]  0.9   37
#>  [10,]  1.0   27
#>  [11,]  1.1   36
#>  [12,]  1.2   38
#>  [13,]  1.3   35
#>  [14,]  1.4   39
#>  [15,]  1.5   34
#>  [16,]  1.6   41
#>  [17,]  1.7   31
#>  [18,]  1.8   36
#>  [19,]  1.9   43
#>  [20,]  2.0   26
#>  [21,]  2.1   32
#>  [22,]  2.2   36
#>  [23,]  2.3   46
#>  [24,]  2.4   32
#>  [25,]  2.5   31
#>  [26,]  2.6   53
#>  [27,]  2.7   35
#>  [28,]  2.8   24
#>  [29,]  2.9   26
#>  [30,]  3.0   40
#>  [31,]  3.1   35
#>  [32,]  3.2   40
#>  [33,]  3.3   42
#>  [34,]  3.4   30
#>  [35,]  3.5   49
#>  [36,]  3.6   38
#>  [37,]  3.7   34
#>  [38,]  3.8   40
#>  [39,]  3.9   26
#>  [40,]  4.0   27
#>  [41,]  4.1   42
#>  [42,]  4.2   23
#>  [43,]  4.3   32
#>  [44,]  4.4   25
#>  [45,]  4.5   35
#>  [46,]  4.6   40
#>  [47,]  4.7   41
#>  [48,]  4.8   42
#>  [49,]  4.9   34
#>  [50,]  5.0   32
#>  [51,]  5.1   44
#>  [52,]  5.2   36
#>  [53,]  5.3   32
#>  [54,]  5.4   26
#>  [55,]  5.5   27
#>  [56,]  5.6   33
#>  [57,]  5.7   48
#>  [58,]  5.8   35
#>  [59,]  5.9   30
#>  [60,]  6.0   33
#>  [61,]  6.1   34
#>  [62,]  6.2   36
#>  [63,]  6.3   27
#>  [64,]  6.4   32
#>  [65,]  6.5   42
#>  [66,]  6.6   26
#>  [67,]  6.7   35
#>  [68,]  6.8   28
#>  [69,]  6.9   32
#>  [70,]  7.0   21
#>  [71,]  7.1   27
#>  [72,]  7.2   27
#>  [73,]  7.3   25
#>  [74,]  7.4   32
#>  [75,]  7.5   34
#>  [76,]  7.6   28
#>  [77,]  7.7   18
#>  [78,]  7.8   20
#>  [79,]  7.9   20
#>  [80,]  8.0   36
#>  [81,]  8.1   29
#>  [82,]  8.2   26
#>  [83,]  8.3   30
#>  [84,]  8.4   31
#>  [85,]  8.5   30
#>  [86,]  8.6   20
#>  [87,]  8.7   25
#>  [88,]  8.8   23
#>  [89,]  8.9   28
#>  [90,]  9.0   24
#>  [91,]  9.1   26
#>  [92,]  9.2   25
#>  [93,]  9.3   23
#>  [94,]  9.4   19
#>  [95,]  9.5   24
#>  [96,]  9.6   22
#>  [97,]  9.7   20
#>  [98,]  9.8   28
#>  [99,]  9.9   23
#> [100,] 10.0   24
#> [101,] 10.1   20
#> [102,] 10.2   26
#> [103,] 10.3   24
#> [104,] 10.4   35
#> [105,] 10.5   31
#> [106,] 10.6   29
#> [107,] 10.7   23
#> [108,] 10.8   29
#> [109,] 10.9   24
#> [110,] 11.0   18
#> [111,] 11.1   34
#> [112,] 11.2   12
#> [113,] 11.3   23
#> [114,] 11.4   19
#> [115,] 11.5   33
#> [116,] 11.6   32
#> [117,] 11.7   30
#> [118,] 11.8   22
#> [119,] 11.9   24
#> [120,] 12.0   29
#> [121,] 12.1   33
#> [122,] 12.2   37
#> [123,] 12.3   30
#> [124,] 12.4   30
#> [125,] 12.5   25
#> [126,] 12.6   33
#> [127,] 12.7   21
#> [128,] 12.8   28
#> [129,] 12.9   19
#> [130,] 13.0   25
#> [131,] 13.1   28
#> [132,] 13.2   29
#> [133,] 13.3   22
#> [134,] 13.4   29
#> [135,] 13.5   19
#> [136,] 13.6   13
#> [137,] 13.7   28
#> [138,] 13.8   18
#> [139,] 13.9   24
#> [140,] 14.0   20
#> [141,] 14.1   30
#> [142,] 14.2   37
#> [143,] 14.3   19
#> [144,] 14.4   23
#> [145,] 14.5   21
#> [146,] 14.6   27
#> [147,] 14.7   20
#> [148,] 14.8   21
#> [149,] 14.9   25
#> [150,] 15.0   29
#> [151,] 15.1   28
#> [152,] 15.2   38
#> [153,] 15.3   27
#> [154,] 15.4   36
#> [155,] 15.5   31
#> [156,] 15.6   25
#> [157,] 15.7   19
#> [158,] 15.8   29
#> [159,] 15.9   24
#> [160,] 16.0   17
#> [161,] 16.1   19
#> [162,] 16.2   37
#> [163,] 16.3   33
#> [164,] 16.4   30
#> [165,] 16.5   31
#> [166,] 16.6   34
#> [167,] 16.7   39
#> [168,] 16.8   35
#> [169,] 16.9   20
#> [170,] 17.0   42
#> [171,] 17.1   27
#> [172,] 17.2   30
#> [173,] 17.3   28
#> [174,] 17.4   25
#> [175,] 17.5   21
#> [176,] 17.6   17
#> [177,] 17.7   35
#> [178,] 17.8   34
#> [179,] 17.9   24
#> [180,] 18.0   27
#> [181,] 18.1   29
#> [182,] 18.2   38
#> [183,] 18.3   32
#> [184,] 18.4   40
#> [185,] 18.5   37
#> [186,] 18.6   27
#> [187,] 18.7   36
#> [188,] 18.8   41
#> [189,] 18.9   45
#> [190,] 19.0   53
#> [191,] 19.1   41
#> [192,] 19.2   51
#> [193,] 19.3   49
#> [194,] 19.4   48
#> [195,] 19.5   43
#> [196,] 19.6   40
#> [197,] 19.7   41
#> [198,] 19.8   47
#> [199,] 19.9   44
#> [200,] 20.0   44
#> [201,] 20.1   41
#> [202,] 20.2   43
#> [203,] 20.3   44
#> [204,] 20.4   50
#> [205,] 20.5   46
#> [206,] 20.6   45
#> [207,] 20.7   61
#> [208,] 20.8   43
#> [209,] 20.9   62
#> [210,] 21.0   50
#> [211,] 21.1   48
#> [212,] 21.2   39
#> [213,] 21.3   44
#> [214,] 21.4   68
#> [215,] 21.5   56
#> [216,] 21.6   60
#> [217,] 21.7   56
#> [218,] 21.8   64
#> [219,] 21.9   60
#> [220,] 22.0   65
#> [221,] 22.1   54
#> [222,] 22.2   52
#> [223,] 22.3   58
#> [224,] 22.4   51
#> [225,] 22.5   58
#> [226,] 22.6   55
#> [227,] 22.7   57
#> [228,] 22.8   50
#> [229,] 22.9   59
#> [230,] 23.0   73
#> [231,] 23.1   65
#> [232,] 23.2   69
#> [233,] 23.3   72
#> [234,] 23.4   66
#> [235,] 23.5   63
#> [236,] 23.6   71
#> [237,] 23.7   76
#> [238,] 23.8   76
#> [239,] 23.9   64
#> [240,] 24.0   76
#> [241,] 24.1   78
#> [242,] 24.2   78
#> [243,] 24.3   68
#> [244,] 24.4   76
#> [245,] 24.5   77
#> [246,] 24.6   79
#> [247,] 24.7   86
#> [248,] 24.8   78
#> [249,] 24.9   80
#> [250,] 25.0   97
#> [251,] 25.1   74
#> [252,] 25.2  108
#> [253,] 25.3   76
#> [254,] 25.4   81
#> [255,] 25.5  100
#> [256,] 25.6  101
#> [257,] 25.7  106
#> [258,] 25.8  112
#> [259,] 25.9   92
#> [260,] 26.0  129
#> [261,] 26.1  105
#> [262,] 26.2  121
#> [263,] 26.3   99
#> [264,] 26.4  109
#> [265,] 26.5  129
#> [266,] 26.6  127
#> [267,] 26.7  117
#> [268,] 26.8  119
#> [269,] 26.9  143
#> [270,] 27.0  152
#> [271,] 27.1  143
#> [272,] 27.2  122
#> [273,] 27.3  122
#> [274,] 27.4  128
#> [275,] 27.5  155
#> [276,] 27.6  156
#> [277,] 27.7  140
#> [278,] 27.8  163
#> [279,] 27.9  158
#> [280,] 28.0  172
#> [281,] 28.1  154
#> [282,] 28.2  163
#> [283,] 28.3  161
#> [284,] 28.4  168
#> [285,] 28.5  172
#> [286,] 28.6  184
#> [287,] 28.7  195
#> [288,] 28.8  179
#> [289,] 28.9  181
#> [290,] 29.0  224
#> [291,] 29.1  214
#> [292,] 29.2  187
#> [293,] 29.3  209
#> [294,] 29.4  200
#> [295,] 29.5  211
#> [296,] 29.6  202
#> [297,] 29.7  202
#> [298,] 29.8  210
#> [299,] 29.9  226
#> [300,] 30.0  222
#> [301,] 30.1  243
#> [302,] 30.2  227
#> [303,] 30.3  218
#> [304,] 30.4  229
#> [305,] 30.5  259
#> [306,] 30.6  222
#> [307,] 30.7  269
#> [308,] 30.8  257
#> [309,] 30.9  267
#> [310,] 31.0  244
#> [311,] 31.1  265
#> [312,] 31.2  260
#> [313,] 31.3  301
#> [314,] 31.4  245
#> [315,] 31.5  296
#> [316,] 31.6  280
#> [317,] 31.7  277
#> [318,] 31.8  296
#> [319,] 31.9  308
#> [320,] 32.0  303
#> [321,] 32.1  277
#> [322,] 32.2  304
#> [323,] 32.3  312
#> [324,] 32.4  342
#> [325,] 32.5  282
#> [326,] 32.6  324
#> [327,] 32.7  367
#> [328,] 32.8  336
#> [329,] 32.9  339
#> [330,] 33.0  356
#> [331,] 33.1  355
#> [332,] 33.2  392
#> [333,] 33.3  322
#> [334,] 33.4  323
#> [335,] 33.5  362
#> [336,] 33.6  380
#> [337,] 33.7  403
#> [338,] 33.8  400
#> [339,] 33.9  400
#> [340,] 34.0  416
#> [341,] 34.1  400
#> [342,] 34.2  422
#> [343,] 34.3  439
#> [344,] 34.4  464
#> [345,] 34.5  435
#> [346,] 34.6  432
#> [347,] 34.7  436
#> [348,] 34.8  408
#> [349,] 34.9  406
#> [350,] 35.0  446
#> [351,] 35.1  498
#> [352,] 35.2  414
#> [353,] 35.3  418
#> [354,] 35.4  428
#> [355,] 35.5  462
#> [356,] 35.6  445
#> [357,] 35.7  493
#> [358,] 35.8  446
#> [359,] 35.9  456
#> [360,] 36.0  416
#> [361,] 36.1  457
#> [362,] 36.2  517
#> [363,] 36.3  491
#> [364,] 36.4  463
#> [365,] 36.5  450
#> [366,] 36.6  495
#> [367,] 36.7  447
#> [368,] 36.8  484
#> [369,] 36.9  488
#> [370,] 37.0  498
#> [371,] 37.1  518
#> [372,] 37.2  467
#> [373,] 37.3  506
#> [374,] 37.4  502
#> [375,] 37.5  469
#> [376,] 37.6  530
#> [377,] 37.7  542
#> [378,] 37.8  522
#> [379,] 37.9  482
#> [380,] 38.0  465
#> [381,] 38.1  452
#> [382,] 38.2  469
#> [383,] 38.3  497
#> [384,] 38.4  478
#> [385,] 38.5  469
#> [386,] 38.6  523
#> [387,] 38.7  484
#> [388,] 38.8  468
#> [389,] 38.9  488
#> [390,] 39.0  555
#> [391,] 39.1  469
#> [392,] 39.2  455
#> [393,] 39.3  484
#> [394,] 39.4  493
#> [395,] 39.5  440
#> [396,] 39.6  430
#> [397,] 39.7  464
#> [398,] 39.8  438
#> [399,] 39.9  462
#> [400,] 40.0  427
#> [401,] 40.1  415
#> [402,] 40.2  395
#> [403,] 40.3  439
#> [404,] 40.4  424
#> [405,] 40.5  401
#> [406,] 40.6  421
#> [407,] 40.7  390
#> [408,] 40.8  396
#> [409,] 40.9  404
#> [410,] 41.0  363
#> [411,] 41.1  326
#> [412,] 41.2  344
#> [413,] 41.3  372
#> [414,] 41.4  342
#> [415,] 41.5  370
#> [416,] 41.6  358
#> [417,] 41.7  339
#> [418,] 41.8  326
#> [419,] 41.9  310
#> [420,] 42.0  271
#> [421,] 42.1  281
#> [422,] 42.2  297
#> [423,] 42.3  318
#> [424,] 42.4  311
#> [425,] 42.5  245
#> [426,] 42.6  267
#> [427,] 42.7  227
#> [428,] 42.8  226
#> [429,] 42.9  225
#> [430,] 43.0  192
#> [431,] 43.1  218
#> [432,] 43.2  226
#> [433,] 43.3  217
#> [434,] 43.4  200
#> [435,] 43.5  200
#> [436,] 43.6  194
#> [437,] 43.7  181
#> [438,] 43.8  180
#> [439,] 43.9  152
#> [440,] 44.0  187
#> [441,] 44.1  152
#> [442,] 44.2  158
#> [443,] 44.3  162
#> [444,] 44.4  131
#> [445,] 44.5  149
#> [446,] 44.6  115
#> [447,] 44.7  119
#> [448,] 44.8  112
#> [449,] 44.9  105
#> [450,] 45.0  107
#> [451,] 45.1   82
#> [452,] 45.2  114
#> [453,] 45.3   87
#> [454,] 45.4   87
#> [455,] 45.5   98
#> [456,] 45.6   76
#> [457,] 45.7   94
#> [458,] 45.8   80
#> [459,] 45.9   85
#> [460,] 46.0   65
#> [461,] 46.1   76
#> [462,] 46.2   71
#> [463,] 46.3   54
#> [464,] 46.4   53
#> [465,] 46.5   57
#> [466,] 46.6   49
#> [467,] 46.7   46
#> [468,] 46.8   35
#> [469,] 46.9   39
#> [470,] 47.0   37
#> [471,] 47.1   36
#> [472,] 47.2   41
#> [473,] 47.3   36
#> [474,] 47.4   25
#> [475,] 47.5   35
#> [476,] 47.6   28
#> [477,] 47.7   46
#> [478,] 47.8   27
#> [479,] 47.9   26
#> [480,] 48.0   25
#> [481,] 48.1   20
#> [482,] 48.2   25
#> [483,] 48.3   31
#> [484,] 48.4   29
#> [485,] 48.5   32
#> [486,] 48.6   38
#> [487,] 48.7   26
#> [488,] 48.8   19
#> [489,] 48.9   28
#> [490,] 49.0   24
#> [491,] 49.1   27
#> [492,] 49.2   32
#> [493,] 49.3   19
#> [494,] 49.4   11
#> [495,] 49.5   23
#> [496,] 49.6   14
#> [497,] 49.7   27
#> [498,] 49.8   25
#> [499,] 49.9   14
#> [500,] 50.0   17
#>  [ reached getOption("max.print") -- omitted 180 rows ]
#> 

# NOT RUN {
##select your BIN-file
file <- file.choose()

##convert
convert_XSYG2CSV(file)

# }