Luminescence Emission Spectra Deconvolution

Luminescence spectra deconvolution on RLum.Data.Spectrum and matrix objects on an energy scale. The function is optimised for emission spectra typially obtained in the context of TL, OSL and RF measurements detected between 200 and 1000 nm.

fit_EmissionSpectra(object, frame = NULL, start_parameters = NULL,
  sub_negative = 0, input_scale = NULL, method_control = list(),
  verbose = TRUE, plot = TRUE, ...)

Arguments

object
RLum.Data.Spectrum, matrix (required): input object. Please note that an energy spectrum is expected

frame
numeric (optional): defines the frame to be analysed

start_parameters
(optional): allows to provide own start parameters for a semi-automated precedure. ##TODO

sub_negative
numeric (with default): substitute negative values in the input object by the number provided here (default: 0). Can be set to NULL, i.e. negative values are kept.

input_scale
character (optional): defines whether your x-values define wavelength or energy values. For the analysis an energy scale is expected, allowed values are 'wavelength' and 'energy'. If nothing (NULL) is defined, the function tries to understand the input automatically.

method_control
list (optional): options to control the fit method, see details

verbose
logical (with default): enable/disable verbose mode

plot
logical (with default): enable/disable plot output

...
further arguments to be passed to control the plot output (supported: main, xlab, ylab, xlim, ylim, log, mtext, legend (TRUE or FALSE), legend.text, legend.pos)

Value

-----------------------------------
[ NUMERICAL OUTPUT ]
-----------------------------------

RLum.Results-object

slot: @data

Element	Type	Description

$data	matrix	the final fit matrix

$fit	nls	the fit object returned by minpack.lm::nls.lm

slot: @info

The original function call

------------------------
[ TERMINAL OUTPUT ]
------------------------

The terminal output provides brief information on the deconvolution process and the obtained results. Terminal output is only shown of the argument verbose = TRUE.

------------------------
[ PLOT OUTPUT ]
------------------------

The function returns a plot showing the raw signal with the detected components. If the fitting failed, a basic plot is returned showing the raw data and indicating the peaks detected for the start parameter estimation.

Details

Used equation

The emission spectra (on an energy scale) can be best described as the sum of multiple Gaussian components:

'$$ y = \Sigma Ci * 1/(\sigma_{i} * \sqrt(2 * \pi)) * exp(1/2 * ((x - \mu_{i})/\sigma_{i}))^2) $$

with the parameters \(\sigma\) (peak width) and \(\mu\) (peak centre) and \(C\) (scaling factor).

Start parameter estimation and fitting algorithm

The spectrum deconvolution consits of the following steps:

Peak finding

Start parameter estimation

Fitting via minpack.lm::nls.lm

The peak finding is realised by an approach (re-)suggested by Petr Pikal via the R-help mailing list (https://stat.ethz.ch/pipermail/r-help/2005-November/thread.html) in November 2005. This goes back to even earlier discussion in 2001 based on Prof Brian Ripley's idea. It smartly uses the functions stats::embed and max.col to identify peaks positions. For the use in this context, the algorithm has been further modified to scale on the input data resolution (cv source code).

The start parameter estimation uses random sampling from a range of meaningful parameters and repeats the fitting until 100 sucessful fits have been produced or the set max.runs value is exceeded.

Currently the best fit is the one with the lowest number for squared residuals.

Supported method_control settings

Parameter	Type	Default	Descritpion

max.runs	integer	1000	maximum allowed search iterations, if exceed the searching stops

trace	logical	FALSE	enables/disables the tracing of the minimisation routine

Note

Beta version, not uet recommended for productive usage

Function version

0.1.0

How to cite

Kreutzer, S. (2019). fit_EmissionSpectra(): Luminescence Emission Spectra Deconvolution. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data AnalysisR package version 0.9.1.9000-11. https://CRAN.R-project.org/package=Luminescence

References

##TODO

See also

Examples

##deconvolution of a TL spectrum
##TODO should be modified ...

# NOT RUN {
##load example data
data(ExampleData.XSYG, envir = environment())

##subtract background
TL.Spectrum@data <- TL.Spectrum@data[] - TL.Spectrum@data[,15]

##replace 0 values
results <- fit_EmissionSpectra(
 object = TL.Spectrum,
 frame = 5, main = "TL spectrum"
)

# }