General accessor function for RLum S4 class objects

Function calls object-specific get functions for RLum S4 class objects.

Method to handle NULL if the user calls get_RLum

get_RLum(object, ...)

# S4 method for list
get_RLum(object, null.rm = FALSE, ...)

# S4 method for `NULL`
get_RLum(object, ...)

Arguments

object
RLum (required): S4 object of class RLum or an object of type list containing only objects of type RLum

...
further arguments that will be passed to the object specific methods. For furter details on the supported arguments please see the class documentation: RLum.Data.Curve, RLum.Data.Spectrum, RLum.Data.Image, RLum.Analysis and RLum.Results

null.rm
logical (with default): option to get rid of empty and NULL objects

Value

Return is the same as input objects as provided in the list.

Details

The function provides a generalised access point for specific RLum objects. Depending on the input object, the corresponding get function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class.

Methods (by class)

list: Returns a list of RLum objects that had been passed to get_RLum

NULL: Returns NULL

Function version

0.3.2 (2018-01-21 17:22:38)

See also

RLum.Data.Curve, RLum.Data.Image, RLum.Data.Spectrum, RLum.Analysis, RLum.Results

How to cite

Kreutzer, S. (2018). get_RLum(): General accessor function for RLum S4 class objects. Function version 0.3.2. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.1. https://CRAN.R-project.org/package=Luminescence

Examples

##Example based using data and from the calc_CentralDose() function

##load example data
data(ExampleData.DeValues, envir = environment())

##apply the central dose model 1st time
temp1 <- calc_CentralDose(ExampleData.DeValues$CA1)

#> 
#>  [calc_CentralDose]
#> 
#> ----------- meta data ----------------
#>  n:                       62
#>  log:                     TRUE
#> ----------- dose estimate ------------
#>  central dose [Gy]:       65.71
#>  SE [Gy]:                 3.05
#>  rel. SE [%]:             4.65
#> ----------- overdispersion -----------
#>  OD [Gy]:                 22.79
#>  SE [Gy]:                 2.27
#>  OD [%]:                  34.69
#>  SE [%]:                  3.46
#> -------------------------------------
#> 

##get results and store them in a new object
temp.get <- get_RLum(object = temp1)