General accessor function for RLum S4 class objects

Function calls object-specific get functions for RisoeBINfileData S4 class objects.

Usage

get_Risoe.BINfileData(object, ...)

Arguments

object

Risoe.BINfileData (required): S4 object of class RLum

...

further arguments that one might want to pass to the specific get function

Value

Return is the same as input objects as provided in the list

Details

The function provides a generalised access point for specific Risoe.BINfileData objects.
Depending on the input object, the corresponding get function will be selected. Allowed arguments can be found in the documentations of the corresponding Risoe.BINfileData class.

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)