Merge function for RLum.Results S4-class objects

Function merges objects of class RLum.Results. The slots in the objects are combined depending on the object type, e.g., for data.frame and matrix rows are appended.

Usage

merge_RLum.Results(objects)

Arguments

objects

list (required): a list of RLum.Results objects

Details

Elements are appended where possible and attributes are preserved if not of similar name as the default attributes of, e.g., a data.frame

Note

The originator is taken from the first element and not reset to merge_RLum

Function version

0.2.1

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany) , RLum Developer Team

How to cite

Kreutzer, S., 2024. merge_RLum.Results(): Merge function for RLum.Results S4-class objects. Function version 0.2.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., 2024. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.25. https://r-lum.github.io/Luminescence/