plot_RLum {Luminescence}	R Documentation

General plot function for RLum S4 class objects

Description

Function calls object specific plot functions for RLum S4 class objects.

Usage


plot_RLum(object, ...)

Arguments

object
RLum (required): S4 object of class RLum. Optional a list containing objects of class RLum can be provided. In this case the function tries to plot every object in this list according to its RLum class. Non-RLum objects are removed.

...
further arguments and graphical parameters that will be passed to the specific plot functions. The only argument that is supported directly is main (setting the plot title). In contrast to the normal behaviour main can be here provided as list and the arguments in the list will dispatched to the plots if the object is of type list as well.

Details

The function provides a generalised access point for plotting specific RLum objects.
Depending on the input object, the corresponding plot function will be selected. Allowed arguments can be found in the documentations of each plot function.

object 	 	 corresponding plot function

RLum.Data.Curve 	 : 	 plot_RLum.Data.Curve

RLum.Data.Spectrum 	 : 	 plot_RLum.Data.Spectrum

RLum.Data.Image 	 : 	 plot_RLum.Data.Image

RLum.Analysis 	 : 	 plot_RLum.Analysis

RLum.Results 	 : 	 plot_RLum.Results

Value

Returns a plot.

Function version

0.4.4

How to cite

Kreutzer, S., 2021. plot_RLum(): General plot function for RLum S4 class objects. Function version 0.4.4. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

The provided plot output depends on the input object.

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

plot_RLum.Data.Curve, RLum.Data.Curve, plot_RLum.Data.Spectrum, RLum.Data.Spectrum, plot_RLum.Data.Image, RLum.Data.Image, plot_RLum.Analysis, RLum.Analysis, plot_RLum.Results, RLum.Results

Examples


#load Example data
data(ExampleData.CW_OSL_Curve, envir = environment())

#transform data.frame to RLum.Data.Curve object
temp <- as(ExampleData.CW_OSL_Curve, "RLum.Data.Curve")

#plot RLum object
plot_RLum(temp)


[Package Luminescence version 0.9.15 Index]