plot_RLum.Data.Image {Luminescence}	R Documentation

Plot function for an RLum.Data.Image S4 class object

Description

The function provides a standardised plot output for image data of an RLum.Data.ImageS4 class object, mainly using the plot functions provided by the raster package.

Details on the plot functions

Image is visualised as 2D plot using generic plot types provided by other packages.

Supported plot types:

plot.type = "plot.raster"

Uses the standard plot function for raster data from the package raster::raster: raster::plot. For each raster layer in a raster brick one plot is produced.

Arguments that are passed through the function call:

main,axes, xlab, ylab, xlim, ylim,zlim, col, stretch (NULL no strech)

plot.type = "plotRGB"

Uses the function raster::plotRGB from the raster::raster package. Only one image plot is produced as all layers in a brick a combined. This plot type is useful to see whether any signal is recorded by the camera.
Arguments that are passed through the function call:

main,axes, xlab, ylab, ext, interpolate, maxpixels, alpha, colNA, stretch

plot.type = "contour"

Uses the function contour plot function from the raster function (raster::contour). For each raster layer one contour plot is produced. Arguments that are passed through the function call:

main,axes, xlab, ylab, xlim, ylim, col

Usage


plot_RLum.Data.Image(object, par.local = TRUE, plot.type = "plot.raster", ...)

Arguments

object
RLum.Data.Image (required): S4 object of class RLum.Data.Image

par.local
logical (with default): use local graphical parameters for plotting, e.g. the plot is shown in one column and one row. If par.local = FALSE global parameters are inherited.

plot.type
character (with default): plot types. Supported types are plot.raster, plotRGB or contour

...
further arguments and graphical parameters that will be passed to the specific plot functions.

Value

Returns a plot.

Function version

0.1.1

How to cite

Kreutzer, S., 2021. plot_RLum.Data.Image(): Plot function for an RLum.Data.Image S4 class object. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Data.Image, plot, plot_RLum, raster::raster

Examples



##load data
data(ExampleData.RLum.Data.Image, envir = environment())

##plot data
plot_RLum.Data.Image(ExampleData.RLum.Data.Image)


[Package Luminescence version 0.9.15 Index]