General replication function for RLum S4 class objects

Function replicates RLum S4 class objects and returns a list for this objects

replicate_RLum(object, times = NULL)

Arguments

object
RLum (required): an RLum object

times
integer (optional): number for times each element is repeated element

Value

Returns a list of the object to be repeated

Function version

0.1.0

See also

How to cite

Kreutzer, S. (2019). replicate_RLum(): General replication function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data AnalysisR package version 0.9.1.9000-11. https://CRAN.R-project.org/package=Luminescence