General setter function for Risoe.BINfileData objects

The function provides a generalised access point for specific Risoe.BINfileData objects. Depending on the input object, the corresponding function will be selected. Allowed arguments can be found in the documentations of the corresponding Risoe.BINfileData class.

Usage

set_Risoe.BINfileData(
  METADATA = data.frame(),
  DATA = list(),
  .RESERVED = list()
)

Arguments

METADATA

x

DATA

x

.RESERVED

x

Value

A Risoe.BINfileData object.

Function version

0.1

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany) , RLum Developer Team

How to cite

Kreutzer, S., 2025. set_Risoe.BINfileData(): General setter function for Risoe.BINfileData objects. Function version 0.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., Steinbuch, L., Boer, A.d., 2025. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 1.1.0. https://r-lum.github.io/Luminescence/