tune_Data {Luminescence}	R Documentation

Tune data for experimental purpose

Description

The error can be reduced and sample size increased for specific purpose.

Usage


tune_Data(data, decrease.error = 0, increase.data = 0)

Arguments

data
data.frame (required): input values, structure: data (values[,1]) and data error (values [,2]) are required

decrease.error
numeric: factor by which the error is decreased, ranges between 0 and 1.

increase.data
numeric: factor by which the error is decreased, ranges between 0 and Inf.

Value

Returns a data.frame with tuned values.

Function version

0.5.0

How to cite

Dietze, M., 2021. tune_Data(): Tune data for experimental purpose. Function version 0.5.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

You should not use this function to improve your poor data set!

Author(s)

Michael Dietze, GFZ Potsdam (Germany) , RLum Developer Team

Examples



## load example data set
data(ExampleData.DeValues, envir = environment())
x <- ExampleData.DeValues$CA1

## plot original data
plot_AbanicoPlot(data = x,
                 summary = c("n", "mean"))

## decrease error by 10 %
plot_AbanicoPlot(data = tune_Data(x, decrease.error = 0.1),
                 summary = c("n", "mean"))

## increase sample size by 200 %
#plot_AbanicoPlot(data = tune_Data(x, increase.data = 2) ,
#                summary = c("n", "mean"))


[Package Luminescence version 0.9.15 Index]