BaseDataSet.ConversionFactors {Luminescence}	R Documentation

Base data set of dose-rate conversion factors

Description

Collection of published dose-rate conversion factors to convert concentrations of radioactive isotopes to dose rate values.

Format

A list with three elements with dose-rate conversion factors sorted by article and radiation type (alpha, beta, gamma):

AdamiecAitken1998: 	 Conversion factors from Tables 5 and 6

Cresswelletal2018: 	 Conversion factors from Tables 5 and 6

Guerinetal2011: 	 Conversion factors from Tables 1, 2 and 3

Liritzisetal2013: 	 Conversion factors from Tables 1, 2 and 3

Version

0.2.0

Source

All gamma conversion factors were carefully read from the tables given in the references above.

References

Adamiec, G., Aitken, M.J., 1998. Dose-rate conversion factors: update. Ancient TL 16, 37-46.

Cresswell., A.J., Carter, J., Sanderson, D.C.W., 2018. Dose rate conversion parameters: Assessment of nuclear data. Radiation Measurements 120, 195-201.

Guerin, G., Mercier, N., Adamiec, G., 2011. Dose-rate conversion factors: update. Ancient TL, 29, 5-8.

Liritzis, I., Stamoulis, K., Papachristodoulou, C., Ioannides, K., 2013. A re-evaluation of radiation dose-rate conversion factors. Mediterranean Archaeology and Archaeometry 13, 1-15.

Examples



## Load data
data("BaseDataSet.ConversionFactors", envir = environment())


[Package Luminescence version 0.9.15 Index]