ExampleData.BINfileData {Luminescence}	R Documentation

Example data from a SAR OSL and SAR TL measurement for the package Luminescence

Description

Example data from a SAR OSL and TL measurement for package Luminescence directly extracted from a Risoe BIN-file and provided in an object of type Risoe.BINfileData

Format

CWOSL.SAR.Data: SAR OSL measurement data

TL.SAR.Data: SAR TL measurement data

Each class object contains two slots: (a) METADATA is a data.frame with all metadata stored in the BIN file of the measurements and (b) DATA contains a list of vectors of the measured data (usually count values).

Version

0.1

Note

Please note that this example data cannot be exported to a BIN-file using the function writeR2BIN as it was generated and implemented in the package long time ago. In the meantime the BIN-file format changed.

Source

CWOSL.SAR.Data

Lab: 	 Luminescence Laboratory Bayreuth

Lab-Code: 	 BT607

Location: 	 Saxony/Germany

Material: 	 Middle grain quartz measured on aluminium cups on a Risø TL/OSL DA-15 reader

Reference: 	 unpublished

TL.SAR.Data

Lab: 	 Luminescence Laboratory of Cologne

Lab-Code: 	 LP1_5

Location: 	 Spain

Material: 	 Flint

Setup: 	 Risoe TL/OSL DA-20 reader (Filter: Semrock Brightline, HC475/50, N2, unpolished steel discs)

Reference: 	 unpublished

Remarks: 	 dataset limited to one position

References

CWOSL.SAR.Data: unpublished data

TL.SAR.Data: unpublished data

Examples



## show first 5 elements of the METADATA and DATA elements in the terminal
data(ExampleData.BINfileData, envir = environment())
CWOSL.SAR.Data@METADATA[1:5,]
CWOSL.SAR.Data@DATA[1:5]


[Package Luminescence version 0.9.15 Index]