ExampleData.CobbleData {Luminescence}	R Documentation

Example data for calc_CobbleDoseRate()

Description

An example data set for the function calc_CobbleDoseRate containing layer specific information for the cobble to be used in the function.

Format

A data.frame. Please see calc_CobbleDoseRate for detailed information on the structure of the data.frame.

Version

0.1.0

Examples



## Load data
data("ExampleData.CobbleData", envir = environment())


[Package Luminescence version 0.9.15 Index]