ExampleData.Fading {Luminescence}	R Documentation

Example data for feldspar fading measurements

Description

Example data set for fading measurements of the IR50, IR100, IR150 and IR225 feldspar signals of sample UNIL/NB123. It further contains regular equivalent dose measurement data of the same sample, which can be used to apply a fading correction to.

Format

A list with two elements, each containing a further list of data.frames containing the data on the fading and equivalent dose measurements:

$fading.data: A named list of data.frames, each having three named columns (LxTx, LxTx.error, timeSinceIrradiation).
..$IR50: Fading data of the IR50 signal.
..$IR100: Fading data of the IR100 signal.
..$IR150: Fading data of the IR150 signal.
..$IR225: Fading data of the IR225 signal.

$equivalentDose.data: A named of data.frames, each having three named columns (dose, LxTx, LxTx.error).
..$IR50: Equivalent dose measurement data of the IR50 signal.
..$IR100: Equivalent dose measurement data of the IR100 signal.
..$IR150: Equivalent dose measurement data of the IR150 signal.
..$IR225: Equivalent dose measurement data of the IR225 signal.

Source

These data were kindly provided by Georgina E. King. Detailed information on the sample UNIL/NB123 can be found in the reference given below. The raw data can be found in the accompanying supplementary information.

References

King, G.E., Herman, F., Lambert, R., Valla, P.G., Guralnik, B., 2016. Multi-OSL-thermochronometry of feldspar. Quaternary Geochronology 33, 76-87. doi:10.1016/j.quageo.2016.01.004

Details

Lab: 	 University of Lausanne

Lab-Code: 	 UNIL/NB123

Location: 	 Namche Barwa (eastern Himalayas)

Material: 	 Coarse grained (180-212 microns) potassium feldspar

Units: 	 Values are given in seconds

Lab Dose Rate: 	 Dose rate of the beta-source at measurement ca. 0.1335 +/- 0.004 Gy/s

Environmental Dose Rate: 	 7.00 +/- 0.92 Gy/ka (includes internal dose rate)

Examples



## Load example data
data("ExampleData.Fading", envir = environment())

## Get fading measurement data of the IR50 signal
IR50_fading <- ExampleData.Fading$fading.data$IR50
head(IR50_fading)

## Determine g-value and rho' for the IR50 signal
IR50_fading.res <- analyse_FadingMeasurement(IR50_fading)

## Show g-value and rho' results
gval <- get_RLum(IR50_fading.res)
rhop <- get_RLum(IR50_fading.res, "rho_prime")

gval
rhop

## Get LxTx values of the IR50 DE measurement
IR50_De.LxTx <- ExampleData.Fading$equivalentDose.data$IR50

## Calculate the De of the IR50 signal
IR50_De <- plot_GrowthCurve(IR50_De.LxTx,
                                mode = "interpolation",
                                fit.method = "EXP")

## Extract the calculated De and its error
IR50_De.res <- get_RLum(IR50_De)
De <- c(IR50_De.res$De, IR50_De.res$De.Error)

## Apply fading correction (age conversion greatly simplified)
IR50_Age <- De / 7.00
IR50_Age.corr <- calc_FadingCorr(IR50_Age, g_value = IR50_fading.res)



[Package Luminescence version 0.9.15 Index]