GitHub API

R Interface to the GitHub API v3.

github_commits(user = "r-lum", repo = "luminescence",
  branch = "master", n = 5)

github_branches(user = "r-lum", repo = "luminescence")

github_issues(user = "r-lum", repo = "luminescence", verbose = TRUE)

Arguments

user
character (with default): GitHub user name (defaults to 'r-lum').

repo
character (with default): name of a GitHub repository (defaults to 'luminescence').

branch
character (with default): branch of a GitHub repository (defaults to 'master').

n
integer (with default): number of commits returned (defaults to 5).

verbose
logical (with default): print the output to the console (defaults to TRUE).

Value

github_commits: data.frame with columns:

[ ,1]	SHA

[ ,2]	AUTHOR

[ ,3]	DATE

[ ,4]	MESSAGE

github_branches: data.frame with columns:

[ ,1]	BRANCH

[ ,2]	SHA

[ ,3]	INSTALL

github_commits: Nested list with n elements. Each commit element is a list with elements:

[[1]]	NUMBER

[[2]]	TITLE

[[3]]	BODY

[[4]]	CREATED

[[5]]	UPDATED

[[6]]	CREATOR

[[7]]	URL

[[8]]	STATUS

Details

These functions can be used to query a specific repository hosted on GitHub.

github_commits lists the most recent n commits of a specific branch of a repository.

github_branches can be used to list all current branches of a repository and returns the corresponding SHA hash as well as an installation command to install the branch in R via the 'devtools' package.

github_issues lists all open issues for a repository in valid YAML.

Function version

0.1.0

How to cite

Burow, C. (2019). GitHub-API(): GitHub API. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data AnalysisR package version 0.9.1.9000-11. https://CRAN.R-project.org/package=Luminescence

References

GitHub Developer API v3. https://developer.github.com/v3/, last accessed: 10/01/2017.

Examples

# NOT RUN {
github_branches(user = "r-lum", repo = "luminescence")
github_issues(user = "r-lum", repo = "luminescence")
github_commits(user = "r-lum", repo = "luminescence", branch = "master", n = 10)
# }