Comprehensive Luminescence Dating Data Analysis

A collection of various R functions for the purpose of luminescence dating data analysis. This includes, amongst others, data import, export, application of age models, curve deconvolution, sequence analysis and plotting of equivalent dose distributions.

Details

Supervisor of the initial version in 2012

Markus Fuchs, Justus-Liebig-University Giessen, Germany

Support contact

developers@r-luminescence.org

https://github.com/R-Lum/Luminescence/discussions

Bug reporting

developers@r-luminescence.org or

https://github.com/R-Lum/Luminescence/issues

Project website

https://r-luminescence.org

Project source code repository

https://github.com/R-Lum/Luminescence

Related package projects

https://cran.r-project.org/package=RLumShiny

https://cran.r-project.org/package=RLumModel

https://cran.r-project.org/package=RLumCarlo

https://cran.r-project.org/package=RCarb

Funding

2011-2013: The initial version of the package was developed, while Sebastian Kreutzer was funded through the DFG programme "Rekonstruktion der Umweltbedingungen des Spätpleistozäns in Mittelsachsen anhand von Löss-Paläobodensequenzen" (DFG id: 46526743)

2014-2018: Cooperation and personal exchange between the developers is gratefully funded by the DFG (SCHM 3051/3-1) in the framework of the program "Scientific Networks". Project title: "RLum.Network: Ein Wissenschaftsnetzwerk zur Analyse von Lumineszenzdaten mit R" (2014-2018)

05/2014-12/2019: The work of Sebastian Kreutzer as maintainer of the package was supported by LabEx LaScArBx (ANR - n. ANR-10-LABX-52).

01/2020-04/2022: Sebastian Kreutzer as maintainer of the package has received funding from the European Union’s Horizon 2020 research and innovation programme under the Marie Skłodowska-Curie grant agreement No 844457 (CREDit), and could continue maintaining the package.

since 03/2023: Sebastian Kreutzer as maintainer of the package receives funding from the DFG Heisenberg programme No 505822867.

All other authors gratefully received additional funding from various public funding bodies.

References

Dietze, M., Kreutzer, S., Fuchs, M.C., Burow, C., Fischer, M., Schmidt, C., 2013. A practical guide to the R package Luminescence. Ancient TL, 31 (1), 11-18.

Dietze, M., Kreutzer, S., Burow, C., Fuchs, M.C., Fischer, M., Schmidt, C., 2016. The abanico plot: visualising chronometric data with individual standard errors. Quaternary Geochronology 31, 1-7. https://doi.org/10.1016/j.quageo.2015.09.003

Fuchs, M.C., Kreutzer, S., Burow, C., Dietze, M., Fischer, M., Schmidt, C., Fuchs, M., 2015. Data processing in luminescence dating analysis: An exemplary workflow using the R package 'Luminescence'. Quaternary International, 362,8-13. https://doi.org/10.1016/j.quaint.2014.06.034

Kreutzer, S., Schmidt, C., Fuchs, M.C., Dietze, M., Fischer, M., Fuchs, M., 2012. Introducing an R package for luminescence dating analysis. Ancient TL, 30 (1), 1-8.

Mercier, N., Kreutzer, S., Christophe, C., Guérin, G., Guibert, P., Lahaye, C., Lanos, P., Philippe, A., Tribolo, C., 2016. Bayesian statistics in luminescence dating: The 'baSAR'-model and its implementation in the R package ’Luminescence’. Ancient TL 34 (2), 14-21.

Mercier, N., Galharret, J.-M., Tribolo, C., Kreutzer, S., Philippe, A., 2022. Luminescence age calculation through Bayesian convolution of equivalent dose and dose-rate distributions: the De_Dr model. Geochronology 4, 297–310. https://doi.org/10.5194/gchron-4-297-2022

Smedley, R.K., 2015. A new R function for the Internal External Uncertainty (IEU) model. Ancient TL, 33 (1), 16-21.

King, E.G., Burow, C., Roberts, H., Pearce, N.J.G., 2018. Age determination using feldspar: evaluating fading-correction model performance. Radiation Measurements 119, 58-73. https://doi.org/10.1016/j.radmeas.2018.07.013

See also

Author

Maintainer: Sebastian Kreutzer maintainer_luminescence@r-luminescence.org (ORCID) [translator, data contributor]

Authors:

Christoph Burow (ORCID) [translator, data contributor]

Michael Dietze (ORCID)

Margret C. Fuchs (ORCID)

Christoph Schmidt (ORCID)

Manfred Fischer [translator]

Johannes Friedrich (ORCID)

Norbert Mercier (ORCID)

Anne Philippe (ORCID)

Svenja Riedesel (ORCID)

Martin Autzen (ORCID)

Dirk Mittelstrass (ORCID)

Harrison J. Gray (ORCID)

Jean-Michel Galharret (ORCID)

Marco Colombo (ORCID)

Other contributors:

Rachel K. Smedley (ORCID) [contributor]

Claire Christophe [contributor]

Antoine Zink (ORCID) [contributor]

Julie Durcan (ORCID) [contributor]

Georgina E. King (ORCID) [contributor, data contributor]

Guillaume Guerin (ORCID) [contributor]

Pierre Guibert (ORCID) [contributor]

Markus Fuchs (ORCID) [thesis advisor]