RLum-class {Luminescence}	R Documentation

Class "RLum"

Description

Abstract class for data in the package Luminescence Subclasses are:

Usage


## S4 method for signature 'RLum'
replicate_RLum(object, times = NULL)

Arguments

object
RLum (required): an object of class RLum

times
integer (optional): number for times each element is repeated element

Details

RLum-class
|
|—-RLum.Data
|—-|– RLum.Data.Curve
|—-|– RLum.Data.Spectrum
|—-|– RLum.Data.Image
|—-RLum.Analysis
|—-RLum.Results

Methods (by generic)

replicate_RLum: Replication method RLum-objects

Slots

originator

Object of class character containing the name of the producing function for the object. Set automatically by using the function set_RLum.

info

Object of class list for additional information on the object itself

.uid

Object of class character for a unique object identifier. This id is usually calculated using the internal function create_UID() if the function set_RLum is called.

.pid

Object of class character for a parent id. This allows nesting RLum-objects at will. The parent id can be the uid of another object.

Objects from the Class

A virtual Class: No objects can be created from it.

Class version

0.4.0

How to cite

Kreutzer, S., 2021. RLum-class(): Class 'RLum'. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

RLum is a virtual class.

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Data, RLum.Data.Curve, RLum.Data.Spectrum, RLum.Data.Image, RLum.Analysis, RLum.Results, methods_RLum

Examples



showClass("RLum")


[Package Luminescence version 0.9.15 Index]