Class "RLum.Data.Image"

Class for representing luminescence image data (TL/OSL/RF). Such data are for example produced by the function read_SPE2R

# S4 method for RLum.Data.Image
show(object)

# S4 method for RLum.Data.Image
set_RLum(
  class,
  originator,
  .uid,
  .pid,
  recordType = "Image",
  curveType = NA_character_,
  data = raster::brick(raster::raster(matrix())),
  info = list()
)

# S4 method for RLum.Data.Image
get_RLum(object, info.object)

# S4 method for RLum.Data.Image
names_RLum(object)

Arguments

object
get_RLum, names_RLum (required): an object of class RLum.Data.Image

class
set_RLum; character: name of the RLum class to create

originator
set_RLum; character (automatic): contains the name of the calling function (the function that produces this object); can be set manually.

.uid
set_RLum; character (automatic): sets an unique ID for this object using the internal C++ function create_UID.

.pid
set_RLum; character (with default): option to provide a parent id for nesting at will.

recordType
set_RLum; character: record type (e.g. "OSL")

curveType
set_RLum; character: curve type (e.g. "predefined" or "measured")

data
set_RLum; matrix: raw curve data. If data is of type RLum.Data.Image this can be used to re-construct the object, i.e. modified parameters except .uid and .pid. The rest will be subject to copy and paste unless provided.

info
set_RLum; list: info elements

info.object
get_RLum; character: name of the info object to returned

Value

set_RLum

Returns an object from class RLum.Data.Image

get_RLum

Returns the data object (raster::brick)

only the info object if info.object was set.

names_RLum

Returns the names of the info elements

Methods (by generic)

show: Show structure of RLum.Data.Image object

set_RLum: Construction method for RLum.Data.Image object. The slot info is optional and predefined as empty list by default.

get_RLum: Accessor method for RLum.Data.Image object. The argument info.object is optional to directly access the info elements. If no info element name is provided, the raw image data (RasterBrick) will be returned.

names_RLum: Returns the names info elements coming along with this curve object

Slots

recordType

Object of class character containing the type of the curve (e.g. "OSL image", "TL image")

curveType

Object of class character containing curve type, allowed values are measured or predefined

data

Object of class raster::brick containing images (raster data).

info

Object of class list containing further meta information objects

Note

The class should only contain data for a set of images. For additional elements the slot info can be used.

Objects from the class

Objects can be created by calls of the form set_RLum("RLum.Data.Image", ...).

Class version

0.4.3

See also

Author

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

How to cite

Kreutzer, S., 2021. RLum.Data.Image-class(): Class 'RLum.Data.Image'. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.16. https://CRAN.R-project.org/package=Luminescence

Examples

showClass("RLum.Data.Image")
#> Class "RLum.Data.Image" [package "Luminescence"]
#> 
#> Slots:
#>                                                                               
#> Name:   recordType   curveType        data        info  originator        .uid
#> Class:   character   character RasterBrick        list   character   character
#>                   
#> Name:         .pid
#> Class:   character
#> 
#> Extends: 
#> Class "RLum.Data", directly
#> Class "RLum", by class "RLum.Data", distance 2

##create empty RLum.Data.Image object
set_RLum(class = "RLum.Data.Image")
#> 
#>  [RLum.Data.Image-class]
#> 	 recordType: Image
#> 	 curveType: NA
#> 	 .. recorded frames: 1
#> 	 .. .. pixel per frame: 1
#> 	 .. .. x dimension [px]: 1
#> 	 .. .. y dimension [px]: 1
#> 	 .. .. full pixel value range: Inf : -Inf
#> 	 additional info elements: 0