RLum.Data.Spectrum-class {Luminescence}	R Documentation

Class "RLum.Data.Spectrum"

Description

Class for representing luminescence spectra data (TL/OSL/RF).

Usage


## S4 method for signature 'RLum.Data.Spectrum'
show(object)

## S4 method for signature 'RLum.Data.Spectrum'
set_RLum(
  class,
  originator,
  .uid,
  .pid,
  recordType = "Spectrum",
  curveType = NA_character_,
  data = matrix(),
  info = list()
)

## S4 method for signature 'RLum.Data.Spectrum'
get_RLum(object, info.object)

## S4 method for signature 'RLum.Data.Spectrum'
names_RLum(object)

## S4 method for signature 'RLum.Data.Spectrum'
bin_RLum.Data(object, bin_size.col = 1, bin_size.row = 1)

Arguments

object
get_RLum, names_RLum (required): an object of class RLum.Data.Spectrum

class
set_RLum; character (automatic): name of the RLum class to create.

originator
character (automatic): contains the name of the calling function (the function that produces this object); can be set manually.

.uid
set_RLum; character (automatic): sets an unique ID for this object using the internal C++ function create_UID.

.pid
set_RLum; character (with default): option to provide a parent id for nesting at will.

recordType
set_RLum; character: record type (e.g. "OSL")

curveType
set_RLum; character: curve type (e.g. "predefined" or "measured")

data
set_RLum; matrix: raw curve data. If data is of type RLum.Data.Spectrum, this can be used to re-construct the object. If the object is reconstructed, .uid, .pid and orginator are always taken from the input object

info
set_RLum list: info elements

info.object
get_RLum; character (optional): the name of the info object to be called

bin_size.col
integer (with default): set number of channels used for each bin, e.g. bin_size.col = 2 means that two channels are binned. Note: The function does not check the input, very large values mean a full column binning (a single sum)

bin_size.row
integer (with default): set number of channels used for each bin, e.g. bin_size.row = 2 means that two channels are binned. Note: The function does not check the input, very large values mean a full row binning (a single sum)

Value

[set_RLum]

An object from the class RLum.Data.Spectrum

[get_RLum]

A matrix with the spectrum values or

only the info object if info.object was set.

[names_RLum]

The names of the info objects

[bin_RLum.Data]

Same object as input, after applying the binning.

Methods (by generic)

show: Show structure of RLum.Data.Spectrum object

set_RLum: Construction method for RLum.Data.Spectrum object. The slot info is optional and predefined as empty list by default

get_RLum: Accessor method for RLum.Data.Spectrum object. The argument info.object is optional to directly access the info elements. If no info element name is provided, the raw curve data (matrix) will be returned

names_RLum: Returns the names info elements coming along with this curve object

bin_RLum.Data: Allows binning of RLum.Data.Spectrum data. Count values and values on the x-axis are summed-up; for wavelength/energy values the mean is calculated.

Slots

recordType

Object of class character containing the type of the curve (e.g. "TL" or "OSL")

curveType

Object of class character containing curve type, allowed values are measured or predefined

data

Object of class matrix containing spectrum (count) values. Row labels indicate wavelength/pixel values, column labels are temperature or time values.

info

Object of class list containing further meta information objects

Objects from the Class

Objects can be created by calls of the form set_RLum("RLum.Data.Spectrum", ...).

Class version

0.5.2

How to cite

Kreutzer, S., 2021. RLum.Data.Spectrum-class(): Class 'RLum.Data.Spectrum'. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

The class should only contain data for a single spectra data set. For additional elements the slot info can be used. Objects from this class are automatically created by, e.g., read_XSYG2R

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum, RLum.Data, plot_RLum

Examples



showClass("RLum.Data.Spectrum")

##show example data
data(ExampleData.XSYG, envir = environment())
TL.Spectrum

##show data matrix
get_RLum(TL.Spectrum)

##plot spectrum
## Not run: 
plot_RLum(TL.Spectrum)

## End(Not run)


[Package Luminescence version 0.9.15 Index]