Class "RLum.Data.Spectrum"

Class for representing luminescence spectra data (TL/OSL/RF).

# S4 method for RLum.Data.Spectrum
show(object)

# S4 method for RLum.Data.Spectrum
set_RLum(class, originator, .uid, .pid,
  recordType = "Spectrum", curveType = NA_character_, data = matrix(),
  info = list())

# S4 method for RLum.Data.Spectrum
get_RLum(object, info.object)

# S4 method for RLum.Data.Spectrum
names_RLum(object)

Arguments

object
get_RLum, names_RLum (required): an object of class RLum.Data.Spectrum

class
set_RLum; character (automatic): name of the RLum class to create.

originator
character (automatic): contains the name of the calling function (the function that produces this object); can be set manually.

.uid
set_RLum; character (automatic): sets an unique ID for this object using the internal C++ function create_UID.

.pid
set_RLum; character (with default): option to provide a parent id for nesting at will.

recordType
set_RLum; character: record type (e.g. "OSL")

curveType
set_RLum; character: curve type (e.g. "predefined" or "measured")

data
set_RLum; matrix: raw curve data. If data is of type RLum.Data.Spectrum, this can be used to re-construct the object.

info
set_RLum list: info elements

info.object
get_RLum; character (optional): the name of the info object to be called

Value

[set_RLum]

An object from the class RLum.Data.Spectrum get_RLum

A matrix with the spectrum values or

only the info object if info.object was set.

names_RLum

The names of the info objects

Methods (by generic)

show: Show structure of RLum.Data.Spectrum object

set_RLum: Construction method for RLum.Data.Spectrum object. The slot info is optional and predefined as empty list by default

get_RLum: Accessor method for RLum.Data.Spectrum object. The argument info.object is optional to directly access the info elements. If no info element name is provided, the raw curve data (matrix) will be returned

names_RLum: Returns the names info elements coming along with this curve object

Slots

recordType

Object of class character containing the type of the curve (e.g. "TL" or "OSL")

curveType

Object of class character containing curve type, allowed values are measured or predefined

data

Object of class matrix containing spectrum (count) values. Row labels indicate wavelength/pixel values, column labels are temperature or time values.

info

Object of class list containing further meta information objects

Note

The class should only contain data for a single spectra data set. For additional elements the slot info can be used. Objects from this class are automatically created by, e.g., read_XSYG2R

Objects from the Class

Objects can be created by calls of the form set_RLum("RLum.Data.Spectrum", ...).

Class version

0.4.0

See also

RLum, RLum.Data, plot_RLum

How to cite

Kreutzer, S. (2018). RLum.Data.Spectrum-class(): Class 'RLum.Data.Spectrum'. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence

Examples

showClass("RLum.Data.Spectrum")

#> Class "RLum.Data.Spectrum" [package "Luminescence"]
#> 
#> Slots:
#>                                                                         
#> Name:  recordType  curveType       data       info originator       .uid
#> Class:  character  character     matrix       list  character  character
#>                  
#> Name:        .pid
#> Class:  character
#> 
#> Extends: 
#> Class "RLum.Data", directly
#> Class "RLum", by class "RLum.Data", distance 2

##show example data
data(ExampleData.XSYG, envir = environment())
TL.Spectrum

#> 
#>  [RLum.Data.Spectrum-class]
#> 	 recordType: TL (Spectrometer)
#> 	 curveType: measured
#> 	 .. recorded frames: 24
#> 	 .. .. measured values per frame: 1024
#> 	 .. .. range wavelength/pixel: 296.5 823.126
#> 	 .. .. range time/temp.: 0.029 460.007
#> 	 .. .. range count values: 554 65405
#> 	 additional info elements: 14

##show data matrix
get_RLum(TL.Spectrum)

#>         0.029 20.027 40.025 60.026 80.025 100.023 120.022 140.022 160.02
#> 296.5     555  556.0    561    568    573     579     573     574    570
#> 297.055   555  556.0    560    568    573     579     574     574    570
#> 297.61    555  556.0    555    564    571     579     575     573    573
#> 298.164   554  555.0    555    557    571     579     576     572    573
#> 298.719   554  555.0    555    556    569     579     578     571    573
#> 299.273   554  555.0    555    556    569     579     578     571    572
#> 299.827   554  555.0    557    556    569     579     579     571    572
#> 300.382   555  558.0    557    557    567     579     579     576    572
#> 300.936   555  562.0    560    562    567     579     579     576    574
#> 301.489   557  562.5    560    562    567     579     579     576    574
#> 302.043   559  562.5    560    562    568     579     579     574    574
#> 302.597   560  563.0    560    562    570     579     579     574    574
#> 303.15    560  563.0    560    562    571     580     579     575    572
#> 303.703   560  563.0    560    562    574     580     579     579    572
#> 304.256   560  563.0    560    564    575     580     581     580    572
#> 304.809   560  564.0    561    564    576     580     581     580    575
#> 305.362   561  564.0    561    564    578     581     581     580    578
#> 305.915   561  565.0    562    565    578     581     582     580    578
#> 306.467   561  565.0    563    566    578     582     582     580    578
#> 307.02    561  565.0    564    566    578     582     582     583    575
#> 307.572   561  565.0    564    568    574     582     582     583    575
#> 308.124   565  565.0    566    568    574     582     582     583    575
#> 308.676   565  565.0    566    569    574     582     581     580    575
#> 309.228   564  565.0    566    570    574     582     581     580    580
#> 309.78    564  565.0    566    570    577     583     581     580    580
#> 310.332   564  565.0    566    570    577     583     581     581    580
#> 310.883   562  565.0    566    570    577     583     584     584    580
#> 311.435   562  565.0    567    572    575     583     584     584    580
#> 311.986   564  565.0    567    572    575     583     584     584    580
#> 312.537   567  565.0    567    572    575     584     584     584    581
#> 313.088   567  565.0    567    572    575     584     584     584    585
#> 313.639   567  566.0    568    572    577     585     584     583    585
#> 314.189   567  566.0    568    572    577     585     586     583    585
#> 314.74    567  566.0    568    572    577     585     586     583    585
#> 315.29    567  566.0    567    572    577     586     586     583    585
#> 315.841   567  566.0    567    571    577     586     586     579    583
#> 316.391   564  567.0    567    571    576     586     586     578    582
#> 316.941   564  567.0    567    571    576     586     585     578    580
#> 317.491   563  567.0    567    569    576     586     585     580    580
#> 318.04    563  567.0    567    569    577     586     585     583    583
#> 318.59    563  567.0    567    569    577     586     585     583    586
#>         180.019 200.02 220.018 240.017 260.016 280.015 300.014 320.014 340.014
#> 296.5       571    562   561.0   562.0     566     571     587     623     690
#> 297.055     571    562   561.0   562.0     565     571     586     623     686
#> 297.61      564    562   557.0   562.0     561     567     583     623     684
#> 298.164     564    562   557.0   562.0     559     567     581     621     684
#> 298.719     564    563   557.0   562.0     559     567     580     613     684
#> 299.273     564    563   560.0   563.0     563     567     580     613     684
#> 299.827     566    563   560.0   563.0     567     568     583     613     684
#> 300.382     566    563   560.0   563.0     567     569     586     613     687
#> 300.936     566    565   560.0   563.5     567     569     586     615     687
#> 301.489     571    566   562.5   563.5     567     569     586     615     687
#> 302.043     571    566   564.0   564.0     567     569     586     619     687
#> 302.597     571    566   565.0   564.0     565     569     586     622     686
#> 303.15      572    566   565.0   564.0     565     569     584     622     686
#> 303.703     572    567   566.0   564.0     565     573     584     622     686
#> 304.256     572    567   566.0   564.0     565     575     584     622     686
#> 304.809     572    570   567.0   566.0     565     576     588     622     686
#> 305.362     572    570   568.0   566.0     568     576     588     619     686
#> 305.915     573    570   568.0   566.0     569     576     588     619     686
#> 306.467     574    568   569.0   567.0     569     576     590     619     686
#> 307.02      574    568   569.0   567.0     569     574     590     618     686
#> 307.572     574    568   570.0   568.0     569     574     590     618     686
#> 308.124     574    572   570.0   569.0     571     574     590     618     687
#> 308.676     574    574   571.0   569.0     571     578     593     619     687
#> 309.228     576    574   571.0   569.0     571     578     593     619     687
#> 309.78      576    574   571.0   569.0     568     578     594     619     687
#> 310.332     577    574   571.0   569.0     567     576     594     619     687
#> 310.883     578    574   571.0   570.0     567     576     596     621     687
#> 311.435     579    574   571.0   571.0     567     576     596     621     684
#> 311.986     579    574   572.0   571.0     569     577     596     621     684
#> 312.537     579    574   572.0   571.0     570     579     598     621     684
#> 313.088     579    574   572.0   571.0     573     579     602     622     684
#> 313.639     579    574   572.0   571.0     573     579     603     622     687
#> 314.189     579    575   573.0   571.0     573     579     599     622     687
#> 314.74      579    575   573.0   571.0     574     577     591     622     687
#> 315.29      579    577   573.0   571.0     574     577     591     621     687
#> 315.841     579    581   574.0   572.0     574     577     591     621     687
#> 316.391     579    581   574.0   572.0     574     577     591     619     687
#> 316.941     579    579   574.0   572.0     574     578     590     619     687
#> 317.491     579    578   574.0   572.0     575     578     590     617     690
#> 318.04      579    578   574.0   572.0     575     578     590     617     690
#> 318.59      579    578   574.0   572.0     575     574     590     617     690
#>         360.012 380.012 400.01 420.009 440.008 460.007
#> 296.5       828    1128   1612    2490    4157    6574
#> 297.055     828    1110   1600    2490    4056    6438
#> 297.61      851    1100   1594    2490    3990    6370
#> 298.164     851    1100   1592    2490    3990    6370
#> 298.719     826    1095   1591    2490    3990    6364
#> 299.273     826    1095   1591    2479    3994    6361
#> 299.827     826    1097   1591    2467    3994    6357
#> 300.382     825    1097   1589    2462    3994    6342
#> 300.936     825    1097   1581    2459    3951    6317
#> 301.489     825    1097   1581    2458    3942    6316
#> 302.043     829    1097   1575    2444    3938    6276
#> 302.597     829    1097   1575    2444    3924    6269
#> 303.15      829    1097   1567    2444    3924    6258
#> 303.703     829    1097   1566    2437    3924    6253
#> 304.256     818    1092   1566    2437    3915    6224
#> 304.809     818    1092   1557    2436    3915    6224
#> 305.362     818    1092   1557    2434    3905    6224
#> 305.915     818    1092   1557    2421    3905    6208
#> 306.467     820    1091   1557    2421    3905    6202
#> 307.02      820    1078   1558    2421    3888    6169
#> 307.572     820    1077   1564    2421    3878    6162
#> 308.124     820    1077   1564    2414    3877    6162
#> 308.676     818    1082   1562    2406    3877    6162
#> 309.228     817    1082   1560    2406    3872    6162
#> 309.78      817    1082   1558    2401    3861    6144
#> 310.332     817    1078   1557    2401    3861    6123
#> 310.883     816    1078   1553    2401    3850    6116
#> 311.435     816    1078   1553    2386    3832    6102
#> 311.986     816    1080   1549    2386    3826    6097
#> 312.537     824    1080   1549    2386    3824    6097
#> 313.088     824    1080   1549    2387    3824    6097
#> 313.639     823    1080   1549    2387    3827    6097
#> 314.189     823    1080   1551    2387    3827    6084
#> 314.74      823    1078   1551    2377    3822    6077
#> 315.29      823    1079   1535    2364    3822    6062
#> 315.841     821    1077   1535    2364    3822    6062
#> 316.391     818    1077   1534    2364    3798    6018
#> 316.941     818    1074   1534    2361    3798    6017
#> 317.491     818    1068   1534    2361    3790    6017
#> 318.04      818    1068   1525    2358    3776    6020
#> 318.59      818    1068   1525    2348    3776    6020
#>  [ reached getOption("max.print") -- omitted 983 rows ]

##plot spectrum

# NOT RUN {
plot_RLum(TL.Spectrum)
# }