RLum.Results-class {Luminescence}	R Documentation

Class "RLum.Results"

Description

Object class contains results data from functions (e.g., analyse_SAR.CWOSL).

Usage


## S4 method for signature 'RLum.Results'
show(object)

## S4 method for signature 'RLum.Results'
set_RLum(class, originator, .uid, .pid, data = list(), info = list())

## S4 method for signature 'RLum.Results'
get_RLum(object, data.object, info.object = NULL, drop = TRUE)

## S4 method for signature 'RLum.Results'
length_RLum(object)

## S4 method for signature 'RLum.Results'
names_RLum(object)

Arguments

object
get_RLum; RLum.Results (required): an object of class RLum.Results to be evaluated

class
set_RLum; character (required): name of the RLum class to create

originator
set_RLum; character (automatic): contains the name of the calling function (the function that produces this object); can be set manually.

.uid
set_RLum; character (automatic): sets an unique ID for this object using the internal C++ function create_UID.

.pid
set_RLum; character (with default): option to provide a parent id for nesting at will.

data
set_RLum; list (optional): a list containing the data to be stored in the object

info
set_RLum; list (optional): a list containing additional info data for the object

data.object
get_RLum; character or numeric: name or index of the data slot to be returned

info.object
get_RLum; character (optional): name of the wanted info element

drop
get_RLum; logical (with default): coerce to the next possible layer (which are data objects, drop = FALSE keeps the original RLum.Results

Value

set_RLum:

Returns an object from the class RLum.Results

get_RLum:

Returns:

Data object from the specified slot

list of data objects from the slots if 'data.object' is vector or

an RLum.Results for drop = FALSE.

length_RLum

Returns the number of data elements in the RLum.Results object.

names_RLum

Returns the names of the data elements in the object.

Methods (by generic)

show: Show structure of RLum.Results object

set_RLum: Construction method for an RLum.Results object.

get_RLum: Accessor method for RLum.Results object. The argument data.object allows directly accessing objects delivered within the slot data. The default return object depends on the object originator (e.g., fit_LMCurve). If nothing is specified always the first data.object will be returned.

Note: Detailed specification should be made in combination with the originator slot in the receiving function if results are pipped.

length_RLum: Returns the length of the object, i.e., number of stored data.objects

names_RLum: Returns the names data.objects

Slots

data

Object of class list containing output data

Objects from the Class

Objects can be created by calls of the form new("RLum.Results", ...).

Class version

0.5.2

How to cite

Kreutzer, S., 2021. RLum.Results-class(): Class 'RLum.Results'. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

The class is intended to store results from functions to be used by other functions. The data in the object should always be accessed by the method get_RLum.

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum, plot_RLum, merge_RLum

Examples



showClass("RLum.Results")

##create an empty object from this class
set_RLum(class = "RLum.Results")

##use another function to show how it works

##Basic calculation of the dose rate for a specific date
 dose.rate <-  calc_SourceDoseRate(
   measurement.date = "2012-01-27",
   calib.date = "2014-12-19",
   calib.dose.rate = 0.0438,
   calib.error = 0.0019)

##show object
dose.rate

##get results
get_RLum(dose.rate)

##get parameters used for the calcualtion from the same object
get_RLum(dose.rate, data.object = "parameters")

##alternatively objects can be accessed using S3 generics, such as
dose.rate$parameters


[Package Luminescence version 0.9.15 Index]