Convert Risoe.BINfileData object to an RLum.Analysis object

Converts values from one specific position of a Risoe.BINfileData S4-class object to an RLum.Analysis object.

Risoe.BINfileData2RLum.Analysis(object, pos = NULL, grain = NULL,
  run = NULL, set = NULL, ltype = NULL, dtype = NULL,
  protocol = "unknown", keep.empty = TRUE, txtProgressBar = FALSE)

Arguments

object
Risoe.BINfileData (required): Risoe.BINfileData object

pos
numeric (optional): position number of the Risoe.BINfileData object for which the curves are stored in the RLum.Analysis object. If length(position)>1 a list of RLum.Analysis objects is returned. If nothing is provided every position will be converted. If the position is not valid NA is returned.

grain
vector, numeric (optional): grain number from the measurement to limit the converted data set (e.g., grain = c(1:48)). Please be aware that this option may lead to unwanted effects, as the output is strictly limited to the choosen grain number for all position numbers

run
vector, numeric (optional): run number from the measurement to limit the converted data set (e.g., run = c(1:48)).

set
vector, numeric (optional): set number from the measurement to limit the converted data set (e.g., set = c(1:48)).

ltype
vector, character (optional): curve type to limit the converted data. Commonly allowed values are: IRSL, OSL, TL, RIR, RBR and USER (see also Risoe.BINfileData)

dtype
vector, character (optional): data type to limit the converted data. Commonly allowed values are listed in Risoe.BINfileData

protocol
character (optional): sets protocol type for analysis object. Value may be used by subsequent analysis functions.

keep.empty
logical (with default): If TRUE (default) an RLum.Analysis object is returned even if it does not contain any records. Set to FALSE to discard all empty objects.

txtProgressBar
logical (with default): enables or disables txtProgressBar.

Value

Returns an RLum.Analysis object.

Details

The RLum.Analysis object requires a set of curves for specific further protocol analyses. However, the Risoe.BINfileData usually contains a set of curves for different aliquots and different protocol types that may be mixed up. Therefore, a conversion is needed.

Note

The protocol argument of the RLum.Analysis object is set to 'unknown' if not stated otherwise.

Function version

0.4.2 (2018-01-21 17:22:38)

See also

How to cite

Kreutzer, S. (2019). Risoe.BINfileData2RLum.Analysis(): Convert Risoe.BINfileData object to an RLum.Analysis object. Function version 0.4.2. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0.104. https://CRAN.R-project.org/package=Luminescence

Examples

##load data
data(ExampleData.BINfileData, envir = environment())

##convert values for position 1
Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos = 1)