analyse_portableOSL {Luminescence}	R Documentation

Analyse portable CW-OSL measurements

Description

The function analyses CW-OSL curve data produced by a SUERC portable OSL reader and produces a combined plot of OSL/IRSL signal intensities, OSL/IRSL depletion ratios and the IRSL/OSL ratio.

Usage


analyse_portableOSL(
  object,
  signal.integral,
  invert = FALSE,
  normalise = FALSE,
  plot = TRUE,
  ...
)

Arguments

object
RLum.Analysis (required): RLum.Analysis object produced by read_PSL2R.

signal.integral
vector (required): A vector of two values specifying the lower and upper channel used to calculate the OSL/IRSL signal. Can be provided in form of c(1, 5) or 1:5.

invert
logical (with default): TRUE to calculate and plot the data in reverse order.

normalise
logical (with default): TRUE to normalise the OSL/IRSL signals by the mean of all corresponding data curves.

plot
logical (with default): enable/disable plot output

...
currently not used.

Details

This function only works with RLum.Analysis objects produced by read_PSL2R. It further assumes (or rather requires) an equal amount of OSL and IRSL curves that are pairwise combined for calculating the IRSL/OSL ratio. For calculating the depletion ratios the cumulative signal of the last n channels (same number of channels as specified by signal.integral) is divided by cumulative signal of the first n channels (signal.integral).

Value

Returns an S4 RLum.Results object.

Function version

0.0.3

How to cite

Burow, C., 2021. analyse_portableOSL(): Analyse portable CW-OSL measurements. Function version 0.0.3. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Christoph Burow, University of Cologne (Germany) , RLum Developer Team

See Also

RLum.Analysis, RLum.Data.Curve

Examples



# (1) load example data set
data("ExampleData.portableOSL", envir = environment())

# (2) merge and plot all RLum.Analysis objects
merged <- merge_RLum(ExampleData.portableOSL)
plot_RLum(merged, combine = TRUE)
merged

# (3) analyse and plot
results <- analyse_portableOSL(merged, signal.integral = 1:5, invert = FALSE, normalise = TRUE)
get_RLum(results)




[Package Luminescence version 0.9.15 Index]