Function to apply spectral efficiency correction to RLum.Data.Spectrum S4 class objects

The function allows spectral efficiency corrections for RLum.Data.Spectrum S4 class objects

apply_EfficiencyCorrection(object, spectral.efficiency)

Arguments

object
RLum.Data.Spectrum (required): S4 object of class RLum.Data.Spectrum

spectral.efficiency
data.frame (required): Data set containing wavelengths (x-column) and relative spectral response values (y-column) in percentage

Value

Returns same object as input (RLum.Data.Spectrum)

Details

The efficiency correction is based on a spectral response dataset provided by the user. Usually the data set for the quantum efficiency is of lower resolution and values are interpolated for the required spectral resolution using the function stats::approx

If the energy calibration differes for both data set NA values are produces that will be removed from the matrix.

Note

Please note that the spectral efficiency data from the camera alone may not sufficiently correct for spectral efficiency of the entire optical system (e.g., spectrometer, camera ...).

Function version

0.1.1 (2017-09-02 23:27:02)

See also

RLum.Data.Spectrum

How to cite

Kreutzer, S., Friedrich, J. (2018). apply_EfficiencyCorrection(): Function to apply spectral efficiency correction to RLum.Data.Spectrum S4 class objects. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence

Examples

##(1) - use with your own data (uncomment for usage)
## spectral.efficiency <- read.csv("your data")
##
## your.spectrum <- apply_EfficiencyCorrection(your.spectrum, )