calc_HomogeneityTest {Luminescence}	R Documentation

Apply a simple homogeneity test after Galbraith (2003)

Description

A simple homogeneity test for De estimates

Usage


calc_HomogeneityTest(data, log = TRUE, ...)

Arguments

data
RLum.Results or data.frame (required): for data.frame: two columns with De (data[,1]) and De error (values[,2])

log
logical (with default): perform the homogeneity test with (un-)logged data

...
further arguments (for internal compatibility only).

Details

For details see Galbraith (2003).

Value

Returns a terminal output. In addition an RLum.Results-object is returned containing the following elements:

summary
data.frame summary of all relevant model results.

data
data.frame original input data

args
list used arguments

call
call the function call

The output should be accessed using the function get_RLum

Function version

0.3.0

How to cite

Burow, C., Kreutzer, S., 2021. calc_HomogeneityTest(): Apply a simple homogeneity test after Galbraith (2003). Function version 0.3.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Christoph Burow, University of Cologne (Germany), Sebastian Kreutzer, IRAMAT-CRP2A, Université Bordeaux Montaigne (France) , RLum Developer Team

References

Galbraith, R.F., 2003. A simple homogeneity test for estimates of dose obtained using OSL. Ancient TL 21, 75-77.

See Also

pchisq

Examples



## load example data
data(ExampleData.DeValues, envir = environment())

## apply the homogeneity test
calc_HomogeneityTest(ExampleData.DeValues$BT998)

## using the data presented by Galbraith (2003)
df <-
 data.frame(
   x = c(30.1, 53.8, 54.3, 29.0, 47.6, 44.2, 43.1),
   y = c(4.8, 7.1, 6.8, 4.3, 5.2, 5.9, 3.0))

calc_HomogeneityTest(df)



[Package Luminescence version 0.9.15 Index]