Calculate the Lx/Tx ratio for a given set of TL curves -beta version-

Calculate Lx/Tx ratio for a given set of TL curves.

<<<<<<< HEAD

calc_TLLxTxRatio(
  Lx.data.signal,
  Lx.data.background = NULL,
  Tx.data.signal,
  Tx.data.background = NULL,
  signal.integral.min,
  signal.integral.max
)

=======

calc_TLLxTxRatio(
  Lx.data.signal,
  Lx.data.background = NULL,
  Tx.data.signal,
  Tx.data.background = NULL,
  signal.integral.min,
  signal.integral.max
)

>>>>>>> dev_0.9.20

Arguments

Lx.data.signal

RLum.Data.Curve or data.frame (required): TL data (x = temperature, y = counts) (TL signal)

<<<<<<< HEAD ======= >>>>>>> dev_0.9.20
Lx.data.background

RLum.Data.Curve or data.frame (optional): TL data (x = temperature, y = counts). If no data are provided no background subtraction is performed.

<<<<<<< HEAD
Tx.data.signal

RLum.Data.Curve or data.frame (required): TL data (x = temperature, y = counts) (TL test signal)

=======
Tx.data.signal

RLum.Data.Curve or data.frame (required): TL data (x = temperature, y = counts) (TL test signal)

>>>>>>> dev_0.9.20
Tx.data.background

RLum.Data.Curve or data.frame (optional): TL data (x = temperature, y = counts). If no data are provided no background subtraction is performed.

<<<<<<< HEAD ======= >>>>>>> dev_0.9.20
signal.integral.min

integer (required): channel number for the lower signal integral bound (e.g. signal.integral.min = 100)

<<<<<<< HEAD ======= >>>>>>> dev_0.9.20
signal.integral.max

integer (required): channel number for the upper signal integral bound (e.g. signal.integral.max = 200)

<<<<<<< HEAD

Value

Returns an S4 object of type RLum.Results. Slot data contains a list with the following structure:

$ LxTx.table
.. $ LnLx
.. $ LnLx.BG
.. $ TnTx
.. $ TnTx.BG
.. $ Net_LnLx
.. $ Net_LnLx.Error

=======

Value

Returns an S4 object of type RLum.Results. Slot data contains a list with the following structure:

$ LxTx.table
.. $ LnLx
.. $ LnLx.BG
.. $ TnTx
.. $ TnTx.BG
.. $ Net_LnLx
.. $ Net_LnLx.Error

>>>>>>> dev_0.9.20

Details

Uncertainty estimation

The standard errors are calculated using the following generalised equation:

$$SE_{signal} <- abs(Signal_{net} * BG_f /BG_{signal}$$

where \(BG_f\) is a term estimated by calculating the standard deviation of the sum of the \(L_x\) background counts and the sum of the \(T_x\) background counts. However, if both signals are similar the error becomes zero.

Note

This function has still BETA status! Please further note that a similar background for both curves results in a zero error and is therefore set to NA.

Function version

0.3.3

See also

Author
<<<<<<< HEAD
Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
=======
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
>>>>>>> dev_0.9.20 Christoph Schmidt, University of Bayreuth (Germany) , RLum Developer Team

How to cite
<<<<<<< HEAD
Kreutzer, S., Schmidt, C., 2022. calc_TLLxTxRatio(): Calculate the Lx/Tx ratio for a given set of TL curves -beta version-. Function version 0.3.3. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2022. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.19. https://CRAN.R-project.org/package=Luminescence
=======
Kreutzer, S., Schmidt, C., 2022. calc_TLLxTxRatio(): Calculate the Lx/Tx ratio for a given set of TL curves -beta version-. Function version 0.3.3. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2022. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.20. https://CRAN.R-project.org/package=Luminescence
>>>>>>> dev_0.9.20

Examples
<<<<<<< HEAD

##load package example data
data(ExampleData.BINfileData, envir = environment())

##convert Risoe.BINfileData into a curve object
temp <- Risoe.BINfileData2RLum.Analysis(TL.SAR.Data, pos = 3)

Lx.data.signal <- get_RLum(temp, record.id=1)
Lx.data.background <- get_RLum(temp, record.id=2)
Tx.data.signal <- get_RLum(temp, record.id=3)
Tx.data.background <- get_RLum(temp, record.id=4)
signal.integral.min <- 210
signal.integral.max <- 230

output <- calc_TLLxTxRatio(Lx.data.signal,
                           Lx.data.background,
                           Tx.data.signal, Tx.data.background,
                           signal.integral.min, signal.integral.max)
get_RLum(output)
=======
    

##load package example data
data(ExampleData.BINfileData, envir = environment())

##convert Risoe.BINfileData into a curve object
temp <- Risoe.BINfileData2RLum.Analysis(TL.SAR.Data, pos = 3)

Lx.data.signal <- get_RLum(temp, record.id=1)
Lx.data.background <- get_RLum(temp, record.id=2)
Tx.data.signal <- get_RLum(temp, record.id=3)
Tx.data.background <- get_RLum(temp, record.id=4)
signal.integral.min <- 210
signal.integral.max <- 230

output <- calc_TLLxTxRatio(Lx.data.signal,
                           Lx.data.background,
                           Tx.data.signal, Tx.data.background,
                           signal.integral.min, signal.integral.max)
get_RLum(output)
>>>>>>> dev_0.9.20
#>     LnLx LnLx.BG  TnTx TnTx.BG net_LnLx net_LnLx.Error net_TnTx net_TnTx.Error
#> 1 257042    4068 82298    2943   252974       49468.92    79355       21449.72
#>       LxTx LxTx.Error
#> 1 3.187877   1.485073
<<<<<<< HEAD

=======

>>>>>>> dev_0.9.20