Calculate De value based on the gSGC by Li et al., 2015

Function returns De value and De value error using the global standardised growth curve (gSGC) assumption proposed by Li et al., 2015 for OSL dating of sedimentary quartz

calc_gSGC(data, gSGC.type = "0-250", gSGC.parameters, n.MC = 100,
  verbose = TRUE, plot = TRUE, ...)

Arguments

data
data.frame (required): input data of providing the following columns: 'LnTn', 'LnTn.error', Lr1Tr1', 'Lr1Tr1.error', 'Dr1' Note: column names are not required. The function expect the input data in the given order

gSGC.type
character (with default): define the function parameters that should be used for the iteration procedure: Li et al., 2015 (Table 2) presented function parameters for two dose ranges: "0-450" and "0-250"

gSGC.parameters
list (optional): option to provide own function parameters used for fitting as named list. Nomenclature follows Li et al., 2015, i.e. list(A,A.error,D0,D0.error,c,c.error,Y0,Y0.error,range), range requires a vector for the range the function is considered as valid, e.g. range = c(0,250) Using this option overwrites the default parameter list of the gSGC, meaning the argument gSGC.type will be without effect

n.MC
integer (with default): number of Monte Carlo simulation runs for error estimation, see details.

verbose
logical: enable or disable terminal output

plot
logical: enable or disable graphical feedback as plot

...
parameters will be passed to the plot output

Value

Returns an S4 object of type RLum.Results. @data $ De.value (data.frame) .. $ De .. $ De.error .. $ Eta $ De.MC (list) contains the matricies from the error estimation. $ uniroot (list) contains the uniroot outputs of the De estimations @info `$ call`` (call) the original function call

Details

The error of the De value is determined using a Monte Carlo simulation approach. Solving of the equation is realised using uniroot. Large values for n.MC will significantly increase the computation time.

Function version

0.1.1 (2018-01-21 17:22:38)

How to cite

Kreutzer, S. (2018). calc_gSGC(): Calculate De value based on the gSGC by Li et al., 2015. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.1. https://CRAN.R-project.org/package=Luminescence

References

Li, B., Roberts, R.G., Jacobs, Z., Li, S.-H., 2015. Potential of establishing a 'global standardised growth curve' (gSGC) for optical dating of quartz from sediments. Quaternary Geochronology 27, 94-104. doi:10.1016/j.quageo.2015.02.011

See also

RLum.Results, get_RLum, uniroot

Examples

results <- calc_gSGC(data = data.frame(
LnTn =  2.361, LnTn.error = 0.087,
Lr1Tr1 = 2.744, Lr1Tr1.error = 0.091,
Dr1 = 34.4))

#> Error in plot.new(): figure margins too large

get_RLum(results, data.object = "De")

#> Error in get_RLum(results, data.object = "De"): object 'results' not found