Export measurement data produced by a Daybreak luminescence reader to CSV-files

This function is a wrapper function around the functions read_Daybreak2R and write_RLum2CSV and it imports an Daybreak-file (TXT-file, DAT-file) and directly exports its content to CSV-files. If nothing is set for the argument path (write_RLum2CSV) the input folder will become the output folder.

convert_Daybreak2CSV(file, ...)

Arguments

file
character (required): name of the Daybreak-file (TXT-file, DAT-file) to be converted to CSV-files

...
further arguments that will be passed to the function read_Daybreak2R and write_RLum2CSV

Value

The function returns either a CSV-file (or many of them) or for the option export = FALSE a list comprising objects of type data.frame and matrix

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

How to cite

Kreutzer, S., 2021. convert_Daybreak2CSV(): Export measurement data produced by a Daybreak luminescence reader to CSV-files. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.16. https://CRAN.R-project.org/package=Luminescence

Examples

if (FALSE) {
##select your BIN-file
file <- file.choose()

##convert
convert_Daybreak2CSV(file)

}