convert_PSL2CSV {Luminescence}	R Documentation

Export PSL-file(s) to CSV-files

Description

This function is a wrapper function around the functions read_PSL2R and write_RLum2CSV and it imports an PSL-file (SUERC portable OSL reader file format) and directly exports its content to CSV-files. If nothing is set for the argument path (write_RLum2CSV) the input folder will become the output folder.

Usage


convert_PSL2CSV(file, ...)

Arguments

file
character (required): name of the PSL-file to be converted to CSV-files

...
further arguments that will be passed to the function read_PSL2R and write_RLum2CSV

Value

The function returns either a CSV-file (or many of them) or for the option export = FALSE a list comprising objects of type data.frame and matrix

Function version

0.1.0

How to cite

Kreutzer, S., 2021. convert_PSL2CSV(): Export PSL-file(s) to CSV-files. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Analysis, RLum.Data, RLum.Results, utils::write.table, write_RLum2CSV, read_PSL2R

Examples



## Not run: 
##select your BIN-file
file <- file.choose()

##convert
convert_PSL2CSV(file)


## End(Not run)


[Package Luminescence version 0.9.15 Index]