convert_RLum2Risoe.BINfileData {Luminescence}	R Documentation

Converts RLum.Analysis-objects and RLum.Data.Curve-objects to RLum2Risoe.BINfileData-objects

Description

The functions converts RLum.Analysis and RLum.Data.Curve objects and a list of those to Risoe.BINfileData objects. The function intends to provide a minimum of compatibility between both formats. The created RLum.Analysis object can be later exported to a BIN-file using the function write_R2BIN.

Usage


convert_RLum2Risoe.BINfileData(object, keep.position.number = FALSE)

Arguments

object
RLum.Analysis or RLum.Data.Curve (required): input object to be converted

keep.position.number
logical (with default): keeps the original position number or re-calculate the numbers to avoid doubling

Value

The function returns a Risoe.BINfileData object.

Function version

0.1.3

How to cite

Kreutzer, S., 2021. convert_RLum2Risoe.BINfileData(): Converts RLum.Analysis-objects and RLum.Data.Curve-objects to RLum2Risoe.BINfileData-objects. Function version 0.1.3. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

The conversion can be never perfect. The RLum objects may contain information which are not part of the Risoe.BINfileData definition.

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Analysis, RLum.Data.Curve, write_R2BIN

Examples



##simple conversion using the example dataset
data(ExampleData.RLum.Analysis, envir = environment())
convert_RLum2Risoe.BINfileData(IRSAR.RF.Data)


[Package Luminescence version 0.9.15 Index]