convert_SG2MG {Luminescence}	R Documentation

Converts Single-Grain Data to Multiple-Grain Data

Description

Conversion of single-grain data to multiple-grain data by adding signals from grains belonging to one disc (unique pairs of position, set and run).

Usage


convert_SG2MG(object, write_file = FALSE, ...)

Arguments

object
Risoe.BINfileData character (required): Risoe.BINfileData object or BIN/BINX-file name

write_file
logical (with default): if the input was a path to a file, the output can be written to a file if TRUE. The multiple grain file will be written into the same folder and with extension -SG to the file name.

...
further arguments passed down to read_BIN2R if input is file path

Value

Risoe.BINfileData object and if write_file = TRUE and the input was a file path, a file is written to origin folder.

Function version

0.1.0

How to cite

Kreutzer, S., Mercier, N., 2021. convert_SG2MG(): Converts Single-Grain Data to Multiple-Grain Data. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom), Norbert Mercier, IRAMAT-CRP2A, UMR 5060, CNRS-Université Bordeaux Montaigne (France); , RLum Developer Team

See Also

Risoe.BINfileData, read_BIN2R, write_R2BIN

Examples


## simple run
## (please not that the example is not using SG data)
data(ExampleData.BINfileData, envir = environment())
convert_SG2MG(CWOSL.SAR.Data)


[Package Luminescence version 0.9.15 Index]