get_Quote {Luminescence}	R Documentation

Function to return essential quotes

Description

This function returns one of the collected essential quotes in the growing library. If called without any parameters, a random quote is returned.

Usage


get_Quote(ID, separated = FALSE)

Arguments

ID
character (optional): quote ID to be returned.

separated
logical (with default): return result in separated form.

Value

Returns a character with quote and respective (false) author.

Function version

0.1.5

How to cite

Dietze, M., Kreutzer, S., 2021. get_Quote(): Function to return essential quotes. Function version 0.1.5. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Michael Dietze, GFZ Potsdam (Germany), Sebastian Kreutzer, Geography & Earth Science, Aberystwyth University (United Kingdom), Dirk Mittelstraß, TU Dresden (Germany) , RLum Developer Team

Examples



## ask for an arbitrary quote
get_Quote()


[Package Luminescence version 0.9.15 Index]