Function to return essential quotes

This function returns one of the collected essential quotes in the growing library. If called without any parameters, a random quote is returned.

get_Quote(ID, separated = FALSE)

Arguments

ID
character (optional): qoute ID to be returned.

separated
logical (with default): return result in separated form.

Value

Returns a character with quote and respective (false) author.

Function version

0.1.4 (2019-04-19 09:59:27)

How to cite

Dietze, M., Kreutzer, S. (2019). get_Quote(): Function to return essential quotes. Function version 0.1.4. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0.104. https://CRAN.R-project.org/package=Luminescence

Examples

## ask for an arbitrary qoute
get_Quote()

#> [1] "The answer to life, the universe and everything: 'get_rightAnswer()'"