Function to return essential quotes

This function returns one of the collected essential quotes in the growing library. If called without any parameters, a random quote is returned.

get_Quote(ID, separated = FALSE)

Arguments

ID
character (optional): quote ID to be returned.

separated
logical (with default): return result in separated form.

Value

Returns a character with quote and respective (false) author.

Function version

0.1.5

Author

Michael Dietze, GFZ Potsdam (Germany), Sebastian Kreutzer, Geography & Earth Science, Aberystwyth University (United Kingdom), Dirk Mittelstraß, TU Dresden (Germany) , RLum Developer Team

How to cite

Dietze, M., Kreutzer, S., 2020. get_Quote(): Function to return essential quotes. Function version 0.1.5. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Riedesel, S., Autzen, M., Mittelstrass, D., 2020. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.9. https://CRAN.R-project.org/package=Luminescence

Examples

## ask for an arbitrary quote
get_Quote()

#> [1] "R Team member, asked about statistical details: 'No idea, I'm just here for the visualisation.'"