Function to return essential quotes

This function returns one of the collected essential quotes in the growing library. If called without any parameters, a random quote is returned.

get_Quote(ID, separated = FALSE)

Arguments

ID
character (optional): qoute ID to be returned.

separated
logical (with default): return result in separated form.

Value

Returns a character with quote and respective (false) author.

Function version

0.1.2 (2018-01-21 17:22:38)

How to cite

Dietze, M. (2018). get_Quote(): Function to return essential quotes. Function version 0.1.2. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.0. https://CRAN.R-project.org/package=Luminescence

Examples

## ask for an arbitrary qoute
get_Quote()

#> [1] "Push Pin: 'Made of used sample carriers'"