get_RLum {Luminescence}	R Documentation

General accessors function for RLum S4 class objects

Description

Function calls object-specific get functions for RLum S4 class objects.

Usage


get_RLum(object, ...)

## S4 method for signature 'list'
get_RLum(object, class = NULL, null.rm = FALSE, ...)

## S4 method for signature ''NULL''
get_RLum(object, ...)

Arguments

object
RLum (required): S4 object of class RLum or an object of type list containing only objects of type RLum

...
further arguments that will be passed to the object specific methods. For further details on the supported arguments please see the class documentation: RLum.Data.Curve, RLum.Data.Spectrum, RLum.Data.Image, RLum.Analysis and RLum.Results

class
character (optional): allows to define the class that gets selected if applied to a list, e.g., if a list consists of different type of RLum-class objects, this arguments allows to make selection. If nothing is provided, all RLum-objects are treated.

null.rm
logical (with default): option to get rid of empty and NULL objects

Details

The function provides a generalised access point for specific RLum objects.
Depending on the input object, the corresponding get function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class.

Value

Return is the same as input objects as provided in the list.

Functions

get_RLum,list-method: Returns a list of RLum objects that had been passed to get_RLum

get_RLum,NULL-method: Returns NULL

Function version

0.3.3

How to cite

Kreutzer, S., 2021. get_RLum(): General accessors function for RLum S4 class objects. Function version 0.3.3. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Data.Curve, RLum.Data.Image, RLum.Data.Spectrum, RLum.Analysis, RLum.Results

Examples



##Example based using data and from the calc_CentralDose() function

##load example data
data(ExampleData.DeValues, envir = environment())

##apply the central dose model 1st time
temp1 <- calc_CentralDose(ExampleData.DeValues$CA1)

##get results and store them in a new object
temp.get <- get_RLum(object = temp1)


[Package Luminescence version 0.9.15 Index]