Import Luminescence Data into R

Convenience wrapper function to provide a quicker and more standardised way of reading data into R by looping through all in the package available data import functions starting with read_.

Usage

import_Data(file, ..., fastForward = TRUE, verbose = FALSE)

Arguments

file

character (required): file to be imported, can be a list

...

arguments to be further passed down to supported functions (please check the functions to determine the correct arguments)

fastForward

logical (with default): option to create RLum objects during import or a list of such objects

verbose

logical (with default): enable/disable verbose mode

Function version

0.1.1

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany) , RLum Developer Team

How to cite

Kreutzer, S., 2024. import_Data(): Import Luminescence Data into R. Function version 0.1.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., 2024. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.25. https://r-lum.github.io/Luminescence/

Examples

## import BINX/BIN
file <- system.file("extdata/BINfile_V8.binx", package = "Luminescence")
temp <- import_Data(file)
#> [src_create_RLumDataCurve_matrix()] BIN/BINX-file non-conform. TL curve may be wrong!
#> [src_create_RLumDataCurve_matrix()] BIN/BINX-file non-conform. TL curve may be wrong!

## RF data
file <- system.file("extdata", "RF_file.rf", package = "Luminescence")
temp <- import_Data(file)