Attempts to install the development version of the 'Luminescence' package

This function is a convenient method for installing the development version of the R package 'Luminescence' directly from GitHub.

<<<<<<< HEAD

install_DevelopmentVersion(force_install = FALSE)

=======

install_DevelopmentVersion(force_install = FALSE)

>>>>>>> dev_0.9.20

Arguments

force_install

logical (optional): If FALSE (the default) the function produces and prints the required code to the console for the user to run manually afterwards. When TRUE and all requirements are fulfilled (see details) this function attempts to install the package itself.

<<<<<<< HEAD

Value

This function requires user input at the command prompt to choose the =======

Value

This function requires user input at the command prompt to choose the >>>>>>> dev_0.9.20 desired development branch to be installed. The required R code to install the package is then printed to the console.

Details

This function uses Luminescence::github_branches to check which development branches of the R package 'Luminescence' are currently available on GitHub. The user is then prompted to choose one of the branches to be installed. It further checks whether the R package 'devtools' is currently installed and available on the system. Finally, it prints R code to the console that the user can copy and paste to the R console in order to install the desired development version of the package.

If force_install=TRUE the functions checks if 'devtools' is available and then attempts to install the chosen development branch via devtools::remote-reexports.

Examples
<<<<<<< HEAD

if (FALSE) {
install_DevelopmentVersion()
}

=======
    

if (FALSE) {
install_DevelopmentVersion()
}

>>>>>>> dev_0.9.20