General accessor function for RLum S4 class objects

Function calls object-specific get functions for RLum S4 class objects.

length_RLum(object)

Arguments

object
RLum (required): S4 object of class RLum

Value

Return is the same as input objects as provided in the list.

Details

The function provides a generalised access point for specific RLum objects.
Depending on the input object, the corresponding get function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class.

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) (France) , RLum Developer Team

How to cite

Kreutzer, S., 2021. length_RLum(): General accessor function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.16. https://CRAN.R-project.org/package=Luminescence