merge_RLum {Luminescence}	R Documentation

General merge function for RLum S4 class objects

Description

Function calls object-specific merge functions for RLum S4 class objects.

Usage


merge_RLum(objects, ...)

Arguments

objects
list of RLum (required): list of S4 object of class RLum

...
further arguments that one might want to pass to the specific merge function

Details

The function provides a generalised access point for merge specific RLum objects. Depending on the input object, the corresponding merge function will be selected. Allowed arguments can be found in the documentations of each merge function. Empty list elements (NULL) are automatically removed from the input list.

object 	 	 corresponding merge function

RLum.Data.Curve 	 : 	 merge_RLum.Data.Curve

RLum.Analysis 	 : 	 merge_RLum.Analysis

RLum.Results 	 : 	 merge_RLum.Results

Value

Return is the same as input objects as provided in the list.

Function version

0.1.2

How to cite

Kreutzer, S., 2021. merge_RLum(): General merge function for RLum S4 class objects. Function version 0.1.2. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

So far not for every RLum object a merging function exists.

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Data.Curve, RLum.Data.Image, RLum.Data.Spectrum, RLum.Analysis, RLum.Results

Examples




##Example based using data and from the calc_CentralDose() function

##load example data
data(ExampleData.DeValues, envir = environment())

##apply the central dose model 1st time
temp1 <- calc_CentralDose(ExampleData.DeValues$CA1)

##apply the central dose model 2nd time
temp2 <- calc_CentralDose(ExampleData.DeValues$CA1)

##merge the results and store them in a new object
temp.merged <- get_RLum(merge_RLum(objects = list(temp1, temp2)))



[Package Luminescence version 0.9.15 Index]