Merge function for RLum.Analysis S4 class objects

Function allows merging of RLum.Analysis objects and adding of allowed objects to an RLum.Analysis.

merge_RLum.Analysis(objects)

Arguments

objects

list of RLum.Analysis (required): list of S4 objects of class RLum.Analysis. Furthermore other objects of class RLum can be added, see details.

Value

Return an RLum.Analysis object.

Details

This function simply allowing to merge RLum.Analysis objects. Additionally other RLum objects can be added to an existing RLum.Analysis object. Supported objects to be added are: RLum.Data.Curve, RLum.Data.Spectrum and RLum.Data.Image.

The order in the new RLum.Analysis object is the object order provided with the input list.

Note

The information for the slot 'protocol' is taken from the first RLum.Analysis object in the input list. Therefore at least one object of type RLum.Analysis has to be provided.

Function version

0.2.0

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany) , RLum Developer Team

How to cite

Kreutzer, S., 2024. merge_RLum.Analysis(): Merge function for RLum.Analysis S4 class objects. Function version 0.2.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2024. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.24.9000-106. https://CRAN.R-project.org/package=Luminescence

Examples

##merge different RLum objects from the example data
data(ExampleData.RLum.Analysis, envir = environment())
data(ExampleData.BINfileData, envir = environment())

object <- Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos=1)
curve <- get_RLum(object)[[2]]

temp.merged <- merge_RLum.Analysis(list(curve, IRSAR.RF.Data, IRSAR.RF.Data))