merge_RLum.Results {Luminescence}	R Documentation

Merge function for RLum.Results S4-class objects

Description

Function merges objects of class RLum.Results. The slots in the objects are combined depending on the object type, e.g., for data.frame and matrix rows are appended.

Usage


merge_RLum.Results(objects)

Arguments

objects
list (required): a list of RLum.Results objects

Function version

0.2.0

How to cite

Kreutzer, S., 2021. merge_RLum.Results(): Merge function for RLum.Results S4-class objects. Function version 0.2.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

The originator is taken from the first element and not reset to merge_RLum

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

[Package Luminescence version 0.9.15 Index]