Merge Risoe.BINfileData objects or Risoe BIN-files

Function allows merging Risoe BIN/BINX files or Risoe.BINfileData objects.

merge_Risoe.BINfileData(
  input.objects,
  output.file,
  keep.position.number = FALSE,
  position.number.append.gap = 0
)

Arguments

input.objects
character with Risoe.BINfileData objects (required): Character vector with path and files names (e.g. input.objects = c("path/file1.bin", "path/file2.bin") or Risoe.BINfileData objects (e.g. input.objects = c(object1, object2)). Alternatively a list is supported.

output.file
character (optional): File output path and name. If no value is given, a Risoe.BINfileData is returned instead of a file.

keep.position.number
logical (with default): Allows keeping the original position numbers of the input objects. Otherwise the position numbers are recalculated.

position.number.append.gap
integer (with default): Set the position number gap between merged BIN-file sets, if the option keep.position.number = FALSE is used. See details for further information.

Value

Returns a file or a Risoe.BINfileData object.

Details

The function allows merging different measurements to one file or one object. The record IDs are recalculated for the new object. Other values are kept for each object. The number of input objects is not limited.

position.number.append.gap option

If the option keep.position.number = FALSE is used, the position numbers of the new data set are recalculated by adding the highest position number of the previous data set to the each position number of the next data set. For example: The highest position number is 48, then this number will be added to all other position numbers of the next data set (e.g. 1 + 48 = 49)

However, there might be cases where an additional addend (summand) is needed before the next position starts. Example:

Position number set (A): 1,3,5,7

Position number set (B): 1,3,5,7

With no additional summand the new position numbers would be: 1,3,5,7,8,9,10,11. That might be unwanted. Using the argument position.number.append.gap = 1 it will become: 1,3,5,7,9,11,13,15,17.

Note

The validity of the output objects is not further checked.

Function version

0.2.7

How to cite

Kreutzer, S., 2020. merge_Risoe.BINfileData(): Merge Risoe.BINfileData objects or Risoe BIN-files. Function version 0.2.7. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Riedesel, S., Autzen, M., 2020. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.8. https://r-luminescence.org

References

Duller, G., 2007. Analyst.

See also

Author

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

Examples

##merge two objects
data(ExampleData.BINfileData, envir = environment())

object1 <- CWOSL.SAR.Data
object2 <- CWOSL.SAR.Data

object.new <- merge_Risoe.BINfileData(c(object1, object2))