Model Luminescence Signals (wrapper)

Wrapper for the function RLumModel::model_LuminescenceSignals from the package RLumModel::RLumModel-package. For the further details and examples please see the manual of this package.

model_LuminescenceSignals(model, sequence, lab.dose_rate = 1,
  simulate_sample_history = FALSE, plot = TRUE, verbose = TRUE,
  show_structure = FALSE, own_parameters = NULL,
  own_state_parameters = NULL, own_start_temperature = NULL, ...)

Arguments

model
character (required): set model to be used. Available models are: "Bailey2001", "Bailey2002", "Bailey2004", "Pagonis2007", "Pagonis2008", "Friedrich2017", "Friedrich2018" and for own models "customized" (or "customised"). Note: When model = "customized" is set, the argument 'own_parameters' has to be set.

sequence
list (required): set sequence to model as list or as *.seq file from the Riso sequence editor. To simulate SAR measurements there is an extra option to set the sequence list (cf. details).

lab.dose_rate
numeric (with default): laboratory dose rate in XXX Gy/s for calculating seconds into Gray in the *.seq file.

simulate_sample_history
logical (with default): FALSE (with default): simulation begins at laboratory conditions, TRUE: simulations begins at crystallization (all levels 0) process

plot
logical (with default): Enables or disables plot output

verbose
logical (with default): Verbose mode on/off

show_structure
logical (with default): Shows the structure of the result. Recommended to show record.id to analyse concentrations.

own_parameters
list (with default): This argument allows the user to submit own parameter sets. The list has to contain the following items:

N: Concentration of electron- and hole traps [cm^(-3)]

E: Electron/Hole trap depth [eV

s: Frequency factor [s^(-1)]

A: Conduction band to electron trap and valence band to hole trap transition probability [s^(-1) * cm^(3)]. CAUTION: Not every publication uses the same definition of parameter A and B! See vignette "RLumModel - Usage with own parameter sets" for further details

B: Conduction band to hole centre transition probability [s^(-1) * cm^(3)].

Th: Photo-eviction constant or photoionisation cross section, respectively

E_th: Thermal assistence energy [eV]

k_B: Boltzman constant 8.617e-05 [eV/K]

W: activation energy 0.64 [eV] (for UV)

K: 2.8e7 (dimensionless constant)

model: "customized"

R (optional): Ionisation rate (pair production rate) equivalent to 1 Gy/s [s^(-1) * cm^(-3)]

For further details see Bailey 2001, Wintle 1975, vignette "RLumModel - Using own parameter sets" and example 3.

own_state_parameters
numeric (with default): Some publications (e.g. Pagonis 2009) offer state parameters. With this argument the user can submit this state parameters. For further details see vignette ""RLumModel - Using own parameter sets" and example 3.

own_start_temperature
numeric (with default): Parameter to control the start temperature (in deg. C) of a simulation. This parameter takes effect only when 'model = "customized"' is choosen.

...
further arguments and graphical parameters passed to plot.default. See details for further information.

Function version

0.1.3