S4-names function for RLum S4 class objects

Function calls object-specific names functions for RLum S4 class objects.

Usage

names_RLum(object)

# S4 method for class 'list'
names_RLum(object)

Arguments

object

RLum (required): S4 object of class RLum

Value

Returns a character

Details

The function provides a generalised access point for specific RLum objects.
Depending on the input object, the corresponding 'names' function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class.

Functions

names_RLum(list): Returns a list of RLum objects that had been passed to names_RLum

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)