plot_DRCSummary {Luminescence}	R Documentation

Create a Dose-Response Curve Summary Plot

Description

While analysing OSL SAR or pIRIR-data the view on the data is limited usually to one dose-response curve (DRC) at the time for one aliquot. This function overcomes this limitation by plotting all DRC from an RLum.Results object created by the function analyse_SAR.CWOSL in one single plot.

Usage


plot_DRCSummary(
  object,
  source_dose_rate = NULL,
  sel_curves = NULL,
  show_dose_points = FALSE,
  show_natural = FALSE,
  n = 51L,
  ...
)

Arguments

object
RLum.Results object (required): input object created by the function analyse_SAR.CWOSL. The input object can be provided as list.

source_dose_rate
numeric (optional): allows to modify the axis and show values in Gy, instead seconds. Only a single numerical values is allowed.

sel_curves
numeric (optional): id of the curves to be plotting in its occurring order. A sequence can be provided for selecting, e.g., only every 2nd curve from the input object

show_dose_points
logical (with default): enable or disable plot of dose points in the graph

show_natural
logical (with default): enable or disable the plot of the natural Lx/Tx values

n
integer (with default): the number of x-values used to evaluate one curve object. Large numbers slow down the plotting process and are usually not needed

...
Further arguments and graphical parameters to be passed.

Details

If you want plot your DRC on an energy scale (dose in Gy), you can either use the option source_dose_rate provided below or your can SAR analysis with the dose points in Gy (better axis scaling).

Value

An RLum.Results object is returned:

Slot: @data

OBJECT 	 TYPE 	 COMMENT

results 	 data.frame 	 with dose and LxTx values

data 	 RLum.Results 	 original input data

Slot: @info

OBJECT 	 TYPE 	 COMMENT

call 	 call 	 the original function call

args 	 list 	 arguments of the original function call

Note: If the input object is a list a list of RLum.Results objects is returned.

Function version

0.2.1

How to cite

Kreutzer, S., Burow, C., 2021. plot_DRCSummary(): Create a Dose-Response Curve Summary Plot. Function version 0.2.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
Christoph Burow, University of Cologne , RLum Developer Team

See Also

RLum.Results, analyse_SAR.CWOSL

Examples



#load data example data
data(ExampleData.BINfileData, envir = environment())

#transform the values from the first position in a RLum.Analysis object
object <- Risoe.BINfileData2RLum.Analysis(CWOSL.SAR.Data, pos=1)

results <- analyse_SAR.CWOSL(
  object = object,
  signal.integral.min = 1,
  signal.integral.max = 2,
   background.integral.min = 900,
   background.integral.max = 1000,
   plot = FALSE
 )

##plot only DRC
plot_DRCSummary(results)


[Package Luminescence version 0.9.15 Index]