Plot single luminescence curves from a BIN file object

Plots single luminescence curves from an object returned by the read_BIN2R function.

plot_Risoe.BINfileData(
  BINfileData,
  position,
  run,
  set,
  sorter = "POSITION",
  ltype = c("IRSL", "OSL", "TL", "RIR", "RBR", "RL"),
  curve.transformation,
  dose_rate,
  temp.lab,
  cex.global = 1,
  ...
)

Arguments

BINfileData

Risoe.BINfileData (required): requires an S4 object returned by the read_BIN2R function.

position

vector (optional): option to limit the plotted curves by position (e.g. position = 1, position = c(1,3,5)).

run

vector (optional): option to limit the plotted curves by run (e.g., run = 1, run = c(1,3,5)).

set

vector (optional): option to limit the plotted curves by set (e.g., set = 1, set = c(1,3,5)).

sorter

character (with default): the plot output can be ordered by "POSITION","SET" or "RUN". POSITION, SET and RUN are options defined in the Risoe Sequence Editor.

ltype

character (with default): option to limit the plotted curves by the type of luminescence stimulation. Allowed values: "IRSL", "OSL","TL", "RIR", "RBR" (corresponds to LM-OSL), "RL". All type of curves are plotted by default.

curve.transformation

character (optional): allows transforming CW-OSL and CW-IRSL curves to pseudo-LM curves via transformation functions. Allowed values are: CW2pLM, CW2pLMi, CW2pHMi and CW2pPMi. See details.

dose_rate

numeric (optional): dose rate of the irradiation source at the measurement date. If set, the given irradiation dose will be shown in Gy. See details.

temp.lab

character (optional): option to allow for different temperature units. If no value is set deg. C is chosen.

cex.global

numeric (with default): global scaling factor.

...

further undocumented plot arguments.

Value

Returns a plot.

Details

Nomenclature

See Risoe.BINfileData

curve.transformation

This argument allows transforming continuous wave (CW) curves to pseudo (linear) modulated curves. For the transformation, the functions of the package are used. Currently, it is not possible to pass further arguments to the transformation functions. The argument works only for ltype OSL and IRSL.

Irradiation time

Plotting the irradiation time (s) or the given dose (Gy) requires that the variable IRR_TIME has been set within the BIN-file. This is normally done by using the 'Run Info' option within the Sequence Editor or by editing in R.

Note

The function has been successfully tested for the Sequence Editor file output version 3 and 4.

Function version

0.4.1

How to cite

Kreutzer, S., Dietze, M., 2022. plot_Risoe.BINfileData(): Plot single luminescence curves from a BIN file object. Function version 0.4.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2022. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.17. https://CRAN.R-project.org/package=Luminescence

References

Duller, G., 2007. Analyst. pp. 1-45.

See also

Author

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom)
Michael Dietze, GFZ Potsdam (Germany) , RLum Developer Team

Examples

##load data
data(ExampleData.BINfileData, envir = environment())

##plot all curves from the first position to the desktop
#pdf(file = "~/Desktop/CurveOutput.pdf", paper = "a4", height = 11, onefile = TRUE)

##example - load from *.bin file
#BINfile<- file.choose()
#BINfileData<-read_BIN2R(BINfile)

#par(mfrow = c(4,3), oma = c(0.5,1,0.5,1))
#plot_Risoe.BINfileData(CWOSL.SAR.Data,position = 1)
#mtext(side = 4, BINfile, outer = TRUE, col = "blue", cex = .7)
#dev.off()