read_PSL2R {Luminescence}	R Documentation

Import PSL files to R

Description

Imports PSL files produced by a SUERC portable OSL reader into R (BETA).

Usage


read_PSL2R(
  file,
  drop_bg = FALSE,
  as_decay_curve = TRUE,
  smooth = FALSE,
  merge = FALSE,
  ...
)

Arguments

file
character (required): path and file name of the PSL file. If input is a vector it should comprise only characters representing valid paths and PSL file names. Alternatively the input character can be just a directory (path). In this case the the function tries to detect and import all PSL files found in the directory.

drop_bg
logical (with default): TRUE to automatically remove all non-OSL/IRSL curves.

as_decay_curve
logical (with default): Portable OSL Reader curves are often given as cumulative light sum curves. Use TRUE (default) to convert the curves to the more usual decay form.

smooth
logical (with default): TRUE to apply Tukey's Running Median Smoothing for OSL and IRSL decay curves. Smoothing is encouraged if you see random signal drops within the decay curves related to hardware errors.

merge
logical (with default): TRUE to merge all RLum.Analysis objects. Only applicable if multiple files are imported.

...
currently not used.

Details

This function provides an import routine for the SUERC portable OSL Reader PSL format. PSL files are just plain text and can be viewed with any text editor. Due to the formatting of PSL files this import function relies heavily on regular expression to find and extract all relevant information. See note.

Value

Returns an S4 RLum.Analysis object containing RLum.Data.Curve objects for each curve.

Function version

0.0.1

How to cite

Burow, C., 2021. read_PSL2R(): Import PSL files to R. Function version 0.0.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

Because this function relies heavily on regular expressions to parse PSL files it is currently only in beta status. If the routine fails to import a specific PSL file please report to <christoph.burow@gmx.net> so the function can be updated.

Author(s)

Christoph Burow, University of Cologne (Germany) , RLum Developer Team

See Also

RLum.Analysis, RLum.Data.Curve, RLum.Data.Curve

Examples



# (1) Import PSL file to R

file <- system.file("extdata", "DorNie_0016.psl", package = "Luminescence")
psl <- read_PSL2R(file, drop_bg = FALSE, as_decay_curve = TRUE, smooth = TRUE, merge = FALSE)
print(str(psl, max.level = 3))
plot(psl, combine = TRUE)


[Package Luminescence version 0.9.15 Index]