Import RF-files to R

Import files produced by the IR-RF 'ImageJ' macro (#TODO ADD REFERENCE) into R and create a list of RLum.Analysis objects

read_RF2R(file)

Arguments

file
character (required): path and file name of the RF file. Alternatively a list of file names can be provided.

Value

Returns an S4 RLum.Analysis object containing RLum.Data.Curve objects for each curve.

Details

The results of spatially resolved IR-RF data are summarised in so-called RF-files (#TODO ADD REFERENCE). This functions provides an easy import to process the data seamlessly with the R package 'Luminescence'. The output of the function can be passed to the function analyse_IRSAR.RF

Function version

0.1.0

How to cite

Kreutzer, S., 2021. read_RF2R(): Import RF-files to R. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.14. https://CRAN.R-project.org/package=Luminescence

References

#TODO ADD REFERENCE

See also

Author

Sebastian Kreutzer, Geography & Earth Science, Aberystwyth University (United Kingdom) , RLum Developer Team

Examples

##Import
file <- system.file("extdata", "RF_file.rf", package = "Luminescence")
temp <- read_RF2R(file)