Import RF-files to R

Import files produced by the IR-RF 'ImageJ' macro (SR-RF.ijm; Mittelstraß and Kreutzer, 2021) into R and create a list of RLum.Analysis objects

read_RF2R(file)

Arguments

file

character (required): path and file name of the RF file. Alternatively a list of file names can be provided.

Value

Returns an S4 RLum.Analysis object containing RLum.Data.Curve objects for each curve.

Details

The results of spatially resolved IR-RF data are summarised in so-called RF-files ((Mittelstraß and Kreutzer, 2021). This functions provides an easy import to process the data seamlessly with the R package 'Luminescence'. The output of the function can be passed to the function analyse_IRSAR.RF

Function version

0.1.0

How to cite

Kreutzer, S., 2023. read_RF2R(): Import RF-files to R. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., 2023. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.22. https://CRAN.R-project.org/package=Luminescence

References

Mittelstraß, D., Kreutzer, S., 2021. Spatially resolved infrared radiofluorescence: single-grain K-feldspar dating using CCD imaging. Geochronology 3, 299–319. doi:10.5194/gchron-3-299-2021

See also

Author

Sebastian Kreutzer, Geography & Earth Science, Aberystwyth University (United Kingdom) , RLum Developer Team

Examples

##Import
file <- system.file("extdata", "RF_file.rf", package = "Luminescence")
temp <- read_RF2R(file)