replicate_RLum {Luminescence}	R Documentation

General replication function for RLum S4 class objects

Description

Function replicates RLum S4 class objects and returns a list for this objects

Usage


replicate_RLum(object, times = NULL)

Arguments

object
RLum (required): an RLum object

times
integer (optional): number for times each element is repeated element

Value

Returns a list of the object to be repeated

Function version

0.1.0

How to cite

Kreutzer, S., 2021. replicate_RLum(): General replication function for RLum S4 class objects. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum

[Package Luminescence version 0.9.15 Index]