set_RLum {Luminescence}	R Documentation

General set function for RLum S4 class objects

Description

Function calls object-specific set functions for RLum S4 class objects.

Usage


set_RLum(class, originator, .uid = create_UID(), .pid = NA_character_, ...)

Arguments

class
RLum (required): name of the S4 class to create

originator
character (automatic): contains the name of the calling function (the function that produces this object); can be set manually.

.uid
character (automatic): sets an unique ID for this object using the internal C++ function create_UID.

.pid
character (with default): option to provide a parent id for nesting at will.

...
further arguments that one might want to pass to the specific set method

Details

The function provides a generalised access point for specific RLum objects.
Depending on the given class, the corresponding method to create an object from this class will be selected. Allowed additional arguments can be found in the documentations of the corresponding RLum class:

RLum.Data.Curve,

RLum.Data.Image,

RLum.Data.Spectrum,

RLum.Analysis,

RLum.Results

Value

Returns an object of the specified class.

Function version

0.3.0

How to cite

Kreutzer, S., 2021. set_RLum(): General set function for RLum S4 class objects. Function version 0.3.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Data.Curve, RLum.Data.Image, RLum.Data.Spectrum, RLum.Analysis, RLum.Results

Examples



##produce empty objects from each class
set_RLum(class = "RLum.Data.Curve")
set_RLum(class = "RLum.Data.Spectrum")
set_RLum(class = "RLum.Data.Spectrum")
set_RLum(class = "RLum.Analysis")
set_RLum(class = "RLum.Results")

##produce a curve object with arbitrary curve values
object <- set_RLum(
class = "RLum.Data.Curve",
curveType = "arbitrary",
recordType = "OSL",
data = matrix(c(1:100,exp(-c(1:100))),ncol = 2))

##plot this curve object
plot_RLum(object)


[Package Luminescence version 0.9.15 Index]