set_Risoe.BINfileData {Luminescence}	R Documentation

General accessor function for RLum S4 class objects

Description

Function calls object-specific get functions for RisoeBINfileData S4 class objects.

Usage


set_Risoe.BINfileData(
  METADATA = data.frame(),
  DATA = list(),
  .RESERVED = list()
)

Arguments

METADATA
x

DATA
x

.RESERVED
x

Details

The function provides a generalised access point for specific Risoe.BINfileData objects.
Depending on the input object, the corresponding get function will be selected. Allowed arguments can be found in the documentations of the corresponding Risoe.BINfileData class.

Value

Return is the same as input objects as provided in the list.

Function version

0.1

How to cite

Kreutzer, S., 2021. set_Risoe.BINfileData(): General accessor function for RLum S4 class objects. Function version 0.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

Risoe.BINfileData

[Package Luminescence version 0.9.15 Index]