smooth_RLum {Luminescence}	R Documentation

Smoothing of data

Description

Function calls the object-specific smooth functions for provided RLum S4-class objects.

Usage


smooth_RLum(object, ...)

## S4 method for signature 'list'
smooth_RLum(object, ...)

Arguments

object
RLum (required): S4 object of class RLum

...
further arguments passed to the specific class method

Details

The function provides a generalised access point for specific RLum objects.
Depending on the input object, the corresponding function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class. The smoothing is based on an internal function called .smoothing.

Value

An object of the same type as the input object is provided

Functions

smooth_RLum,list-method: Returns a list of RLum objects that had been passed to smooth_RLum

Function version

0.1.0

How to cite

Kreutzer, S., 2021. smooth_RLum(): Smoothing of data. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Note

Currently only RLum objects of class RLum.Data.Curve and RLum.Analysis (with curve data) are supported!

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Data.Curve, RLum.Analysis

Examples



##load example data
data(ExampleData.CW_OSL_Curve, envir = environment())

##create RLum.Data.Curve object from this example
curve <-
  set_RLum(
      class = "RLum.Data.Curve",
      recordType = "OSL",
      data = as.matrix(ExampleData.CW_OSL_Curve)
  )

##plot data without and with smoothing
plot_RLum(curve)
plot_RLum(smooth_RLum(curve))


[Package Luminescence version 0.9.15 Index]