Smoothing of data

Function calls the object-specific smooth functions for provided RLum S4-class objects.

smooth_RLum(object, ...)

# S4 method for list
smooth_RLum(object, ...)

Arguments

object
RLum (required): S4 object of class RLum

...
further arguments passed to the specifc class method

Value

An object of the same type as the input object is provided

Details

The function provides a generalised access point for specific RLum objects. Depending on the input object, the corresponding function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class. The smoothing is based on an internal function called .smoothing.

Methods (by class)

list: Returns a list of RLum objects that had been passed to smooth_RLum

Note

Currenlty only RLum objects of class RLum.Data.Curve and RLum.Analysis (with curve data) are supported!

Function version

0.1.0 (2018-01-30 16:13:33)

See also

RLum.Data.Curve, RLum.Analysis

How to cite

Kreutzer, S. (2018). smooth_RLum(): Smoothing of data. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.1. https://CRAN.R-project.org/package=Luminescence

Examples

##load example data
data(ExampleData.CW_OSL_Curve, envir = environment())

##create RLum.Data.Curve object from this example
curve <-
  set_RLum(
      class = "RLum.Data.Curve",
      recordType = "OSL",
      data = as.matrix(ExampleData.CW_OSL_Curve)
  )

##plot data without and with smoothing
plot_RLum(curve)

#> Error in plot.new(): figure margins too large

plot_RLum(smooth_RLum(curve))

#> Error in plot.new(): figure margins too large