structure_RLum {Luminescence}	R Documentation

General structure function for RLum S4 class objects

Description

Function calls object-specific get functions for RLum S4 class objects.

Usage


structure_RLum(object, ...)

## S4 method for signature 'list'
structure_RLum(object, ...)

Arguments

object
RLum (required): S4 object of class RLum

...
further arguments that one might want to pass to the specific structure method

Details

The function provides a generalised access point for specific RLum objects.
Depending on the input object, the corresponding structure function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class.

Value

Returns a data.frame with structure of the object.

Functions

structure_RLum,list-method: Returns a list of RLum objects that had been passed to structure_RLum

Function version

0.2.0

How to cite

Kreutzer, S., 2021. structure_RLum(): General structure function for RLum S4 class objects. Function version 0.2.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

See Also

RLum.Data.Curve, RLum.Data.Image, RLum.Data.Spectrum, RLum.Analysis, RLum.Results

Examples



##load example data
data(ExampleData.XSYG, envir = environment())

##show structure
structure_RLum(OSL.SARMeasurement$Sequence.Object)


[Package Luminescence version 0.9.15 Index]