General structure function for RLum S4 class objects

Function calls object-specific get functions for RLum S4 class objects.

structure_RLum(object, ...)

# S4 method for list
structure_RLum(object, ...)

Arguments

object
RLum (required): S4 object of class RLum

...
further arguments that one might want to pass to the specific structure method

Value

Returns a data.frame with structure of the object.

Details

The function provides a generalised access point for specific RLum objects. Depending on the input object, the corresponding structure function will be selected. Allowed arguments can be found in the documentations of the corresponding RLum class.

Methods (by class)

list: Returns a list of RLum objects that had been passed to structure_RLum

Function version

0.2.0 (2018-01-30 16:22:52)

See also

RLum.Data.Curve, RLum.Data.Image, RLum.Data.Spectrum, RLum.Analysis, RLum.Results

How to cite

Kreutzer, S. (2018). structure_RLum(): General structure function for RLum S4 class objects. Function version 0.2.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2018). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.8.1. https://CRAN.R-project.org/package=Luminescence

Examples

##load example data
data(ExampleData.XSYG, envir = environment())

##show structure
structure_RLum(OSL.SARMeasurement$Sequence.Object)

#>      id       recordType  curveType protocol.step n.channels x.min   x.max
#> 1     1       TL (UVVIS)   measured            NA       1180 0.100 118.000
#> 2     2          TL (NA) predefined            NA          5 0.000 268.000
#> 3     3          TL (NA)   measured            NA       1740 0.000 173.900
#> 4     4      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 5     5         OSL (NA) predefined            NA          5 0.000 216.000
#> 6     6         OSL (NA)   measured            NA       1150 0.000 114.900
#> 7     7         OSL (NA) predefined            NA          2 0.000  50.000
#> 8     8         OSL (NA)   measured            NA        150 0.497  74.997
#> 9     9 irradiation (NA) predefined            NA          2 0.000  15.000
#> 10   10       TL (UVVIS)   measured            NA        680 0.100  68.000
#> 11   11          TL (NA) predefined            NA          4 0.000 198.000
#> 12   12          TL (NA)   measured            NA        990 0.000  98.900
#> 13   13      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 14   14         OSL (NA) predefined            NA          5 0.000 216.000
#> 15   15         OSL (NA)   measured            NA       1150 0.000 114.900
#> 16   16         OSL (NA) predefined            NA          2 0.000  50.000
#> 17   17         OSL (NA)   measured            NA        151 0.012  75.012
#> 18   18 irradiation (NA) predefined            NA          2 0.000  60.000
#> 19   19       TL (UVVIS)   measured            NA       1180 0.100 118.000
#> 20   20          TL (NA) predefined            NA          5 0.000 268.000
#> 21   21          TL (NA)   measured            NA       1730 0.000 172.900
#> 22   22      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 23   23         OSL (NA) predefined            NA          5 0.000 216.000
#> 24   24         OSL (NA)   measured            NA       1160 0.000 115.900
#> 25   25         OSL (NA) predefined            NA          2 0.000  50.000
#> 26   26         OSL (NA)   measured            NA        153 0.012  76.012
#> 27   27 irradiation (NA) predefined            NA          2 0.000  15.000
#> 28   28       TL (UVVIS)   measured            NA        680 0.100  68.000
#> 29   29          TL (NA) predefined            NA          4 0.000 198.000
#> 30   30          TL (NA)   measured            NA       1000 0.000  99.900
#> 31   31      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 32   32         OSL (NA) predefined            NA          5 0.000 216.000
#> 33   33         OSL (NA)   measured            NA       1150 0.000 114.900
#> 34   34         OSL (NA) predefined            NA          2 0.000  50.000
#> 35   35         OSL (NA)   measured            NA        151 0.014  75.014
#> 36   36 irradiation (NA) predefined            NA          2 0.000 130.000
#> 37   37       TL (UVVIS)   measured            NA       1180 0.100 118.000
#> 38   38          TL (NA) predefined            NA          5 0.000 268.000
#> 39   39          TL (NA)   measured            NA       1740 0.000 173.900
#> 40   40      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 41   41         OSL (NA) predefined            NA          5 0.000 216.000
#> 42   42         OSL (NA)   measured            NA       1150 0.000 114.900
#> 43   43         OSL (NA) predefined            NA          2 0.000  50.000
#> 44   44         OSL (NA)   measured            NA        151 0.012  75.012
#> 45   45 irradiation (NA) predefined            NA          2 0.000  15.000
#> 46   46       TL (UVVIS)   measured            NA        680 0.100  68.000
#> 47   47          TL (NA) predefined            NA          4 0.000 198.000
#> 48   48          TL (NA)   measured            NA        990 0.000  98.900
#> 49   49      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 50   50         OSL (NA) predefined            NA          5 0.000 216.000
#> 51   51         OSL (NA)   measured            NA       1130 0.000 112.900
#> 52   52         OSL (NA) predefined            NA          2 0.000  50.000
#> 53   53         OSL (NA)   measured            NA        147 0.012  73.012
#> 54   54 irradiation (NA) predefined            NA          2 0.000 230.000
#> 55   55       TL (UVVIS)   measured            NA       1180 0.100 118.000
#> 56   56          TL (NA) predefined            NA          5 0.000 268.000
#> 57   57          TL (NA)   measured            NA       1740 0.000 173.900
#> 58   58      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 59   59         OSL (NA) predefined            NA          5 0.000 216.000
#> 60   60         OSL (NA)   measured            NA       1150 0.000 114.900
#> 61   61         OSL (NA) predefined            NA          2 0.000  50.000
#> 62   62         OSL (NA)   measured            NA        151 0.012  75.012
#> 63   63 irradiation (NA) predefined            NA          2 0.000  15.000
#> 64   64       TL (UVVIS)   measured            NA        680 0.100  68.000
#> 65   65          TL (NA) predefined            NA          4 0.000 198.000
#> 66   66          TL (NA)   measured            NA        990 0.000  98.900
#> 67   67      OSL (UVVIS)   measured            NA        500 0.100  50.000
#> 68   68         OSL (NA) predefined            NA          5 0.000 216.000
#> 69   69         OSL (NA)   measured            NA       1150 0.000 114.900
#> 70   70         OSL (NA) predefined            NA          2 0.000  50.000
#> 71   71         OSL (NA)   measured            NA        151 0.012  75.012
#> 72   72 irradiation (NA) predefined            NA          2 0.000 300.000
#> 73   73       TL (UVVIS)   measured            NA       1180 0.100 118.000
#> 74   74          TL (NA) predefined            NA          5 0.000 268.000
#> 75   75          TL (NA)   measured            NA       1750 0.000 174.900
#> 76   76      OSL (UVVIS)   measured            NA        500 0.100  50.000
#>     y.min  y.max    originator                                 .uid
#> 1       0   82.0 read_XSYG2R()   2016-01-30-10:54.0.226260372670367
#> 2      25  260.0 read_XSYG2R()    2016-01-30-10:54.0.77414278825745
#> 3      26  262.0 read_XSYG2R()   2016-01-30-10:54.0.429038936737925
#> 4       0  891.0 read_XSYG2R()   2016-01-30-10:54.0.336071660276502
#> 5      25  125.0 read_XSYG2R()   2016-01-30-10:54.0.832267640857026
#> 6      43  127.0 read_XSYG2R()   2016-01-30-10:54.0.339185907738283
#> 7      40   40.0 read_XSYG2R()   2016-01-30-10:54.0.517060354584828
#> 8       0   40.2 read_XSYG2R()   2016-01-30-10:54.0.537775189615786
#> 9       1    1.0 read_XSYG2R()    2016-01-30-10:54.0.20047601335682
#> 10      4  555.0 read_XSYG2R()   2016-01-30-10:54.0.868833578424528
#> 11     25  160.0 read_XSYG2R()   2016-01-30-10:54.0.742296873126179
#> 12     34  160.0 read_XSYG2R()   2016-01-30-10:54.0.589681873098016
#> 13      0  122.0 read_XSYG2R()   2016-01-30-10:54.0.630885894177482
#> 14     25  125.0 read_XSYG2R()   2016-01-30-10:54.0.663541194982827
#> 15     42  127.0 read_XSYG2R()   2016-01-30-10:54.0.905940495431423
#> 16     40   40.0 read_XSYG2R()   2016-01-30-10:54.0.381334827048704
#> 17      0   40.2 read_XSYG2R()   2016-01-30-10:54.0.328897463157773
#> 18      1    1.0 read_XSYG2R()   2016-01-30-10:54.0.870911038015038
#> 19     11 1591.0 read_XSYG2R()   2016-01-30-10:54.0.932349309092388
#> 20     25  260.0 read_XSYG2R()   2016-01-30-10:54.0.297850674483925
#> 21     33  262.0 read_XSYG2R()   2016-01-30-10:54.0.605329250684008
#> 22      0  473.0 read_XSYG2R()   2016-01-30-10:54.0.289562621153891
#> 23     25  125.0 read_XSYG2R()   2016-01-30-10:54.0.925942022586241
#> 24     43  128.0 read_XSYG2R()   2016-01-30-10:54.0.291019570548087
#> 25     40   40.0 read_XSYG2R()   2016-01-30-10:54.0.320679558906704
#> 26      0   40.2 read_XSYG2R()   2016-01-30-10:54.0.968130854191259
#> 27      1    1.0 read_XSYG2R()  2016-01-30-10:54.0.0512687624432147
#> 28      3  736.0 read_XSYG2R()    2016-01-30-10:54.0.42879255511798
#> 29     25  160.0 read_XSYG2R()  2016-01-30-10:54.0.0236395532265306
#> 30     33  160.0 read_XSYG2R()   2016-01-30-10:54.0.585047567961738
#> 31      0  149.0 read_XSYG2R()   2016-01-30-10:54.0.522002540295944
#> 32     25  125.0 read_XSYG2R()   2016-01-30-10:54.0.454029348213226
#> 33     42  127.0 read_XSYG2R()   2016-01-30-10:54.0.710487167816609
#> 34     40   40.0 read_XSYG2R()   2016-01-30-10:54.0.830461921636015
#> 35      0   40.1 read_XSYG2R()  2016-01-30-10:54.0.0631769173778594
#> 36      1    1.0 read_XSYG2R()   2016-01-30-10:54.0.593671935843304
#> 37     13 3495.0 read_XSYG2R()   2016-01-30-10:54.0.229677951894701
#> 38     25  260.0 read_XSYG2R()   2016-01-30-10:54.0.921212903689593
#> 39     27  263.0 read_XSYG2R()   2016-01-30-10:54.0.462135725654662
#> 40      0 1152.0 read_XSYG2R()   2016-01-30-10:54.0.173798899399117
#> 41     25  125.0 read_XSYG2R()   2016-01-30-10:54.0.905658074887469
#> 42     43  127.0 read_XSYG2R()   2016-01-30-10:54.0.341222140239552
#> 43     40   40.0 read_XSYG2R()     2016-01-30-10:54.0.4814964144025
#> 44      0   40.2 read_XSYG2R()  2016-01-30-10:54.0.0450054712127894
#> 45      1    1.0 read_XSYG2R()   2016-01-30-10:54.0.268596603535116
#> 46      9  912.0 read_XSYG2R()   2016-01-30-10:54.0.370865847915411
#> 47     25  160.0 read_XSYG2R()   2016-01-30-10:54.0.512568636797369
#> 48     33  160.0 read_XSYG2R()   2016-01-30-10:54.0.869454299798235
#> 49      0  215.0 read_XSYG2R()   2016-01-30-10:54.0.186754534952343
#> 50     25  125.0 read_XSYG2R()   2016-01-30-10:54.0.931624148506671
#> 51     43  127.0 read_XSYG2R()   2016-01-30-10:54.0.561583923641592
#> 52     40   40.0 read_XSYG2R()   2016-01-30-10:54.0.465734504396096
#> 53      0   40.1 read_XSYG2R()   2016-01-30-10:54.0.929129796102643
#> 54      1    1.0 read_XSYG2R()    2016-01-30-10:54.0.29691950744018
#> 55     25 5564.0 read_XSYG2R()   2016-01-30-10:54.0.534651084803045
#> 56     25  260.0 read_XSYG2R()   2016-01-30-10:54.0.792043952038512
#> 57     27  262.0 read_XSYG2R()   2016-01-30-10:54.0.208661240758374
#> 58      0 2346.0 read_XSYG2R()   2016-01-30-10:54.0.937022477388382
#> 59     25  125.0 read_XSYG2R()   2016-01-30-10:54.0.633003524737433
#> 60     43  127.0 read_XSYG2R()   2016-01-30-10:54.0.744804858230054
#> 61     40   40.0 read_XSYG2R()   2016-01-30-10:54.0.451362358173355
#> 62      0   40.3 read_XSYG2R()    2016-01-30-10:54.0.91636705910787
#> 63      1    1.0 read_XSYG2R()  2016-01-30-10:54.0.0463254018686712
#> 64     12 1146.0 read_XSYG2R()   2016-01-30-10:54.0.528479704400524
#> 65     25  160.0 read_XSYG2R()   2016-01-30-10:54.0.478199049830437
#> 66     34  160.0 read_XSYG2R()   2016-01-30-10:54.0.122971840202808
#> 67      0  292.0 read_XSYG2R()   2016-01-30-10:54.0.879391883034259
#> 68     25  125.0 read_XSYG2R()   2016-01-30-10:54.0.242106179939583
#> 69     42  128.0 read_XSYG2R()   2016-01-30-10:54.0.482210234506056
#> 70     40   40.0 read_XSYG2R()   2016-01-30-10:54.0.936310092685744
#> 71      0   40.1 read_XSYG2R()    2016-01-30-10:54.0.58761327737011
#> 72      1    1.0 read_XSYG2R()   2016-01-30-10:54.0.662958577508107
#> 73     47 7423.0 read_XSYG2R()   2016-01-30-10:54.0.349518121918663
#> 74     25  260.0 read_XSYG2R()   2016-01-30-10:54.0.395675572101027
#> 75     25  261.0 read_XSYG2R()   2016-01-30-10:54.0.832506845006719
#> 76      0 3505.0 read_XSYG2R()   2016-01-30-10:54.0.673705363180488
#>                                   .pid
#> 1   2016-01-30-10:56.0.474548307945952
#> 2   2016-01-30-10:56.0.474548307945952
#> 3   2016-01-30-10:56.0.474548307945952
#> 4   2016-01-30-10:56.0.474548307945952
#> 5   2016-01-30-10:56.0.474548307945952
#> 6   2016-01-30-10:56.0.474548307945952
#> 7   2016-01-30-10:56.0.474548307945952
#> 8   2016-01-30-10:56.0.474548307945952
#> 9   2016-01-30-10:56.0.474548307945952
#> 10  2016-01-30-10:56.0.474548307945952
#> 11  2016-01-30-10:56.0.474548307945952
#> 12  2016-01-30-10:56.0.474548307945952
#> 13  2016-01-30-10:56.0.474548307945952
#> 14  2016-01-30-10:56.0.474548307945952
#> 15  2016-01-30-10:56.0.474548307945952
#> 16  2016-01-30-10:56.0.474548307945952
#> 17  2016-01-30-10:56.0.474548307945952
#> 18  2016-01-30-10:56.0.474548307945952
#> 19  2016-01-30-10:56.0.474548307945952
#> 20  2016-01-30-10:56.0.474548307945952
#> 21  2016-01-30-10:56.0.474548307945952
#> 22  2016-01-30-10:56.0.474548307945952
#> 23  2016-01-30-10:56.0.474548307945952
#> 24  2016-01-30-10:56.0.474548307945952
#> 25  2016-01-30-10:56.0.474548307945952
#> 26  2016-01-30-10:56.0.474548307945952
#> 27  2016-01-30-10:56.0.474548307945952
#> 28  2016-01-30-10:56.0.474548307945952
#> 29  2016-01-30-10:56.0.474548307945952
#> 30  2016-01-30-10:56.0.474548307945952
#> 31  2016-01-30-10:56.0.474548307945952
#> 32  2016-01-30-10:56.0.474548307945952
#> 33  2016-01-30-10:56.0.474548307945952
#> 34  2016-01-30-10:56.0.474548307945952
#> 35  2016-01-30-10:56.0.474548307945952
#> 36  2016-01-30-10:56.0.474548307945952
#> 37  2016-01-30-10:56.0.474548307945952
#> 38  2016-01-30-10:56.0.474548307945952
#> 39  2016-01-30-10:56.0.474548307945952
#> 40  2016-01-30-10:56.0.474548307945952
#> 41  2016-01-30-10:56.0.474548307945952
#> 42  2016-01-30-10:56.0.474548307945952
#> 43  2016-01-30-10:56.0.474548307945952
#> 44  2016-01-30-10:56.0.474548307945952
#> 45  2016-01-30-10:56.0.474548307945952
#> 46  2016-01-30-10:56.0.474548307945952
#> 47  2016-01-30-10:56.0.474548307945952
#> 48  2016-01-30-10:56.0.474548307945952
#> 49  2016-01-30-10:56.0.474548307945952
#> 50  2016-01-30-10:56.0.474548307945952
#> 51  2016-01-30-10:56.0.474548307945952
#> 52  2016-01-30-10:56.0.474548307945952
#> 53  2016-01-30-10:56.0.474548307945952
#> 54  2016-01-30-10:56.0.474548307945952
#> 55  2016-01-30-10:56.0.474548307945952
#> 56  2016-01-30-10:56.0.474548307945952
#> 57  2016-01-30-10:56.0.474548307945952
#> 58  2016-01-30-10:56.0.474548307945952
#> 59  2016-01-30-10:56.0.474548307945952
#> 60  2016-01-30-10:56.0.474548307945952
#> 61  2016-01-30-10:56.0.474548307945952
#> 62  2016-01-30-10:56.0.474548307945952
#> 63  2016-01-30-10:56.0.474548307945952
#> 64  2016-01-30-10:56.0.474548307945952
#> 65  2016-01-30-10:56.0.474548307945952
#> 66  2016-01-30-10:56.0.474548307945952
#> 67  2016-01-30-10:56.0.474548307945952
#> 68  2016-01-30-10:56.0.474548307945952
#> 69  2016-01-30-10:56.0.474548307945952
#> 70  2016-01-30-10:56.0.474548307945952
#> 71  2016-01-30-10:56.0.474548307945952
#> 72  2016-01-30-10:56.0.474548307945952
#> 73  2016-01-30-10:56.0.474548307945952
#> 74  2016-01-30-10:56.0.474548307945952
#> 75  2016-01-30-10:56.0.474548307945952
#> 76  2016-01-30-10:56.0.474548307945952
#>                                                                                     info
#> 1   state parentID startDate curveType detector curveDescripter offset interval duration
#> 2                   state parentID startDate curveType stimulator curveDescripter offset
#> 3                          state parentID startDate curveType stimulator curveDescripter
#> 4   state parentID startDate curveType detector curveDescripter offset interval duration
#> 5                   state parentID startDate curveType stimulator curveDescripter offset
#> 6                          state parentID startDate curveType stimulator curveDescripter
#> 7                   state parentID startDate curveType stimulator curveDescripter offset
#> 8                   state parentID startDate curveType stimulator curveDescripter offset
#> 9                   state parentID startDate curveType stimulator curveDescripter offset
#> 10  state parentID startDate curveType detector curveDescripter offset interval duration
#> 11                  state parentID startDate curveType stimulator curveDescripter offset
#> 12                         state parentID startDate curveType stimulator curveDescripter
#> 13  state parentID startDate curveType detector curveDescripter offset interval duration
#> 14                  state parentID startDate curveType stimulator curveDescripter offset
#> 15                         state parentID startDate curveType stimulator curveDescripter
#> 16                  state parentID startDate curveType stimulator curveDescripter offset
#> 17                  state parentID startDate curveType stimulator curveDescripter offset
#> 18                  state parentID startDate curveType stimulator curveDescripter offset
#> 19  state parentID startDate curveType detector curveDescripter offset interval duration
#> 20                  state parentID startDate curveType stimulator curveDescripter offset
#> 21                         state parentID startDate curveType stimulator curveDescripter
#> 22  state parentID startDate curveType detector curveDescripter offset interval duration
#> 23                  state parentID startDate curveType stimulator curveDescripter offset
#> 24                         state parentID startDate curveType stimulator curveDescripter
#> 25                  state parentID startDate curveType stimulator curveDescripter offset
#> 26                  state parentID startDate curveType stimulator curveDescripter offset
#> 27                  state parentID startDate curveType stimulator curveDescripter offset
#> 28  state parentID startDate curveType detector curveDescripter offset interval duration
#> 29                  state parentID startDate curveType stimulator curveDescripter offset
#> 30                         state parentID startDate curveType stimulator curveDescripter
#> 31  state parentID startDate curveType detector curveDescripter offset interval duration
#> 32                  state parentID startDate curveType stimulator curveDescripter offset
#> 33                         state parentID startDate curveType stimulator curveDescripter
#> 34                  state parentID startDate curveType stimulator curveDescripter offset
#> 35                  state parentID startDate curveType stimulator curveDescripter offset
#> 36                  state parentID startDate curveType stimulator curveDescripter offset
#> 37  state parentID startDate curveType detector curveDescripter offset interval duration
#> 38                  state parentID startDate curveType stimulator curveDescripter offset
#> 39                         state parentID startDate curveType stimulator curveDescripter
#> 40  state parentID startDate curveType detector curveDescripter offset interval duration
#> 41                  state parentID startDate curveType stimulator curveDescripter offset
#> 42                         state parentID startDate curveType stimulator curveDescripter
#> 43                  state parentID startDate curveType stimulator curveDescripter offset
#> 44                  state parentID startDate curveType stimulator curveDescripter offset
#> 45                  state parentID startDate curveType stimulator curveDescripter offset
#> 46  state parentID startDate curveType detector curveDescripter offset interval duration
#> 47                  state parentID startDate curveType stimulator curveDescripter offset
#> 48                         state parentID startDate curveType stimulator curveDescripter
#> 49  state parentID startDate curveType detector curveDescripter offset interval duration
#> 50                  state parentID startDate curveType stimulator curveDescripter offset
#> 51                         state parentID startDate curveType stimulator curveDescripter
#> 52                  state parentID startDate curveType stimulator curveDescripter offset
#> 53                  state parentID startDate curveType stimulator curveDescripter offset
#> 54                  state parentID startDate curveType stimulator curveDescripter offset
#> 55  state parentID startDate curveType detector curveDescripter offset interval duration
#> 56                  state parentID startDate curveType stimulator curveDescripter offset
#> 57                         state parentID startDate curveType stimulator curveDescripter
#> 58  state parentID startDate curveType detector curveDescripter offset interval duration
#> 59                  state parentID startDate curveType stimulator curveDescripter offset
#> 60                         state parentID startDate curveType stimulator curveDescripter
#> 61                  state parentID startDate curveType stimulator curveDescripter offset
#> 62                  state parentID startDate curveType stimulator curveDescripter offset
#> 63                  state parentID startDate curveType stimulator curveDescripter offset
#> 64  state parentID startDate curveType detector curveDescripter offset interval duration
#> 65                  state parentID startDate curveType stimulator curveDescripter offset
#> 66                         state parentID startDate curveType stimulator curveDescripter
#> 67  state parentID startDate curveType detector curveDescripter offset interval duration
#> 68                  state parentID startDate curveType stimulator curveDescripter offset
#> 69                         state parentID startDate curveType stimulator curveDescripter
#> 70                  state parentID startDate curveType stimulator curveDescripter offset
#> 71                  state parentID startDate curveType stimulator curveDescripter offset
#> 72                  state parentID startDate curveType stimulator curveDescripter offset
#> 73  state parentID startDate curveType detector curveDescripter offset interval duration
#> 74                  state parentID startDate curveType stimulator curveDescripter offset
#> 75                         state parentID startDate curveType stimulator curveDescripter
#> 76  state parentID startDate curveType detector curveDescripter offset interval duration
#>  [ reached getOption("max.print") -- omitted 48 rows ]