Simple Subsetting of Single Grain Data from Risø BIN/BINX files

Most measured single grains do not exhibit light and it makes usually sense to subset single grain datasets using a table of position and grain pairs

Usage

subset_SingleGrainData(object, selection)

Arguments

object

Risoe.BINfileData (required): input object with the data to subset

selection

data.frame (required): selection table with two columns for position (1st column) and grain (2nd column) (columns names do not matter)

Value

A subset Risoe.BINfileData object

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany) , RLum Developer Team

How to cite

Kreutzer, S., 2024. subset_SingleGrainData(): Simple Subsetting of Single Grain Data from Risø BIN/BINX files. Function version 0.1.0. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., Galharret, J., Colombo, M., 2024. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.25. https://r-lum.github.io/Luminescence/

Examples

## load example data
data(ExampleData.BINfileData, envir = environment())

## set POSITION/GRAIN pair dataset
selection <- data.frame(POSITION = c(1,5,7), GRAIN = c(0,0,0))

##subset
subset_SingleGrainData(object = CWOSL.SAR.Data, selection = selection)
#> 
#> [Risoe.BINfileData object]
#> 
#> 	BIN/BINX version:     03
#> 	Object date:          060920, 070920, 080920, 090920, 100920
#> 	User:                 Default
#> 	System ID:            0 (unknown)
#> 	Overall records:      90
#> 	Records type:         IRSL	(n = 3)
#> 	                      OSL	(n = 42)
#> 	                      TL	(n = 45)
#> 	Position range:       1 : 7
#> 	Run range:            1 : 8