template_DRAC {Luminescence}	R Documentation

Create a DRAC input data template (v1.2)

Description

This function returns a DRAC input template (v1.2) to be used in conjunction with the use_DRAC function

Usage


template_DRAC(nrow = 1, preset = NULL, notification = TRUE)

Arguments

nrow
integer (with default): specifies the number of rows of the template (i.e., the number of data sets you want to submit).

preset
character (optional): By default, all values of the template are set to NA, which means that the user needs to fill in all data first before submitting to DRAC using use_DRAC(). To reduce the number of values that need to be provided, preset can be used to create a template with at least a minimum of reasonable preset values.

preset can be one of the following:

quartz_coarse

quartz_fine

feldspar_coarse

polymineral_fine

DRAC-example_quartz

DRAC-example_feldspar

DRAC-example_polymineral

Note that the last three options can be used to produce a template with values directly taken from the official DRAC input .csv file.

notification
logical (with default): show or hide the notification

Value

A list.

How to cite

Burow, C., Kreutzer, S., 2021. template_DRAC(): Create a DRAC input data template (v1.2). In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J., Mercier, N., Philippe, A., Riedesel, S., Autzen, M., Mittelstrass, D., Gray, H.J., 2021. Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.15. https://CRAN.R-project.org/package=Luminescence

Author(s)

Christoph Burow, University of Cologne (Germany), Sebastian Kreutzer, Geography & Earth Sciences, Aberystwyth University (United Kingdom) , RLum Developer Team

References

Durcan, J.A., King, G.E., Duller, G.A.T., 2015. DRAC: Dose Rate and Age Calculator for trapped charge dating. Quaternary Geochronology 28, 54-61. doi:10.1016/j.quageo.2015.03.012

See Also

as.data.frame, list

Examples



# create a new DRAC input input
input <- template_DRAC(preset = "DRAC-example_quartz")

# show content of the input
print(input)
print(input$`Project ID`)
print(input[[4]])


## Example: DRAC Quartz example
# note that you only have to assign new values where they
# are different to the default values
input$`Project ID` <- "DRAC-Example"
input$`Sample ID` <- "Quartz"
input$`Conversion factors` <- "AdamiecAitken1998"
input$`External U (ppm)` <- 3.4
input$`errExternal U (ppm)` <- 0.51
input$`External Th (ppm)` <- 14.47
input$`errExternal Th (ppm)` <- 1.69
input$`External K (%)` <- 1.2
input$`errExternal K (%)` <- 0.14
input$`Calculate external Rb from K conc?` <- "N"
input$`Calculate internal Rb from K conc?` <- "N"
input$`Scale gammadoserate at shallow depths?` <- "N"
input$`Grain size min (microns)` <- 90
input$`Grain size max (microns)` <- 125
input$`Water content ((wet weight - dry weight)/dry weight) %` <- 5
input$`errWater content %` <- 2
input$`Depth (m)` <- 2.2
input$`errDepth (m)` <- 0.22
input$`Overburden density (g cm-3)` <- 1.8
input$`errOverburden density (g cm-3)` <- 0.1
input$`Latitude (decimal degrees)` <- 30.0000
input$`Longitude (decimal degrees)` <- 70.0000
input$`Altitude (m)` <- 150
input$`De (Gy)` <- 20
input$`errDe (Gy)` <- 0.2

# use DRAC
## Not run: 
output <- use_DRAC(input)

## End(Not run)


[Package Luminescence version 0.9.15 Index]