Export Risoe.BINfileData into Risø BIN/BINX-file

Exports a Risoe.BINfileData object in a *.bin or *.binx file that can be opened by the Analyst software or other Risoe software.

write_R2BIN(object, file, version, compatibility.mode = FALSE,
  txtProgressBar = TRUE)

Arguments

object
Risoe.BINfileData (required): input object to be stored in a bin file.

file
character (required): file name and path of the output file

[WIN]: write_R2BIN(object, "C:/Desktop/test.bin")

[MAC/LINUX]: write_R2BIN("/User/test/Desktop/test.bin")

version
character (optional): version number for the output file. If no value is provided the highest version number from the Risoe.BINfileData is taken automatically.

Note: This argument can be used to convert BIN-file versions.

compatibility.mode
logical (with default): this option recalculates the position values if necessary and set the max. value to 48. The old position number is appended as comment (e.g., 'OP: 70). This option accounts for potential compatibility problems with the Analyst software. It further limits the maximum number of points per curve to 9,999. If a curve contains more data the curve data got binned using the smallest possible bin width.

txtProgressBar
logical (with default): enables or disables txtProgressBar.

Value

Write a binary file.

Details

The structure of the exported binary data follows the data structure published in the Appendices of the Analyst manual p. 42.

If LTYPE, DTYPE and LIGHTSOURCE are not of type character, no transformation into numeric values is done.

Note

The function just roughly checks the data structures. The validity of the output data depends on the user.

The validity of the file path is not further checked. BIN-file conversions using the argument version may be a lossy conversion, depending on the chosen input andoutput data (e.g., conversion from version 08 to 07 to 06 to 05 to 04 or 03).

Warning

Although the coding was done carefully it seems that the BIN/BINX-files produced by Risø DA 15/20 TL/OSL readers slightly differ on the byte level. No obvious differences are observed in the METADATA, however, the BIN/BINX-file may not fully compatible, at least not similar to the once directly produced by the Risoe readers!

ROI definitions (introduced in BIN-file version 8) are not supported! There are furthermore ignored by the function read_BIN2R.

Function version

0.5.1 (2019-03-07 12:45:56)

How to cite

Kreutzer, S. (2019). write_R2BIN(): Export Risoe.BINfileData into Risø BIN/BINX-file. Function version 0.5.1. In: Kreutzer, S., Burow, C., Dietze, M., Fuchs, M.C., Schmidt, C., Fischer, M., Friedrich, J. (2019). Luminescence: Comprehensive Luminescence Dating Data Analysis. R package version 0.9.0.90. https://CRAN.R-project.org/package=Luminescence

References

DTU Nutech, 2016. The Squence Editor, Users Manual, February, 2016. http://www.nutech.dtu.dk/english/products-and-services/radiation-instruments/tl_osl_reader/manuals

See also

Examples

# NOT RUN {
##load exampled dataset
data(ExampleData.BINfileData, envir = environment())

##create temporary filepath
##(for usage replace by own path)
temp_file <- temp_file <- tempfile(pattern = "output", fileext = ".bin")

##export to temporary file path
write_R2BIN(CWOSL.SAR.Data, file = temp_file)

# }