Export RLum.Data.Image and RLum.Data.Spectrum objects to TIFF Images

Simple wrapper around tiff::writeTIFF to export suitable RLum-class objects to TIFF images. Per default 16-bit TIFF files are exported.

Usage

write_R2TIFF(object, file = tempfile(), norm = 65535, ...)

Arguments

object

RLum.Data.Image or RLum.Data.Spectrum object (required): input object, can be a list of such objects

file

character (required): the file name and path

norm

numeric (with default): normalisation values. Values in TIFF files must range between 0-1, however, usually in imaging applications the pixel values are real integer count values. The normalisation to the to the highest 16-bit integer values -1 ensures that the numerical values are retained in the exported image. If 1 nothing is normalised.

...

further arguments to be passed to tiff::writeTIFF.

Value

A TIFF file

Function version

0.1.0

See also

Author

Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)

Examples

data(ExampleData.RLum.Data.Image, envir = environment())
write_R2TIFF(ExampleData.RLum.Data.Image, file = tempfile())